### 2023-12-19 #6161 @editorialbot
[PRE REVIEW]: AutoEIS: An automated tool for analysis of electrochemical impedance spectroscopy using evolutionary algorithms and Bayesian inference
**Submitting author:** @ma-sadeghi (<a href="http://orcid.org/0000-0002-6756-9117">Mohammad Amin Sadeghi</a>)
**Repository:** https://github.com/AUTODIAL/AutoEIS/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.18
