### 2023-12-19 4c34fde
Add function to tabulate summary statistics of MCMC chains
