### 2023-12-19 4a5e052
Add function to find identical rows in a 2D array
