### 2023-12-19 cdb476a
Add function to check if two circuits are equivalent
