### 2023-12-19 6274d32
Timeout `fit_circuit_parameters` after 20 seconds (-> diverged)
