### 2023-12-19 8800e97
Refactor `parse_circuit_dataframe` for readability and conciseness
