### 2023-12-19 29a64b6
Archive more functions under legacy.py
