### 2023-12-19 178c70a
Use CPU by default, GPU is slower on EIS datasets
