### 2023-12-19 a671472
Refactor complex heuristic filters into `merge_identical_circuits`
