### 2023-12-19 95c443a
Update `apply_heuristic_filters` to use the refactored methods
