### 2023-12-19 cc5750b
Expose `num_warmup`, `num_samples`, and `progress_bar` in `perform_bayesian_inference` as args
