### 2023-12-19 fabc79e
Cover more cases in `find_series_elements` unit test
