### 2023-12-19 4ebb87c
Archive `generate_mathematical_expression` in legacy.py
