### 2023-12-19 4e18244
Remove old helper functions
