+++ autoeis/core.py
@@ -499,288 +499,6 @@ def precise_rank_the_structure(circuit_array: np.array) -> np.array:
     return ranks_array[0]
 
 
-def precise_extractor(circuit_array: np.ndarray, precise_ranks_array: np.ndarray) -> list:
-    """Extracts the index information of each level circuit according to
-    ranking array in a more precise way.
-
-    Parameters
-    ----------
-    circuit_array: np.ndarray
-        The ndarray that stores the circuit configurations
-    precise_ranks_array:np.array
-        The ndarray that stores the level information of given circuits
-
-    Returns
-    -------
-    level_lists: list
-        The list that stores the index information of each level circuit
-    """
-    level_lists = []
-    if circuit_array[0] == 5:
-        # detect [
-        # print(1,0)
-        level_list = []
-        level_list.append(int(precise_ranks_array[0]))
-        for j in range(0, len(precise_ranks_array) - 1):
-            if (
-                precise_ranks_array[j] > int(precise_ranks_array[j])
-                and int(precise_ranks_array[j]) == precise_ranks_array[0]
-            ):
-                # print("1 end:",j)
-                level_list.extend(z for z in circuit_array[0 + 1 : j])
-                level_lists.append(level_list)
-                break
-    for i in range(len(precise_ranks_array) - 1):
-        if int(precise_ranks_array[i + 1]) - int(precise_ranks_array[i]) == 1:
-            # detect[
-            # print(1,i)
-            level_list = []
-            level_list.append(int(precise_ranks_array[i + 1]))
-            for j in range(i, len(precise_ranks_array) - 1):
-                if (
-                    precise_ranks_array[j] > int(precise_ranks_array[j])
-                    and int(precise_ranks_array[j]) == precise_ranks_array[i + 1]
-                ):
-                    # print("1 end:",j)
-                    level_list.extend(z for z in circuit_array[i + 2 : j])
-                    level_lists.append(level_list)
-                    break
-        elif round(precise_ranks_array[i + 1] - precise_ranks_array[i], 1) == 0.1:
-            # detect ","
-            # print(2,i)
-            level_list = []
-            level_list.append(int(precise_ranks_array[i + 1]))
-            for j in range(i + 2, len(precise_ranks_array)):
-                if (
-                    precise_ranks_array[j] > precise_ranks_array[j - 1]
-                    and int(precise_ranks_array[j]) == int(precise_ranks_array[j - 1])
-                    and int(precise_ranks_array[j]) == int(precise_ranks_array[i + 1])
-                    or precise_ranks_array[j] == int(precise_ranks_array[i] - 1)
-                ):
-                    # print("2 end:", j)
-                    level_list.extend(z for z in circuit_array[i + 2 : j])
-                    level_lists.append(level_list)
-                    break
-    return level_lists
-
-
-def sort_level_lists(level_lists: "list") -> "list":
-    """Sorts the level lists.
-
-    Parameters
-    ----------
-    level_lists: list
-        The list that stores the index information of each level circuit
-
-    Returns
-    -------
-    level_lists: list
-        The sorted list that stores the index information of each level circuit
-    """
-    for i in range(len(level_lists)):
-        level_lists[i] = sorted(level_lists[i])
-    level_lists = sorted(level_lists)
-    return level_lists
-
-
-def feature_store(circuit: str) -> dict:
-    """Extract the features of the circuit and store them as a dictionary
-     with the following keys:
-
-        - Feature 1: the component numbers (R/C/L/P/-/,/[/]) of a given circuit
-        - Feature 2: the component numbers in series part of a given circuit
-        - Feature 2.5: the parallel parts of a given circuit at different levels
[…]
