### 2023-12-19 ff00a00
Remove unused function
