### 2023-12-19 1a0cc2d
Fix a bug in `circuit_complexity` (account for '-' in circuit string)
