### 2023-12-19 ae0f578
Add unit test for `circuit_complexity`
