### 2023-12-19 c9dea5e
Make `plot_impedance_combo` tight layout
