### 2023-12-19 0d66955
Cover more edge cases in `circuit_complexity` unit test
