### 2023-12-19 766de84
Update unit tests based on recent API changes
