+++ tests/test_core.py
@@ -1,14 +1,12 @@
 import re
 
 import numpy as np
-import pandas as pd
-import pytest
 from impedance.models.circuits import CustomCircuit
 
 import autoeis as ae
 
 
-def test_generate_mathematical_expression():
+def test_generate_circuit_fn():
     circuit = "R0-C1-[P2,R3]-[P4-[P5,C6],[L7,R8]]"
 
     resistors = re.findall(r"R\d+", circuit)
@@ -20,11 +18,8 @@ def test_generate_mathematical_expression():
     freq = np.array([1, 10, 100])
     p = np.random.rand(num_params)
 
-    circuit_df = pd.DataFrame([circuit], columns=["circuitstring"])
-    func_str = ae.core.generate_mathematical_expression(circuit_df)
-    func_str = func_str["Mathematical expressions"][0]
-    def func(X, F): return eval(func_str)
-    Z_py = func(p, freq)
+    circuit_fn = ae.utils.generate_circuit_fn(circuit)
+    Z_py = circuit_fn(p, freq)
 
     ec = ae.julia_helpers.import_backend()
     Z_jl = np.array([ec.get_target_impedance(circuit, p, f) for f in freq])
@@ -42,10 +37,9 @@ def test_compute_ohmic_resistance():
     freq = np.logspace(-3, 3, 1000)
     Z = circuit.predict(freq)
     R = ae.core.compute_ohmic_resistance(Z, freq)
-    assert np.isclose(R, R1, rtol=0.15)
+    np.testing.assert_allclose(R, R1, rtol=0.15)
 
 
-# FIXME: We don't have a robust way to handle missing high frequency data
 def test_compute_ohmic_resistance_missing_high_freq():
     circuit_string = "R1-[P2,P3-R4]"
     circuit_string = ae.utils.impedancepy_circuit(circuit_string)
@@ -53,7 +47,8 @@ def test_compute_ohmic_resistance_missing_high_freq():
     parameters = np.array([R1, 1e-3, 0.1, 5e-5, 0.8, 10])
     circuit = CustomCircuit(circuit_string, initial_guess=parameters)
     circuit.parameters_ = parameters
-    freq = np.logspace(-3, 1, 1000)
+    freq = np.logspace(-3, 0, 1000)
     Z = circuit.predict(freq)
-    # with pytest.raises(ValueError):
-    #     R = ae.core.compute_ohmic_resistance(Z, freq)
+    R = ae.core.compute_ohmic_resistance(Z, freq)
+    Zreal_at_high_freq = Z.real[-1]
+    np.testing.assert_allclose(R, Zreal_at_high_freq)
