### 2023-12-19 2c479c7
Make `compute_ohmic_resistance` more robust when high frequency data missing
