### 2023-12-19 0f763e6
Identify Python shell vs notebook more robustly
