### 2023-12-19 8f6754e
Only apply timeout on non-Windows OS
