### 2023-12-19 7ba1500
Update `perform_full_analysis` to work with new API
