### 2023-12-19 0ca3045
Remove thin IO wrappers
