### 2023-12-19 372c00a
Add helper functions to load a test dataset
