### 2023-12-19 e86d474
Use the new `load_test_dataset` throughout
