### 2023-12-19 8964d5c
Add unit tests for `get_assets_path` and `load_test_dataset`
