+++ _sources/generated/generated/autoeis.utils.reset_storage.rst.txt
@@ -1,6 +1,0 @@
-reset\_storage
-==============
-
-.. currentmodule:: autoeis.utils
-
-.. autofunction:: reset_storage
