+++ _sources/generated/generated/autoeis.io.get_assets_path.rst.txt
@@ -1,0 +1,6 @@
+get\_assets\_path
+=================
+
+.. currentmodule:: autoeis.io
+
+.. autofunction:: get_assets_path
