+++ _sources/generated/generated/autoeis.utils.get_filename.rst.txt
@@ -1,6 +1,0 @@
-get\_filename
-=============
-
-.. currentmodule:: autoeis.utils
-
-.. autofunction:: get_filename
