+++ _sources/generated/generated/autoeis.utils.make_dir.rst.txt
@@ -1,6 +1,0 @@
-make\_dir
-=========
-
-.. currentmodule:: autoeis.utils
-
-.. autofunction:: make_dir
