+++ _sources/generated/generated/autoeis.utils.get_dirname.rst.txt
@@ -1,6 +1,0 @@
-get\_dirname
-============
-
-.. currentmodule:: autoeis.utils
-
-.. autofunction:: get_dirname
