+++ _sources/generated/generated/autoeis.visualization.plot_nyquist.rst.txt
@@ -1,0 +1,6 @@
+plot\_nyquist
+=============
+
+.. currentmodule:: autoeis.visualization
+
+.. autofunction:: plot_nyquist
