### 2023-12-19 a9b3293
Add missing DOI, format bib file
