+++ assets/paper.bib
@@ -85,19 +85,20 @@
 	year={1995}
 }
 @article{zhang2023,
-	doi = {10.1149/1945-7111/aceab2},
-	url = {https://dx.doi.org/10.1149/1945-7111/aceab2},
-	year = {2023},
-	month = {aug},
-	publisher = {IOP Publishing},
-	volume = {170},
-	number = {8},
-	pages = {086502},
-	author = {Runze Zhang and Robert Black and Debashish Sur and Parisa Karimi and Kangming Li and Brian DeCost and John R. Scully and Jason Hattrick-Simpers},
-	title = {AutoEIS: Automated Bayesian Model Selection and Analysis for Electrochemical Impedance Spectroscopy},
-	journal = {Journal of The Electrochemical Society},
+	doi={10.1149/1945-7111/aceab2},
+	url={https://dx.doi.org/10.1149/1945-7111/aceab2},
+	year={2023},
+	month={aug},
+	publisher={IOP Publishing},
+	volume={170},
+	number={8},
+	pages={086502},
+	author={Runze Zhang and Robert Black and Debashish Sur and Parisa Karimi and Kangming Li and Brian DeCost and John R. Scully and Jason Hattrick-Simpers},
+	title={AutoEIS: Automated Bayesian Model Selection and Analysis for Electrochemical Impedance Spectroscopy},
+	journal={Journal of The Electrochemical Society},
 }
 @article{van2021practical,
+	doi={10.1109/TIM.2021.3113116},
   	title={Practical equivalent electrical circuit identification for electrochemical impedance spectroscopy analysis with gene expression programming},
   	author={Van Haeverbeke, Maxime and Stock, Michiel and De Baets, Bernard},
   	journal={IEEE Transactions on Instrumentation and Measurement},
