+++ doc/requirements.txt
@@ -1,6 +1,0 @@
-furo
-linkify-it-py
-myst-parser
-sphinx
-sphinx-copybutton
-sphinx-autodoc2
