+++ .github/workflows/tests.yml
@@ -33,9 +33,8 @@ jobs:
       - name: "Install AutoEIS"
         run: |
             python -m pip install --upgrade pip
-            pip install -r requirements.txt
-            pip install . --no-deps
+            pip install -e .[test]
             python -m autoeis install
       - name: "Run tests"
         run: |
-          pytest .
+          pytest -v .
