### 2023-12-20 34c691d
Remove unused functions, fix docstrings, undocument internal functions
