### 2023-12-20 0ed6c4f
Add type hints throughout the package
