### 2023-12-20 4182765
Move metrics-like functions to metrics module
