### 2023-12-21 ed3b143
Move circuit parsing-related functions to parser module
