+++ autoeis/utils.py
@@ -6,31 +6,13 @@ Collection of utility functions used throughout the package.
 .. autosummary::
    :toctree: generated/
 
-    parse_component
-    parse_parameter
-    count_parameters
     circuit_complexity
-    impedancepy_circuit
-    validate_parameter_label
-    validate_circuit_string
-    get_component_labels
-    get_component_types
-    get_parameter_types
-    get_parameter_labels
-    group_parameters_by_type
-    group_parameters_by_component
-    find_series_elements
-    find_ohmic_resistors
     generate_circuit_fn
     generate_circuit_fn_impy
     fit_circuit_parameters
     is_equivalent
     initialize_priors
     initialize_priors_from_posteriors
-    mape_score
-    mse_score
-    rmse_score
-    r2_score
 
 """
 import logging
@@ -49,9 +31,10 @@ import pandas as pd
 import rich.traceback
 from impedance.models.circuits import CustomCircuit
 from numpy import pi
-from pyparsing import nested_expr
 from rich.logging import RichHandler
 from scipy.stats import lognorm, norm
+
+from autoeis import parser
 
 # >>> Logging utils
 
@@ -76,7 +59,7 @@ def setup_rich_tracebacks():
 # <<< Logging utils
 
 
-# >>> Filesystem utils
+# >>> General utils
 
 def flatten(xs: list) -> list:
     """Returns a list of all elements in a nested iterable."""
@@ -150,110 +133,10 @@ def timeout(seconds):
         return wrapper
     return decorator
 
-# <<< Filesystem utils
+# <<< General utils
 
 
 # >>> Circuit utils
-
-def validate_parameter_label(p: str) -> bool:
-    """Checks if a parameter label is valid."""
-    # Check if parameter label is a string
-    assert isinstance(p, str), "Parameter label must be a string."
-    # Check if parameter label is not empty
-    assert len(p) > 0, "Parameter label is empty."
-    # Check if parameter label is valid
-    pattern = r"(?:R\d+|C\d+|L\d+|P\d+[wn])"
-    assert re.fullmatch(pattern, p), f"Invalid parameter label: {p}"
-    
-
-def parse_component(c:str) -> str:
-    """Returns the type of a component label, e.g., R1 -> R"""
-    return re.match(r"[A-Za-z]+", c).group()
-
-
-def parse_parameter(p:str, by: str = "type") -> str:
-    """Returns the type/component of a parameter label, e.g., P4n -> Pn/P"""
-    assert by in ["type", "component"], "by must be either 'type' or 'component'."
-    validate_parameter_label(p)
-    if p.startswith(("R", "C", "L")):
-        ptype, ctype = p[0], p[0]
-    elif p.startswith("P") and p.endswith("w"):
-        ptype, ctype = "Pw", "P"
-    elif p.startswith("P") and p.endswith("n"):
-        ptype, ctype = "Pn", "P"
-    return ptype if by == "type" else ctype
-
-
-def get_component_labels(circuit: str, types: list[str] = None) -> list[str]:
-    """Returns a list of labels for all components in a circuit string."""
-    types = [types] if isinstance(types, str) else types
-    types = ["R", "C", "L", "P"] if types is None else types
-    assert isinstance(types, list), "types must be a list of strings."
-    pattern = rf'\b(?:{"|".join(types)})\d+\b'
-    return re.findall(pattern, circuit)
-
-
-def get_component_types(circuit: str, unique: bool = False) -> list[str]:
-    """Returns a list of component types in a circuit string."""
-    types = re.findall(r"[A-Za-z]+", circuit)
-    return list(set(types)) if unique else types
-
-
-def get_parameter_types(circuit: str, unique: bool = False) -> list[str]:
[…]
