+++ tests/test_utils.py
@@ -1,8 +1,7 @@
 import numpy as np
-import pytest
 from impedance.models.circuits import CustomCircuit
 
-from autoeis import utils
+from autoeis import parser, utils
 
 # Real numbers
 x1 = np.random.rand(10)
@@ -16,7 +15,7 @@ circuit_string = "R1-[P2,R3]"
 p0_dict = {"R1": 250, "P2w": 1e-3, "P2n": 0.5, "R3": 10}
 p0_vals = list(p0_dict.values())
 circuit = CustomCircuit(
-    utils.impedancepy_circuit(circuit_string),
+    parser.impedancepy_circuit(circuit_string),
     initial_guess=p0_vals,
 )
 circuit.parameters_ = p0_vals
@@ -38,123 +37,22 @@ def test_fit_circuit_parameters_with_initial_guess():
     assert np.allclose(p_fit, p0_vals, rtol=0.01)
 
 
-def test_get_component_labels():
-    circuit = "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
-    # Pass default types = None (all components)
-    components = utils.get_component_labels(circuit)
-    components_gt = ["R1", "R2", "P12", "L2", "R22", "R6", "C7", "L8", "R5", "L9", "R3"]
-    assert components == components_gt
-    # Pass string as types
-    components = utils.get_component_labels(circuit, types="P")
-    components_gt = ["P12"]
-    assert components == components_gt
-    # Pass list as types
-    components = utils.get_component_labels(circuit, types=["L", "P"])
-    components_gt = ["P12", "L2", "L8", "L9"]
-    assert components == components_gt
-
-
-def test_get_component_types():
-    circuit = "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
-    types_gt = ["R", "R", "P", "L", "R", "R", "C", "L", "R", "L", "R"]
-    types = utils.get_component_types(circuit)
-    assert types == types_gt
-
-
-def test_get_parameter_labels():
-    circuit = "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
-    # Use default types = None (all parameters)
-    variables = utils.get_parameter_labels(circuit)
-    variables_gt = ["R1", "R2", "P12w", "P12n", "L2", "R22", "R6", "C7", "L8", "R5", "L9", "R3"]
-    assert variables == variables_gt
-    # Pass string as types
-    variables = utils.get_parameter_labels(circuit, types="R")
-    variables_gt = ["R1", "R2", "R22", "R6", "R5", "R3"]
-    assert variables == variables_gt
-    # Pass list as types
-    variables = utils.get_parameter_labels(circuit, types=["R", "P"])
-    variables_gt = ["R1", "R2", "P12w", "P12n", "R22", "R6", "R5", "R3"]
-    assert variables == variables_gt
-
-
-def test_count_parameters():
-    circuit = "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
-    num_params_gt = 12
-    num_params = utils.count_parameters(circuit)
-    assert num_params == num_params_gt
-    circuit = ""
-    num_params_gt = 0
-    num_params = utils.count_parameters(circuit)
-    assert num_params == num_params_gt
-    circuit = "[P1,P2]-P3-R4"
-    num_params_gt = 7
-    num_params = utils.count_parameters(circuit)
-    assert num_params == num_params_gt    
-
-
-def test_validate_circuit_string():
-    # Valid circuit
[…]
