+++ tests/test_parser.py
@@ -52,53 +52,121 @@ def test_get_parameter_types():
     assert types_unique == types_gt_unique
 
 
+def test_group_parameters_by_component():
+    circuit = "[R1,R2-P3]-L4-L5-[P6,R7]"
+    g = {
+        "R": ["R1", "R2", "R7"],
+        "P": ["P3w", "P3n", "P6w", "P6n"],
+        "L": ["L4", "L5"]
+    }
+    assert parser.group_parameters_by_component(circuit) == g
+
+
+def test_group_parameters_by_type():
+    circuit = "[R1,R2-P3]-L4-L5-[P6,R7]"
+    g = {
+        "R": ["R1", "R2", "R7"],
+        "Pw": ["P3w", "P6w"],
+        "Pn": ["P3n", "P6n"],
+        "L": ["L4", "L5"]
+    }
+    assert parser.group_parameters_by_type(circuit) == g
+
+
 def test_count_parameters():
-    circuit = "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
-    num_params_gt = 12
-    num_params = parser.count_parameters(circuit)
-    assert num_params == num_params_gt
-    circuit = ""
-    num_params_gt = 0
-    num_params = parser.count_parameters(circuit)
-    assert num_params == num_params_gt
-    circuit = "[P1,P2]-P3-R4"
-    num_params_gt = 7
-    num_params = parser.count_parameters(circuit)
-    assert num_params == num_params_gt    
+    d = {
+        "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3": 12,
+        "": 0,
+        "[P1,P2]-P3-R4": 7
+    }
+    for circuit, num_params_gt in d.items():
+        assert parser.count_parameters(circuit) == num_params_gt
 
 
 def test_validate_circuit():
-    # Valid circuit
-    circuit = "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
-    parser.validate_circuit(circuit)
-    # Empty circuit
-    circuit = ""
-    with pytest.raises(AssertionError):
+    # Valid circuits
+    circuits_valid = [
+        "[R1,R2-P12]-L2-R22-[R6,C7-[L8,R5],L9]-R3"
+    ]
+    for circuit in circuits_valid:
         parser.validate_circuit(circuit)
-    # Duplicate component
-    circuit = "[R1,R2]-R1"
-    with pytest.raises(AssertionError):
-        parser.validate_circuit(circuit)
+    # Invalid circuits
+    circuits_invalid = [
+        "",
+        "[R1,R2]-R1"
+    ]
+    for circuit in circuits_invalid:
+        with pytest.raises(AssertionError):
+            parser.validate_circuit(circuit)
 
+
+def test_validate_parameter():
+    # Valid parameters
+    parameters_valid = ["R1", "C5", "L12", "P1n", "P22w"]
+    for p in parameters_valid:
+        parser.validate_parameter(p)
+    # Invalid parameters
+    parameter_invalid = ["R1w", "C", "L2r", "Pn", "P2s"]
+    for p in parameter_invalid:
+        with pytest.raises(AssertionError):
+            parser.validate_parameter(p)
+
+
+def test_parse_component():
+    d = {
+        "R2": "R",
+        "C5": "C",
+        "L12": "L",
+        "P1n": "P",
+        "P22w": "P",
+        "P4": "P"
+    }
+    for k, v in d.items():
+        assert parser.parse_component(k) == v
+
+
[…]
