### 2023-12-21 dee4683
Add pytest config to pyproject.toml
