### 2023-12-21 536df41
Move `parse_ec_output` to io module
