### 2023-12-21 5074d7a
Include parser module in the docs
