### 2023-12-21 ac4ca0f
Increase timeout for parameter fitter to 60 seconds
