+++ _sources/generated/generated/autoeis.parser.convert_to_impedance_format.rst.txt
@@ -1,0 +1,6 @@
+convert\_to\_impedance\_format
+==============================
+
+.. currentmodule:: autoeis.parser
+
+.. autofunction:: convert_to_impedance_format
