+++ _sources/generated/autoeis.parser.rst.txt
@@ -1,0 +1,6 @@
+﻿:noindex:
+
+parser
+======
+
+.. automodule:: autoeis.parser
