+++ _sources/generated/generated/autoeis.parser.get_parameter_types.rst.txt
@@ -1,0 +1,6 @@
+get\_parameter\_types
+=====================
+
+.. currentmodule:: autoeis.parser
+
+.. autofunction:: get_parameter_types
