### 2023-12-21 2a08874
Remove bin folder, move scripts under assets
