### 2023-12-21 1364f15
Skip Windows nightly tests for Python 3.9
