+++ .github/workflows/nightly.yml
@@ -21,24 +21,29 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - name: "Set up Julia"
+        if: ${{ matrix.os != 'windows-latest' || matrix.python-version != '3.9' }}
         uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.julia-version }}
       - name: "Cache Julia"
+        if: ${{ matrix.os != 'windows-latest' || matrix.python-version != '3.9' }}
         uses: julia-actions/cache@v1
         with:
           cache-name: ${{ matrix.os }}-test-${{ matrix.julia-version }}-${{ matrix.python-version }}
           cache-packages: false
       - name: "Set up Python"
+        if: ${{ matrix.os != 'windows-latest' || matrix.python-version != '3.9' }}
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
           cache: pip
       - name: "Install AutoEIS"
+        if: ${{ matrix.os != 'windows-latest' || matrix.python-version != '3.9' }}
         run: |
             python -m pip install --upgrade pip
             pip install -e .[test]
             python -m autoeis install
       - name: "Run tests"
+        if: ${{ matrix.os != 'windows-latest' || matrix.python-version != '3.9' }}
         run: |
           pytest -v .
