### 2024-01-22 #6256 @editorialbot
[REVIEW]: AutoEIS: Automated equivalent circuit modeling from electrochemical impedance spectroscopy data using statistical machine learning
**Submitting author:** @ma-sadeghi (<a href="http://orcid.org/0000-0002-6756-9117">Mohammad Amin Sadeghi</a>)
**Repository:** https://github.com/AUTODIAL/AutoEIS/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.38
