### 2024-01-30 43164ea
Factor out num_chains as arg
