### 2024-01-30 9e09ac3
Better default values for num_warmup and num_samples
