### 2024-01-30 7f497c4
Remove unncessary postprocessings in `perform_bayesian_inference`
