### 2024-01-30 b69e755
Minor refactor
