### 2024-01-30 9f47da0
Better defaults when initializing priors
