### 2024-01-30 60bb65b
Factor out `dist_type` as arg in `initialize_priors_from_posteriors`
