+++ autoeis/utils.py
@@ -276,24 +276,37 @@ def initialize_priors(
 
 def initialize_priors_from_posteriors(
     posterior: dict[str, np.ndarray[float]],
-    variables: list[str]
+    variables: list[str],
+    dist_type: str = "lognormal",
 ) -> dict[str, dist.Distribution]:
     """Creates new priors based on the posterior distributions."""
     priors = {}
     for var in variables:
-        sample = posterior[var]
+        samples = posterior[var]
         # Fit data to a truncated normal distribution for exponents of CPE elements
         # HACK: for better convergence (fewer parameters), fit a normal and truncate it
         if "n" in var:
             # Exponent of CPE elements is bounded between 0 and 1
-            loc, scale = norm.fit(sample)
-            priors[var] = dist.TruncatedNormal(loc=loc, scale=2*scale, low=0, high=1)
+            loc, scale = stats.norm.fit(samples)
+            priors[var] = dist.TruncatedNormal(loc=loc, scale=1*scale, low=0, high=1)
         # Fit data to a log-normal distribution for all other parameters
         else:
             # NOTE: s and scale in scipy.stats -> scale and np.exp(loc) in numpyro
             # NOTE: above conversion is only valid when loc = 0
-            s, loc, scale = lognorm.fit(sample, floc=0)
-            priors[var] = dist.LogNormal(loc=np.log(scale), scale=2*s)
+            if dist_type == "lognormal":
+                s, loc, scale = stats.lognorm.fit(samples, floc=0)
+                priors[var] = dist.LogNormal(loc=np.log(scale), scale=8*s)
+            elif dist_type == "normal":
+                loc, scale = stats.norm.fit(samples)
+                priors[var] = dist.TruncatedNormal(loc=loc, scale=1*scale, low=0, high=np.inf)
+            elif dist_type == "weibull":
+                c, loc, scale = stats.weibull_min.fit(samples, floc=1)
+                priors[var] = dist.Weibull(scale=scale, concentration=c)
+            elif dist_type == "t":
+                df, loc, scale = stats.t.fit(samples)
+                priors[var] = dist.StudentT(df=df, loc=loc, scale=scale)
+            else:
+                raise ValueError(f"Unknown distribution: {dist_type}")
     return priors
 
 # <<< Statistics utils
