### 2024-01-30 209dbfe
Add models.py to host acquisition functions
