### 2024-01-30 f173b78
Bump Julia version to 1.10, test on Apple silicon as well
