+++ .github/workflows/tests.yml
@@ -10,9 +10,9 @@ jobs:
         shell: bash
     strategy:
       matrix:
-        julia-version: ['1.9']
+        julia-version: ['1.10']
         python-version: ['3.10']
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-latest, macos-latest, macost-14, windows-latest]
 
     steps:
       - uses: actions/checkout@v4
