### 2024-01-30 cd9b7dd
Bump version to v0.0.19
