### 2024-01-30 ec382b7
Augment circuit fitter to accept iters as argument
