### 2024-01-31 c3bf59d
Implement _generate_ecm_parallel2 based on multiprocessing
