### 2024-01-31 6500523
Don't assert backend version if installed from GitHub branch
