### 2024-01-31 fb392d5
Don't redirect Julia's stdout/stderr (leads to cryptic JuliaIO error)
