### 2024-01-31 d751d17
Add unit tests for GEP functions
