+++ tests/test_core.py
@@ -1,6 +1,7 @@
 import numpy as np
+import pandas as pd
 
-from autoeis import core, utils
+from autoeis import core, io, utils
 
 
 def test_compute_ohmic_resistance():
@@ -25,3 +26,33 @@ def test_compute_ohmic_resistance_missing_high_freq():
     # When high frequency measurements are missing, Re(Z) @ max(freq) is good approximation
     Zreal_at_high_freq = Z.real[np.argmax(freq)]
     np.testing.assert_allclose(R, Zreal_at_high_freq)
+
+
+def test_gep_serial():
+    Z, freq = io.load_test_dataset()
+    Z, freq, rmse = core.preprocess_impedance_data(Z, freq, threshold=5e-2)
+    kwargs = {
+        "iters": 2,
+        "complexity": 12,
+        "population_size": 10,
+        "generations": 5,
+        "tol": 1e-2,
+        "parallel": False,
+    }
+    circuits = core.generate_equivalent_circuits(Z, freq, **kwargs)
+    assert isinstance(circuits, pd.DataFrame)
+
+
+def test_gep_parallel():
+    Z, freq = io.load_test_dataset()
+    Z, freq, rmse = core.preprocess_impedance_data(Z, freq, threshold=5e-2)
+    kwargs = {
+        "iters": 2,
+        "complexity": 12,
+        "population_size": 10,
+        "generations": 5,
+        "tol": 1e-2,
+        "parallel": True,
+    }
+    circuits = core.generate_equivalent_circuits(Z, freq, **kwargs)
+    assert isinstance(circuits, pd.DataFrame)
