+++ autoeis/version.py
@@ -1,2 +1,2 @@
-__version__ = "0.0.19"
+__version__ = "0.0.20"
 __equivalent_circuits_jl_version__ = "0.3.1"
