### 2024-02-01 3bd3648
Rename `apply_heuristic_rules` to `filter_implausible_circuits` for clarity
