### 2024-02-01 828794e
Expose hard-coded parameters as argument in `perform_full_analysis`
