+++ autoeis/core.py
@@ -109,6 +109,7 @@ def compute_ohmic_resistance(Z: np.ndarray[complex], freq: np.ndarray[float]) ->
     return R
 
 
+# TODO: Needs heavy refactoring
 def preprocess_impedance_data(
     impedance: np.ndarray[complex],
     freq: np.ndarray[float],
@@ -259,7 +260,7 @@ def generate_equivalent_circuits(
     population_size: int = 100,
     seed: int = None,
 ) -> pd.DataFrame:
-    """Generate potential ECMs using evolutionary algorithms.
+    """Generates candidate circuits that fit impedance data using genetic algorithm.
 
     Parameters
     ----------
@@ -285,7 +286,7 @@ def generate_equivalent_circuits(
     Returns
     -------
     pd.DataFrame or None
-        DataFrame containing ECM solutions or None if no solutions are found.
+        Dataframe containing circuits.
     """
     log.info("Generating equivalent circuits via evolutionary algorithms.")
 
@@ -313,7 +314,7 @@ def generate_equivalent_circuits(
 
 
 def _generate_ecm_serial(impedance, freq, iters, ec_kwargs, seed):
-    """Generates candidate circuit models in serial."""
+    """Generates candidate circuits in serial."""
     Main = julia_helpers.init_julia()
     ec = julia_helpers.import_backend(Main)
 
@@ -340,7 +341,7 @@ def _generate_ecm_serial(impedance, freq, iters, ec_kwargs, seed):
 
 # TODO: This function is deprecated, use _generate_ecm_parallel_julia instead
 def _generate_ecm_parallel_mpire(impedance, freq, iters, ec_kwargs, seed):
-    """Generates candidate circuit models in parallel via Python multiprocessing."""
+    """Generates candidate circuits in parallel via Python multiprocessing."""
 
     def circuit_evolution(seed: int):
         """Closure to generate a single circuit to be used with multiprocessing."""
@@ -392,7 +393,7 @@ def _generate_ecm_parallel_mpire(impedance, freq, iters, ec_kwargs, seed):
 
 
 def _generate_ecm_parallel_julia(impedance, freq, iters, ec_kwargs, seed):
-    """Generates candidate circuit models in parallel directly from Julia."""
+    """Generates candidate circuits in parallel using Julia multiprocessing."""
     Main = julia_helpers.init_julia()
     # Set random seed for reproducibility (Python and Julia)
     # FIXME: This doesn't work when multiprocessing, use @everywhere instead
@@ -438,7 +439,7 @@ def _generate_ecm_parallel_julia(impedance, freq, iters, ec_kwargs, seed):
 
 
 def split_components(circuits: pd.DataFrame) -> pd.DataFrame:
-    """Adds an individual column for each component in the circuit with its value."""
+    """Augments the circuits dataframe with additional columns for each component."""
     # Initialize lists to populate the component columns later
     components = {"R": [], "C": [], "L": [], "P": []}
     labels = {"R": "Resistors", "C": "Capacitors", "L": "Inductors", "P": "CPEs"}
@@ -459,18 +460,7 @@ def split_components(circuits: pd.DataFrame) -> pd.DataFrame:
 
 
 def capacitance_filter(circuits: pd.DataFrame) -> pd.DataFrame:
-    """Exclude ideal capacitors from the circuits dataframe.
-
-    Parameters
-    ----------
-    circuits: pd.DataFrame
-       Dataframe containing ECMs (6 columns)
-
-    Returns
-    -------
-    df_circuits: pd.DataFrame
-       Dataframe containing ECMs without ideal capacitors (6 columns)
-    """
+    """Excludes ideal capacitors from the circuits dataframe."""
     circuits = circuits.copy()
 
[…]
