### 2024-02-01 4d0db1b
`filter_implausible_circuits` shouldn't modify dataframe columns
