+++ autoeis/core.py
@@ -67,13 +67,13 @@ def compute_ohmic_resistance(Z: np.ndarray[complex], freq: np.ndarray[float]) ->
     Z : np.ndarray[complex]
         Impedance measurements.
     freq : np.ndarray[float]
-        Frequencies corresponding to impedance measurements.        
+        Frequencies corresponding to impedance measurements.
 
     Returns
     -------
     ohmic_resistance: float
         The ohmic resistance of impedance data.
-    
+
     Raises
     ------
     ValueError
@@ -91,8 +91,8 @@ def compute_ohmic_resistance(Z: np.ndarray[complex], freq: np.ndarray[float]) ->
     y = 1 / Z.real[:]
 
     def func(x, a, b, c):
-        return (a*x + b) / (x + c)
-    
+        return (a * x + b) / (x + c)
+
     fallback = False
     bounds = ([0, -np.inf, -np.inf], np.inf)
 
@@ -101,7 +101,7 @@ def compute_ohmic_resistance(Z: np.ndarray[complex], freq: np.ndarray[float]) ->
         R = 1 / a
     except RuntimeError:
         fallback = True
-    if R < 0 or R < 0.2*Z.real[0]:
+    if R < 0 or R < 0.2 * Z.real[0]:
         fallback = True
     if fallback:
         R = Z.real[0]
@@ -228,7 +228,9 @@ def preprocess_impedance_data(
             ohmic_resistance = compute_ohmic_resistance(Z_mask, freq_mask)
             ohmic_resistance_found = True
         except ValueError:
-            log.error("Ohmic resistance not found. Check data or increase KK threshold.")
+            log.error(
+                "Ohmic resistance not found. Check data or increase KK threshold."
+            )
             ohmic_resistance_found = False
 
         # Convert the data to a dataframe for easier manipulation
@@ -292,7 +294,7 @@ def generate_equivalent_circuits(
 
     # Set the seed for reproducibility (if not set, use current time in nanoseconds)
     seed = seed or time.time_ns() % 2**32
-        
+
     ec_kwargs = {
         "head": complexity,
         "terminals": "RLP",
@@ -320,8 +322,8 @@ def _generate_ecm_serial(impedance, freq, iters, ec_kwargs, seed):
 
     # Set random seed for reproducibility (Python and Julia)
     np.random.seed(seed)
-    Main.eval(f"import Random; Random.seed!({seed})")            
-    
+    Main.eval(f"import Random; Random.seed!({seed})")
+
     circuits = []
     for _ in tqdm(range(iters), desc="Circuit Evolution"):
         try:
@@ -370,7 +372,7 @@ def _generate_ecm_parallel_mpire(impedance, freq, iters, ec_kwargs, seed):
 
     # Julia cannot be initialized in the main process -> guard against this
     runtime_error = False
-    
+
     # Set a different seed for each process
     seed = [seed + i for i in range(iters)]
 
@@ -381,7 +383,9 @@ def _generate_ecm_parallel_mpire(impedance, freq, iters, ec_kwargs, seed):
             runtime_error = True
 
     if runtime_error:
-        raise RuntimeError("Julia must not be manually initialized, restart the kernel.")
+        raise RuntimeError(
+            "Julia must not be manually initialized, restart the kernel."
+        )
 
     # Remove None values
     circuits = [circuit for circuit in circuits if circuit is not None]
@@ -417,7 +421,9 @@ def _generate_ecm_parallel_julia(impedance, freq, iters, ec_kwargs, seed):
         sys.stderr.flush()
         for iters_ in iters_chunked:
             try:
[…]
