+++ autoeis/utils.py
@@ -39,13 +39,14 @@ from autoeis import parser
 
 # >>> Logging utils
 
+
 def get_logger(name: str) -> logging.Logger:
     """Get a logger with the given name."""
     logger = logging.getLogger(name)
     if logger.hasHandlers():
         # If logger has handlers, do not add another to avoid duplicate logs
         return logger
-    
+
     logger.setLevel(logging.WARNING)
     handler = RichHandler(rich_tracebacks=True)
     handler.setFormatter(logging.Formatter("%(message)s", datefmt="[%X]"))
@@ -57,13 +58,16 @@ def setup_rich_tracebacks():
     """Set up rich traceback for nicer exception printing."""
     rich.traceback.install()
 
+
 # <<< Logging utils
 
 
 # >>> General utils
 
+
 def flatten(xs: list) -> list:
     """Returns a list of all elements in a nested iterable."""
+
     def _flatten(xs):
         """Returns a generator that flattens a nested iterable."""
         for x in xs:
@@ -71,6 +75,7 @@ def flatten(xs: list) -> list:
                 yield from flatten(x)
             else:
                 yield x
+
     return list(_flatten(xs))
 
 
@@ -81,7 +86,7 @@ def find_identical_rows(a: np.ndarray) -> list[list[int]]:
     for i in range(a.shape[0]):
         if i not in flatten(idx):
             idx.append([i])
-        for j in range(i+1, a.shape[0]):
+        for j in range(i + 1, a.shape[0]):
             if np.allclose(a[i, :], a[j, :]):
                 idx[-1].append(j)
     return idx
@@ -91,8 +96,8 @@ class _SuppressOutput:
     def __enter__(self):
         self._original_stdout = sys.stdout
         self._original_stderr = sys.stderr
-        sys.stdout = open(os.devnull, 'w')
-        sys.stderr = open(os.devnull, 'w')
+        sys.stdout = open(os.devnull, "w")
+        sys.stderr = open(os.devnull, "w")
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         sys.stdout.close()
@@ -103,18 +108,22 @@ class _SuppressOutput:
 
 def suppress_output(func):
     """Suppresses the output of a function."""
+
     @wraps(func)
     def wrapped(*args, **kwargs):
         with _SuppressOutput():
             return func(*args, **kwargs)
+
     return wrapped
 
 
 class TimeoutException(Exception):
     pass
 
+
 def timeout(seconds):
     """Raises a TimeoutException if decorated function doesn't return in time."""
+
     def decorator(func):
         def _handle_timeout(signum, frame):
             raise TimeoutException("Didn't converge in time!")
@@ -132,19 +141,22 @@ def timeout(seconds):
             return result
 
         return wrapper
+
     return decorator
+
 
 # <<< General utils
 
 
 # >>> Circuit utils
 
+
 def fit_circuit_parameters(
     circuit: str,
     Z: np.ndarray[complex],
     freq: np.ndarray[float],
     p0: Union[np.ndarray[float], dict[str, float]] = None,
-    iters: int = 1
+    iters: int = 1,
 ) -> dict[str, float]:
     """Fits a circuit to impedance data and returns the parameters."""
     # Deal with initial guess
@@ -157,14 +169,13 @@ def fit_circuit_parameters(
 
     # Fit circuit parameters
     circuit_impy = CustomCircuit(
-        circuit=parser.convert_to_impedance_format(circuit),
-        initial_guess=p0
+        circuit=parser.convert_to_impedance_format(circuit), initial_guess=p0
     )
     # HACK: Use multiple random initial guesses to avoid local minima
     err_min = np.inf
     for _ in range(iters):
[…]
