### 2024-02-01 13c0052
Add validate_circuits_dataframe helper to utils
