### 2024-02-01 85690b1
`perform_bayesian_inference` now returns exit code to handle MCMC failure
