### 2024-02-01 bb935da
Rename Numpyro models for clarity
