+++ pyproject.toml
@@ -81,3 +81,6 @@ testpaths = [
     "tests",
     "integration",
 ]
+
+[tool.ruff]
+line-length = 88
