### 2024-02-01 c815190
Fix type hint syntax
