### 2024-02-01 0566f53
Fix error handling in perform_bayesian_inference function
