### 2024-02-02 1b393a0
Add load_test_circuits helper to load toy circuits
