### 2024-02-02 f28b126
No need to set numpy random seed, jax is enough
