### 2024-02-02 613c659
Refactor of circuit function to be jittable!
