+++ autoeis/__init__.py
@@ -2,6 +2,3 @@ from . import core, io, metrics, parser, utils, visualization  # noqa: F401
 from .core import *
 from .version import __equivalent_circuits_jl_version__, __version__  # noqa: F401
 from .visualization import rich_print  # noqa: F401
-
-# Set up rich traceback for nicer exception printing
-utils.setup_rich_tracebacks()
