### 2024-02-02 498d995
Fix bug in circuit validation where required_columns were being overriden
