### 2024-02-02 aa81b47
Add parallelized version of inference
