### 2024-02-02 5eebbe1
Log warning when no circuits survive the post filteration
