### 2024-02-02 5d4eaec
Use the parallelized version throughout the package
