### 2024-02-02 60fd1f0
Update mpire dependency to use dill (pickle breaks multiprocessing)
