+++ _sources/generated/generated/autoeis.core.perform_bayesian_inference_batch.rst.txt
@@ -1,0 +1,6 @@
+perform\_bayesian\_inference\_batch
+===================================
+
+.. currentmodule:: autoeis.core
+
+.. autofunction:: perform_bayesian_inference_batch
