### 2024-02-04 03d7efa
Make defaults consistent across functions
