+++ autoeis/core.py
@@ -520,7 +520,7 @@ def perform_bayesian_inference(
     Z: np.ndarray[complex],
     freq: np.ndarray[float],
     p0: Union[np.ndarray[float], dict[str, float]] = None,
-    num_warmup=1000,
+    num_warmup=2500,
     num_samples=1000,
     num_chains=1,
     seed: Union[int, jax.Array] = None,
