### 2024-02-04 489b3d3
Unify serial and parallel inference helpers
