### 2024-02-04 c263032
Update `perform_full_analysis` to use the unified API
