### 2024-02-04 7d4b929
Add github action to test notebooks
