+++ .github/workflows/notebooks.yml
@@ -1,0 +1,40 @@
+name: Notebooks
+
+on: pull_request
+
+jobs:
+  test:
+    runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash
+    strategy:
+      matrix:
+        julia-version: ['1.10']
+        python-version: ['3.10']
+        os: [ubuntu-latest]
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: "Set up Julia"
+        uses: julia-actions/setup-julia@v1
+        with:
+          version: ${{ matrix.julia-version }}
+      - name: "Cache Julia"
+        uses: julia-actions/cache@v1
+        with:
+          cache-name: ${{ matrix.os }}-test-${{ matrix.julia-version }}-${{ matrix.python-version }}
+          cache-packages: false
+      - name: "Set up Python"
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: pip
+      - name: "Install AutoEIS"
+        run: |
+            python -m pip install --upgrade pip
+            pip install -e .[test]
+            python -m autoeis install
+      - name: "Run tests"
+        run: |
+          pytest -v examples/*.ipynb
