### 2024-02-04 802cd3c
Change default iters in docs to 24
