### 2024-02-04 4193cb0
Revert back to the old `import_package` implementation
