+++ .github/workflows/notebooks.yml
@@ -37,4 +37,4 @@ jobs:
             python -m autoeis install
       - name: "Run tests"
         run: |
-          pytest -v examples/*.ipynb
+          pytest -v --nbmake examples/*.ipynb
