### 2024-02-05 2bf4958
Refactor max-retries when importing packages on Windows
