+++ .github/workflows/nightly.yml
@@ -46,4 +46,4 @@ jobs:
       - name: "Run tests"
         if: ${{ matrix.os != 'windows-latest' || matrix.python-version != '3.9' }}
         run: |
-          pytest -v .
+          pytest -v tests/
