+++ .github/workflows/conda.yml
@@ -1,0 +1,34 @@
+name: Conda
+
+on: pull_request
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        shell: bash
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: "Set up Julia"
+        uses: julia-actions/setup-julia@v1
+        with:
+          version: '1.10'
+      - name: "Cache Julia"
+        uses: julia-actions/cache@v1
+        with:
+          cache-name: 'ubuntu-latest-test-1.10-3.10'
+          cache-packages: false
+      - name: "Set up Python"
+        uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniforge-version: latest
+      - name: "Install AutoEIS"
+        run: |
+            python -m pip install --upgrade pip
+            pip install -e .[test]
+            python -m autoeis install
+      - name: "Run tests"
+        run: |
+          pytest -v tests/
