### 2024-02-09 427fce7
Use Python 3.10 for conda CI
