### 2024-02-09 7e793ce
Add function to identify Python runtime environment
