### 2024-02-09 fb0582d
Remove ruff-related lines from pyproject.toml in favor of ruff.toml
