### 2024-02-09 ad8104b
Initialize Julia and import EquivalentCircuits on first import
