### 2024-02-09 4bc3504
Make Main mandatory arg in julia helper functions
