### 2024-02-09 c425117
Catch Julia's ImportError the first time package is installed
