+++ autoeis/julia_helpers_pyjulia.py
@@ -1,0 +1,368 @@
+# This code is based on the file julia_helpers.py from the PySR project.
+# It has been adapted to suit the requirements of AutoEIS.
+# The original file is under the Apache 2.0 license.
+# Acknowledge the original authors and license when utilizing this code.
+# Original repository: https://github.com/MilesCranmer/PySR
+# Commit reference: 976f8d8 dated 2023-09-16.
+
+import importlib
+import os
+import subprocess
+import sys
+from pathlib import Path
+
+from julia.api import JuliaError
+
+import autoeis.utils as utils
+
+from .version import __equivalent_circuits_jl_version__, __version__
+
+MAX_RETRIES = 2
+juliainfo = None
+julia_initialized = False
+julia_kwargs_at_initialization = None
+julia_activated_env = None
+
+log = utils.get_logger(__name__)
+# TODO: For virtualenvs see https://github.com/JuliaPy/PyCall.jl?tab=readme-ov-file#python-virtual-environments
+
+
+def install(
+    julia_project=None, quiet=False, precompile=None, offline=False
+):  # pragma: no cover
+    """Install all required dependencies for EquivalentCircuits.jl."""
+    import julia
+
+    _julia_version_assertion()
+    # Set JULIA_PROJECT so that we install in the autoeis environment
+    processed_julia_project, is_shared = _process_julia_project(julia_project)
+    _set_julia_project_env(processed_julia_project, is_shared)
+
+    if precompile is False:
+        os.environ["JULIA_PKG_PRECOMPILE_AUTO"] = "0"
+
+    if offline:
+        os.environ["JULIA_PKG_OFFLINE"] = "true"
+
+    try:
+        julia.install(quiet=quiet)
+    except julia.tools.PyCallInstallError:
+        # Attempt to reset PyCall.jl's build:
+        subprocess.run(
+            [
+                "julia",
+                "-e",
+                f'ENV["PYTHON"] = "{sys.executable}"; import Pkg; Pkg.build("PyCall")',
+            ],
+        )
+        # Try installing again:
+        julia.install(quiet=quiet)
+
+    Main, init_log = init_julia(julia_project, quiet=quiet, return_aux=True)
+    io_arg = _get_io_arg(quiet)
+
+    if precompile is None:
+        precompile = init_log["compiled_modules"]
+
+    if not precompile:
+        Main.eval('ENV["JULIA_PKG_PRECOMPILE_AUTO"] = 0')
+
+    if is_shared:
+        # Install EquivalentCircuits.jl:
+        _add_ec_to_julia_project(Main, io_arg)
+
+    Main.eval("using Pkg")
+    Main.eval(f"Pkg.instantiate({io_arg})")
+
+    if precompile:
+        Main.eval(f"Pkg.precompile({io_arg})")
+
+    if not quiet:
+        log.warning(
+            "It is recommended to restart Python after installing AutoEIS "
+            "dependencies, so that the Julia environment is properly initialized."
+        )
+
+
+def init_julia(julia_project=None, quiet=False, julia_kwargs=None, return_aux=False):
+    """Initialize julia binary, turning off compiled modules if needed."""
+    global julia_initialized
+    global julia_kwargs_at_initialization
+    global julia_activated_env
+
+    if julia_kwargs is None:
+        julia_kwargs = {"optimize": 3}
+
+    from julia.core import JuliaInfo, UnsupportedPythonError
+
+    _julia_version_assertion()
+    processed_julia_project, is_shared = _process_julia_project(julia_project)
+    _set_julia_project_env(processed_julia_project, is_shared)
+
+    try:
+        info = JuliaInfo.load(julia="julia")
+    except FileNotFoundError:
[…]
