### 2024-02-09 9b88c69
Rewrite cli and julia helpers based on juliacall and juliapkg
