+++ autoeis/cli.py
@@ -10,29 +10,5 @@ def autoeis_installer(context):
 
 
 @autoeis_installer.command("install", help="Install Julia dependencies for AutoEIS.")
-@click.option(
-    "-p",
-    "julia_project",
-    "--project",
-    default=None,
-    type=str,
-    help="Install in a specific Julia project (e.g., a local copy of EquivalentProject.jl).",
-    metavar="PROJECT_DIRECTORY",
-)
-@click.option("-q", "--quiet", is_flag=True, default=False, help="Disable logging.")
-@click.option(
-    "--precompile",
-    "precompile",
-    flag_value=True,
-    default=None,
-    help="Force precompilation of Julia libraries.",
-)
-@click.option(
-    "--no-precompile",
-    "precompile",
-    flag_value=False,
-    default=None,
-    help="Disable precompilation.",
-)
-def install_cli(julia_project, quiet, precompile):
-    install(julia_project=julia_project, quiet=quiet, precompile=precompile)
+def install_cli():
+    install()
