### 2024-02-09 d7dc77e
Add option for backend installer to install a local copy of EquivalentCircuits for development
