+++ autoeis/cli.py
@@ -9,6 +9,12 @@ def autoeis_installer(context):
     ctx = context
 
 
+@click.option(
+    "--ec-path",
+    default=None,
+    type=str,
+    help="Installs a local copy of EquivalentCircuits instead of the remote version.",
+)
 @autoeis_installer.command("install", help="Install Julia dependencies for AutoEIS.")
-def install_cli():
-    install()
+def install_cli(ec_path):
+    install(ec_path=ec_path)
