### 2024-02-11 54b9fdb
Ensure unit tests find valid GEPs (used to pass even when no circuit was found)
