### 2024-02-11 1d862c2
Raise ImportError instead of assert when Julia/backend not found
