### 2024-02-11 223532b
Fix pep8, minor refactor
