### 2024-02-11 5776f83
Refactor `parse_ec_output` to accept list[str] instead of dataframe
