### 2024-02-11 636ad6e
Convert EquivalentCircuits output parser to dataframe
