### 2024-02-11 00d69ba
Circuit generator wrappers should all return list[str], EquivalentCircuits parser takes care of conversion
