### 2024-02-11 37ad1d0
Use my patched up fork until EquivalentCircuits is patched
