+++ autoeis/version.py
@@ -1,3 +1,3 @@
 __version__ = "0.0.23"
 # Either use a branch name or a tag name for EquivalentCircuits.jl
-__equivalent_circuits_jl_version__ = "master"  # "0.3.1"
+__equivalent_circuits_jl_version__ = "bug/fix-ndarray-serialization"  # "0.3.1"
