### 2024-02-11 0bd636e
Avoid importing autoeis when not needed
