+++ autoeis/julia_helpers.py
@@ -1,7 +1,7 @@
 import shutil
 
-from autoeis.utils import get_logger
-from autoeis.version import __equivalent_circuits_jl_version__
+from .utils import get_logger
+from .version import __equivalent_circuits_jl_version__
 
 log = get_logger(__name__)
 
