### 2024-02-11 6e037c9
Automatically install Julia and dependencies if not found
