+++ autoeis/julia_helpers.py
@@ -6,9 +6,13 @@ from .version import __equivalent_circuits_jl_version__
 log = get_logger(__name__)
 
 
-def install(ec_path=None):
+def install(ec_path=None, install_julia=True):
     """Installs Julia dependencies for AutoEIS."""
-    assert_julia_installed()
+    if not install_julia:
+        is_julia_installed(error=True)
+    if not is_julia_installed(error=False):
+        log.warning("Julia not found in PATH, installing Julia...")
+    # The next line automatically installs Julia if it is not found
     import juliapkg
 
     kwargs = {"name": "EquivalentCircuits", "uuid": "da5bd070-f609-4e16-a30d-de86b3faa756"}
@@ -28,43 +32,50 @@ def install(ec_path=None):
 
 def init_julia():
     """Initializes Julia and returns the Main module."""
-    assert_julia_installed()
+    is_julia_installed()
     import juliacall
 
     return juliacall.Main
 
 
-def import_package(package_name, Main):
+def import_package(package_name, Main, error=False):
     """Imports a package in Julia and returns the module."""
     from juliacall import JuliaError
 
     try:
         Main.seval(f"using {package_name}")
         return eval(f"Main.{package_name}")
-    except JuliaError:
-        return None
+    except JuliaError as e:
+        if error:
+            raise e
+    return None
 
 
 def import_backend(Main=None):
     """Imports EquivalentCircuits package from Julia."""
+    is_julia_installed(error=True)
     Main = init_julia() if Main is None else Main
-    assert_backend_installed()
+    is_backend_installed(error=True)
     return import_package("EquivalentCircuits", Main)
 
 
-def assert_julia_installed():
+def is_julia_installed(error=False):
     """Asserts that Julia is installed."""
+    if shutil.which("julia"):
+        return True
     msg = "Julia not found. Visit https://github.com/JuliaLang/juliaup and install Julia."
-    if shutil.which("julia") is None:
+    if error:
         raise ImportError(msg)
-    # return shutil.which("julia") is not None
+    return False
 
 
-def assert_backend_installed(Main=None):
+def is_backend_installed(Main=None, error=False):
     """Asserts that EquivalentCircuits.jl is installed."""
-    assert_julia_installed()
+    is_julia_installed(error=True)
     Main = init_julia() if Main is None else Main
-    msg = "EquivalentCircuits.jl not installed."
-    if import_package("EquivalentCircuits", Main) is None:
+    if import_package("EquivalentCircuits", Main) is not None:
+        return True
+    msg = "EquivalentCircuits.jl not found, run 'python -m autoeis install'"
+    if error:
         raise ImportError(msg)
-    # return import_package("EquivalentCircuits", Main) is not None
+    return False
