### 2024-02-11 84bdd75
Try different juliapkg API to see if Julia has been already installed
