+++ autoeis/cli.py
@@ -1,6 +1,6 @@
 import click
 
-from .julia_helpers import install
+from .julia_helpers import install_backend, install_julia
 
 
 @click.group("autoeis")
@@ -17,4 +17,5 @@ def autoeis_installer(context):
 )
 @autoeis_installer.command("install", help="Install Julia dependencies for AutoEIS.")
 def install_cli(ec_path):
-    install(ec_path=ec_path)
+    install_julia()
+    install_backend(ec_path=ec_path)
