### 2024-02-11 2adfa0a
init_julia should throw error if Julia is not found
