### 2024-02-11 2ae5499
Minor refactor
