### 2024-02-12 ad55ada
Compare against Python None instead of Julia nothing
