### 2024-02-12 55b997d
Bump version to v0.0.24
