### 2024-02-12 181f2a5
Add Apple M-chip machines to CI [no ci]
