+++ .github/workflows/nightly.yml
@@ -16,7 +16,7 @@ jobs:
       matrix:
         julia-version: ['1.10']
         python-version: ['3.9', '3.10', '3.11']
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-latest, macos-latest, macos-14, windows-latest]
 
     steps:
       - uses: actions/checkout@v4
