### 2024-02-12 0a39d15
Make Julia executable look-up more robust
