### 2024-02-13 ceb06b5
Add type hint to circuit fn factory
