### 2024-02-13 66b7b74
Expose `maxfev ` as arg in circuit fitter
