### 2024-02-13 f0d4a74
Don't use multiprocessing for single circuit inputs
