### 2024-02-13 d4e98b4
Don't spawn more workers than the number of circuits
