### 2024-02-13 a3127d3
Add type hint
