### 2024-02-13 507b6bb
Use literal p instead of x when in circuit expression
