+++ .github/workflows/nightly.yml
@@ -15,7 +15,7 @@ jobs:
     strategy:
       matrix:
         julia-version: ['1.10']
-        python-version: ['3.9', '3.10', '3.11']
+        python-version: ['3.10', '3.11', '3.12']
         os: [ubuntu-latest, macos-latest, macos-14, windows-latest]
 
     steps:
