### 2024-02-13 2516a46
Use M1 instead of Ubuntu runners for notebook and docs CI for speed
