+++ .github/workflows/gh-pages.yml
@@ -9,7 +9,7 @@ on:
 jobs:
   deploy-docs:
 
-    runs-on: ubuntu-latest
+    runs-on: macos-14
 
     steps:
       - uses: actions/checkout@v4
