+++ .github/workflows/conda.yml
@@ -4,7 +4,7 @@ on: pull_request
 
 jobs:
   test:
-    runs-on: ubuntu-latest
+    runs-on: macos-14
     defaults:
       run:
         shell: bash
