### 2024-02-13 3c6dc10
Remove a strange if statement in CI workflows
