+++ .github/workflows/conda.yml
@@ -16,6 +16,7 @@ jobs:
         with:
           python-version: '3.10'
           miniforge-version: latest
+          installer-url: https://github.com/conda-forge/miniforge/releases/download/23.11.0-0/Mambaforge-23.11.0-0-MacOSX-arm64.sh
       - name: "Install AutoEIS"
         run: |
             python -m pip install --upgrade pip
