### 2024-02-13 f40521a
Set default shell for conda runner
