### 2024-02-13 9b88759
Refactor generate_circuit_fn to optionally return concatenated impedance (Re, Im)
