### 2024-02-13 880dc79
Add brand new circuit fitter ~20x faster than impedance.py (keep as legacy)
