### 2024-02-13 16a6344
Fix a bug in parse_initial_guess where it didn't account for ndarrays as valid input
