### 2024-02-13 1464ece
Use parse_initial_guess in circuit fitter
