### 2024-02-13 b9753da
Fix incorrectly defined objective function in new circuit fitter
