### 2024-02-13 527e7ae
Standardize circuit function signature
