+++ tests/test_utils.py
@@ -15,11 +15,11 @@ p0_dict = {"R1": 250, "P2w": 1e-3, "P2n": 0.5, "R3": 10}
 p0_vals = list(p0_dict.values())
 circuit_fn_gt = utils.generate_circuit_fn_impedance_backend(circuit_string)
 freq = np.logspace(-3, 3, 1000)
-Z = circuit_fn_gt(p0_vals, freq)
+Z = circuit_fn_gt(freq, p0_vals)
 
 
 def test_fit_circuit_parameters_without_x0():
-    p_dict = utils.fit_circuit_parameters(circuit_string, Z, freq, iters=5)
+    p_dict = utils.fit_circuit_parameters(circuit_string, freq, Z, iters=10)
     p_fit = list(p_dict.values())
     assert np.allclose(p_fit, p0_vals, rtol=0.01)
 
@@ -27,7 +27,7 @@ def test_fit_circuit_parameters_without_x0():
 def test_fit_circuit_parameters_with_x0():
     # Add some noise to the initial guess to test robustness
     p0 = p0_vals + np.random.rand(len(p0_vals)) * p0_vals * 0.5
-    p_dict = utils.fit_circuit_parameters(circuit_string, Z, freq, p0)
+    p_dict = utils.fit_circuit_parameters(circuit_string, freq, Z, p0)
     p_fit = list(p_dict.values())
     assert np.allclose(p_fit, p0_vals, rtol=0.01)
 
@@ -38,7 +38,7 @@ def test_generate_circuit_fn():
     freq = np.array([1, 10, 100])
     p = np.random.rand(num_params)
     circuit_fn = utils.generate_circuit_fn(circuit)
-    Z_py = circuit_fn(p, freq)
+    Z_py = circuit_fn(freq, p)
     Main = julia_helpers.init_julia()
     ec = julia_helpers.import_backend(Main)
     Z_jl = np.array([ec.get_target_impedance(circuit, p, f) for f in freq])
