### 2024-02-13 445c599
Enforce double precision for circuit fitter to work
