### 2024-02-13 6eb55de
Enforce float as dtype of initial guess
