### 2024-02-13 9ec9006
Refactor circuit fitter
