### 2024-02-13 e0f72e7
Optionally refine the initial guess before inference
