### 2024-02-13 e218124
Add docstrings for perform_bayesian_inference
