### 2024-02-13 6461909
Move parallel BI to top of the file, for some reason CI fails otherwise!
