+++ tests/test_core.py
@@ -6,6 +6,17 @@ import pytest
 from autoeis import core, io, utils
 
 
+def test_bayesian_inference_batch():
+    Z, freq = io.load_test_dataset()
+    # Only test first three circuits to save time in CI
+    circuits = io.load_test_circuits(filtered=True).iloc[:3]
+    mcmc_results = core.perform_bayesian_inference(circuits, freq, Z, refine_p0=True)
+    assert len(mcmc_results) == len(circuits)
+    for mcmc, exist_code in mcmc_results:
+        assert exist_code in [-1, 0]
+        assert isinstance(mcmc, numpyro.infer.mcmc.MCMC)
+
+
 def test_compute_ohmic_resistance():
     circuit_string = "R1-[P2,P3-R4]"
     circuit_fn = utils.generate_circuit_fn_impedance_backend(circuit_string)
@@ -81,21 +92,10 @@ def test_bayesian_inference_single():
         "num_samples": 1000,
         "progress_bar": False,
     }
-    mcmc, exist_code = core._perform_bayesian_inference(
-        circuit, freq, Z, p0, **kwargs_mcmc
-    )
-    assert exist_code in [-1, 0]
+    mcmcs = core.perform_bayesian_inference(circuit, freq, Z, p0, **kwargs_mcmc)
+    mcmc, exit_code = mcmcs[0]
+    assert exit_code in [-1, 0]
     assert isinstance(mcmc, numpyro.infer.mcmc.MCMC)
-
-
-def test_bayesian_inference_batch():
-    Z, freq = io.load_test_dataset()
-    circuits = io.load_test_circuits(filtered=True)
-    mcmc_results = core.perform_bayesian_inference(circuits, freq, Z)
-    assert len(mcmc_results) == len(circuits)
-    for mcmc, exist_code in mcmc_results:
-        assert exist_code in [-1, 0]
-        assert isinstance(mcmc, numpyro.infer.mcmc.MCMC)
 
 
 @pytest.mark.skip(reason="This test is too slow!")
