### 2024-02-13 203fd11
Go back to Ubuntu runner, macos-14 not reliable for conda yet
