+++ .github/workflows/conda.yml
@@ -4,10 +4,10 @@ on: pull_request
 
 jobs:
   test:
-    runs-on: macos-14
+    runs-on: ubuntu-latest
     defaults:
       run:
-        shell: sh -l {0}
+        shell: bash
 
     steps:
       - uses: actions/checkout@v4
@@ -15,7 +15,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v3
         with:
           python-version: '3.10'
-          installer-url: https://github.com/conda-forge/miniforge/releases/download/23.11.0-0/Mambaforge-23.11.0-0-MacOSX-arm64.sh
+          miniforge-version: latest
       - name: "Install AutoEIS"
         run: |
             python -m pip install --upgrade pip
