### 2024-02-13 c5bc503
Bump version to v0.0.25
