### 2024-02-27 27954e5
Add `scatter` and `label` as optional kwargs to `plot_impedance_combo`
