### 2024-02-27 8ceaf98
Refactor logging mechanism
