### 2024-03-01 AUTODIAL/AutoEIS#96 @dap-biospec
`suppress_output` isn't compatible with Windows, causing install-at-import to fail
Fresh installation using:

```shell
pip install -U autoeis
... [success]
```

followed by:

```python
import numpy as np
import autoeis as ae
```

```python
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "D:\Python\Python312\Lib\site-packages\autoeis\__init__.py", line 1, in <module>
    from . import core, io, metrics, parser, utils, visualization  # noqa: F401, E402
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  ....
  File "D:\Python\Python312\Lib\site-packages\juliapkg\install_julia.py", line 23, in log
    print(prefix, *args)
OSError: [WinError 1] Incorrect function
[juliapkg] Locating Julia ^1.6.1

 # Load test dataset shipped with AutoEIS
 Z, freq = ae.io.load_test_dataset()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'ae' is not defined
```

Same behavior with standalone or Microsoft store installation; with or without Julia pre-installed.

Recommendations?
