### 2024-03-01 #6256 @dap-biospec
@ma-sadeghi @lucydot So far I am unable to verify functionality due to an OSError  [posted](https://github.com/AUTODIAL/AutoEIS/issues/96).
