### 2024-03-06 #6256 @lucydot
Hi @ma-sadeghi:

- you can see @dap-biospec comment above re: install with Win64 Python

- I have read through the [previous comment](https://github.com/openjournals/joss-reviews/issues/6256#issuecomment-1969660317) from @DevT-0. They raise an issue around the approach of using data-driven rather than physico-chemical rules for model construction. This seems to be a comment on the fundamental method underlying your code, and not one which can be addressed in this review process. However I can see that the more data-driven approach is adopted widely in the community, and so I do not see as an acceptance blocker to publication. However, if not already done, I suggest you make clear in your documentation the possible pitfalls of data-driven approaches (such as overfitting).

@DevT-0 I can see there are a number of tick boxes which are unticked under Documentation and Software Paper. What needs to be done (at a minimum) so that these can be ticked? There is a "Statement of Need" and "State of the field" (existing related software projects) listed in the paper, so I think these can be ticked off? You also indicate that there is good coverage of testing (it does not need to be complete), so I infer from this "automated tests" description is also met?
