### 2024-03-22 a13ed84
Force JAX to use CPU to avoid annoying GPU warnings
