### 2024-03-22 480dd8f
Don't show legend when there's none
