### 2024-03-22 abbe1b0
Force FP64 otherwise circuit fitter diverges
