### 2024-03-22 51cd0f2
Bump version to v0.0.26
