+++ pyproject.toml
@@ -1,7 +1,3 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
 [project]
 name = "autoeis"
 dynamic = ["version"]
@@ -15,9 +11,7 @@ authors = [
     { name = "Robert Black", email = "robert.black@nrc-cnrc.gc.ca" },
     { name = "Jason Hattrick-Simpers", email = "jason.hattrick.simpers@utoronto.ca" },
 ]
-maintainers = [
-  { name = "Amin Sadeghi", email = "amin.sadeghi@live.com" }
-]
+maintainers = [{ name = "Amin Sadeghi", email = "amin.sadeghi@live.com" }]
 keywords = [
     "bayesian inference",
     "electrochemical impedance spectroscopy",
@@ -55,10 +49,8 @@ dependencies = [
     "multiprocess==0.70.15",
 ]
 [project.optional-dependencies]
-test = [
-  "pytest",
-  "nbmake",
-]
+build = ["hatch"]
+test = ["pytest", "nbmake"]
 doc = [
     "furo",
     "linkify-it-py",
@@ -69,20 +61,35 @@ doc = [
 ]
 [project.urls]
 "Homepage" = "https://github.com/AUTODIAL/AutoEIS"
+Repository = "https://github.com/AUTODIAL/AutoEIS"
 "Bug Tracker" = "https://github.com/AUTODIAL/AutoEIS/issues"
+Documentation = "https://autodial.github.io/AutoEIS"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.rye]
+managed = true
+dev-dependencies = ["pytest>=8.1.1", "hatch>=1.9.4"]
 
 [tool.hatch.version]
 path = "autoeis/version.py"
 
 [tool.hatch.build.targets.sdist]
-include = [
-    "/autoeis",
-]
+include = ["/autoeis"]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
 addopts = "-ra -v"
-testpaths = [
-    "tests",
-    "integration",
-]
+python_files = "*.py"
+python_classes = "*Test"
+python_functions = "test_*"
+testpaths = ["tests", "integration"]
+norecursedirs = [".git", ".github", ".ipynb_checkpoints", "build", "dist"]
+
+[tool.ruff]
+exclude = [".git", ".github", ".venv", "build"]
+line-length = 92
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["E402", "F401", "F403"]
