+++ doc/basic_usage.md
@@ -8,9 +8,15 @@ import autoeis as ae
 Z, freq = ae.io.load_test_dataset()
 ae.visualization.plot_impedance_combo(Z, freq)
 
-# Perform the full analysis
-circuits = ae.perform_full_analysis(Z, freq, iters=24, parallel=True)
+# Perform automated EIS analysis
+circuits = ae.perform_full_analysis(freq, Z, iters=24, parallel=True)
 print(circuits)
+
+# Print summary of the results
+mcmcs, status = circuits["MCMC (chain)"], circuits["MCMC (status)"]
+for mcmc, stat, circuit in zip(mcmcs, status, circuits.circuitstring):
+    if stat == 0:
+        ae.visualization.print_summary_statistics(mcmc, circuit)
 ```
 
 :::{seealso}
