### 2024-03-23 c5faa7c
Undo accidental changes to core.py
