+++ src/autoeis/core.py
@@ -27,7 +27,6 @@ import psutil
 from impedance.validation import linKK
 from jax import config
 from mpire import WorkerPool
-from numpyro.distributions import Distribution
 from numpyro.infer import MCMC, NUTS
 from scipy.optimize import curve_fit
 from tqdm.auto import tqdm
@@ -498,7 +497,6 @@ def perform_bayesian_inference(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
     p0: Union[np.ndarray, dict, list[dict], list[np.ndarray]] = None,
-    priors: dict[str, Distribution] = None,
     num_warmup=2500,
     num_samples=1000,
     num_chains=1,
@@ -583,9 +581,8 @@ def perform_bayesian_inference(
 
     if len(circuits) == 1:
         # Single inference gets slowed down by progress bar
-        # bi_kwargs["progress_bar"] = False
-        # NOTE: For single inference, DON'T return a list -> crashes multiprocessing
-        return _perform_bayesian_inference(circuits[0], p0=p0[0], **bi_kwargs)
+        bi_kwargs["progress_bar"] = False
+        return [_perform_bayesian_inference(circuits[0], p0=p0[0], **bi_kwargs)]
     return _perform_bayesian_inference_batch(circuits, p0=p0, **bi_kwargs)
 
 
@@ -594,7 +591,6 @@ def _perform_bayesian_inference(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
     p0: Union[np.ndarray[float], dict[str, float]] = None,
-    priors: dict[str, Distribution] = None,
     num_warmup=2500,
     num_samples=1000,
     num_chains=1,
@@ -639,19 +635,10 @@ def _perform_bayesian_inference(
         p0 = utils.fit_circuit_parameters(circuit, freq, Z)
     assert isinstance(p0, dict), "p0 must be a dictionary"
 
-    # TODO: Remove this, circuit fitting must be done in the public API
-    # Deal with initial values for the circuit parameters
-    if priors is None:
-        if p0 is None:
-            p0 = utils.fit_circuit_parameters(circuit, freq, Z)
-        assert isinstance(p0, dict), "p0 must be a dictionary"
-        # Create priors for the circuit parameters based on the initial guess
-        priors = utils.initialize_priors(p0, variables=p0.keys())
-    else:
-        assert isinstance(priors, dict), "'priors' must be a dictionary"
-
     circuit_fn = utils.generate_circuit_fn(circuit, jit=True)
 
+    # Compute prior predictive distribution using the initial guess
+    priors = utils.initialize_priors(p0, variables=p0.keys())
     nuts_kernel = NUTS(
         model=circuit_regression_wrapped,
         init_strategy=numpyro.infer.init_to_median,
@@ -684,7 +671,6 @@ def _perform_bayesian_inference_batch(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
     p0: list[dict[str, float]] = None,
-    priors: list[dict[str, Distribution]] = None,
     num_warmup=2500,
     num_samples=1000,
     num_chains=1,
@@ -723,7 +709,6 @@ def _perform_bayesian_inference_batch(
         "freq": [freq] * N,
         "Z": [Z] * N,
         "p0": p0 if isinstance(p0, list) else [p0] * N,
-        "priors": priors if isinstance(priors, list) else [priors] * N,
         "num_warmup": [num_warmup] * N,
         "num_samples": [num_samples] * N,
         "num_chains": [num_chains] * N,
