### 2024-03-23 408999e
Account for the missing 'v' when comparing most recent tag and version number [no ci]
