### 2024-03-23 0477b85
Force checkout action also fetch tags [no ci]
