+++ .github/workflows/pypi.yml
@@ -17,6 +17,7 @@ jobs:
     - uses: actions/checkout@v4
       with:
         ref: main
+        fetch-tags: true
 
     - name: Set up Python
       uses: actions/setup-python@v5
