### 2024-03-23 42b7c8c
fetch-tags didn't seem to work, fetch everything [no ci]
