+++ .github/workflows/pypi.yml
@@ -17,6 +17,7 @@ jobs:
     - uses: actions/checkout@v4
       with:
         ref: main
+        fetch-depth: 0
         fetch-tags: true
 
     - name: Set up Python
