+++ doc/conf.py
@@ -1,30 +1,15 @@
 from datetime import date
 
-
-def get_version_from_file(file_path):
-    version = None
-    with open(file_path, 'r') as file:
-        for line in file:
-            if line.startswith('__version__'):
-                # Assuming the line is in the form of '__version__ = "0.0.17"'
-                version = line.split('=')[1].strip().strip('"').strip("'")
-                break
-    return version
-
-
-# Configuration file for the Sphinx documentation builder.
-#
-# For the full list of built-in configuration values, see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
+from autoeis.version import __version__
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'AutoEIS'
+project = "AutoEIS"
 copyright = f"{date.today().year}, AutoEIS developers"
-author = 'Runze Zhang, Amin Sadeghi, Jason Hattrick-Simpers'
-version = get_version_from_file("../autoeis/version.py")
-release = version
+author = "Runze Zhang, Amin Sadeghi, Jason Hattrick-Simpers"
+version = __version__
+release = __version__
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -33,11 +18,11 @@ release = version
 suppress_warnings = ["myst.header"]
 
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.autosummary',
-    'myst_parser',
-    'sphinx_copybutton',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autosummary",
+    "myst_parser",
+    "sphinx_copybutton",
     # 'autodoc2',
     # 'numpydoc',
 ]
@@ -59,8 +44,8 @@ myst_enable_extensions = [
     "tasklist",
 ]
 
-templates_path = ['_templates']
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+templates_path = ["_templates"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 autodoc2_packages = [
     "../autoeis",
@@ -69,10 +54,10 @@ autodoc2_packages = [
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'furo'
+html_theme = "furo"
 # html_title = ''
-html_static_path = ['_static']
-html_css_files = ['custom.css']
+html_static_path = ["_static"]
+html_css_files = ["custom.css"]
 html_theme_options = {
     "sidebar_hide_name": True,
     "light_logo": "logo-light-mode.png",
