### 2024-03-25 8fbc324
Pin juliacall version, possible upstream bug
