### 2024-03-25 AUTODIAL/AutoEIS#96 @ma-sadeghi
Since our automated tests are currently passing for the combination of Windows and Python 3.12, I'll close this issue, but feel free to re-open if you still can't import `autoeis`.
