### 2024-03-25 #6256 @ma-sadeghi
Hi @lucydot, I believe the Win64/Python3.12 issue has been fixed (I don't have a Windows machine to check myself, but our automated tests are passing).

Hi @dap-biospec, could you please double check?
