### 2024-03-26 57c8309
Add `eval_posterior_predictive` helper function
