### 2024-03-26 f36a640
Add unit test for `eval_posterior_predictive`
