### 2024-03-26 fdcc9d2
Properly override matplotlib color names
