### 2024-03-26 d6dfea1
Cycle over linestyle as well as colors
