### 2024-03-26 8512989
Escape special characters in warnings regex
