+++ src/autoeis/io.py
@@ -12,6 +12,7 @@ Collection of functions for importing and exporting EIS data/results.
     parse_ec_output
 
 """
+
 import os
 import re
 from pathlib import Path
@@ -32,7 +33,13 @@ def get_assets_path() -> Path:
 
 
 def load_test_dataset() -> tuple[np.ndarray[complex], np.ndarray[float]]:
-    """Returns a test dataset as a tuple of impedance and frequency arrays."""
+    """Returns a test dataset as a tuple of impedance and frequency arrays.
+
+    Returns
+    -------
+    tuple[np.ndarray[complex], np.ndarray[float]]
+        The impedance and frequency arrays.
+    """
     PATH = get_assets_path()
     fpath = os.path.join(PATH, "test_data.txt")
     freq, Zreal, Zimag = np.loadtxt(fpath, skiprows=1, unpack=True, usecols=(0, 1, 2))
@@ -42,7 +49,19 @@ def load_test_dataset() -> tuple[np.ndarray[complex], np.ndarray[float]]:
 
 
 def load_test_circuits(filtered=False) -> pd.DataFrame:
-    """Returns equivalent circuits fitted to test dataset for testing."""
+    """Returns candidate ECMs fitted to test dataset for testing.
+
+    Parameters
+    ----------
+    filtered: bool, optional
+        If True, only physically plausible circuits are returned. Default is False.
+
+    Returns
+    -------
+    pd.DataFrame
+        Dataframe containing the ECMs.
+
+    """
     PATH = get_assets_path()
     fname = "circuits_filtered.csv" if filtered else "circuits_unfiltered.csv"
     fpath = os.path.join(PATH, fname)
@@ -53,7 +72,18 @@ def load_test_circuits(filtered=False) -> pd.DataFrame:
 
 
 def parse_ec_output(circuits: list[str]) -> list[tuple[str, dict[str, float]]]:
-    """Parses the output of EquivalentCircuits.jl's ``circuit_evolution``."""
+    """Parses the output of EquivalentCircuits.jl's ``circuit_evolution``.
+
+    Parameters
+    ----------
+    circuits: list[str]
+        List of stringified output of EquivalentCircuits.jl's ``circuit_evolution``.
+
+    Returns
+    -------
+    pd.DataFrame
+        Dataframe containing ECMs (cols: "circuitstring" and "Parameters")
+    """
     # Example: 'EquivalentCircuit("R1", (R1 = 1.0,))' -> ('R1', {'R1': 1.0})
     circuits = [circuits] if isinstance(circuits, str) else circuits
     parsed = []
