### 2024-03-26 4c30109
Add missing docstrings to visualization module
