### 2024-03-27 b8e9ee5
Move logging setup to pkg init
