### 2024-03-27 db3397c
Hide get_runtime under is_notebook
