### 2024-03-27 e83aff7
Remove the unused find_identical_rows helper
