### 2024-03-27 0bdf99c
Remove unused timeout machinery
