+++ src/autoeis/utils.py
@@ -155,36 +155,6 @@ def suppress_output():
         os.close(original_stderr_fd)
 
 
-class TimeoutException(Exception):
-    pass
-
-
-def timeout(seconds):
-    """Raises a TimeoutException if decorated function doesn't return in time."""
-
-    def decorator(func):
-        timeout_msg = f"{func.__name__} didn't converge in time!"
-
-        def _handle_timeout(signum, frame):
-            raise TimeoutException(timeout_msg)
-
-        def wrapper(*args, **kwargs):
-            signal.signal(signal.SIGALRM, _handle_timeout)
-            signal.alarm(seconds)
-            try:
-                result = func(*args, **kwargs)
-            except TimeoutException:
-                log.warning(timeout_msg)
-                result = None
-            finally:
-                signal.alarm(0)
-            return result
-
-        return wrapper
-
-    return decorator
-
-
 # <<< General utils
 
 
