### 2024-03-27 0a42f0f
Add missing docstrings to utils module + small fixes
