+++ src/autoeis/utils.py
@@ -20,10 +20,10 @@ Collection of utility functions used throughout the package.
 import logging
 import os
 import re
-import signal
 import sys
 from collections.abc import Iterable
 from contextlib import contextmanager
+from dataclasses import dataclass
 from functools import wraps
 from typing import Union
 
@@ -32,14 +32,13 @@ import jax.numpy as jnp
 import numpy as np
 import numpyro.distributions as dist
 import pandas as pd
+import psutil
 from impedance.models.circuits import CustomCircuit
 from impedance.models.circuits.fitting import set_default_bounds
 from jax import random
 from numpy import pi  # NOQA: F401
 from numpyro.distributions import Distribution
 from numpyro.infer import MCMC, Predictive
-from rich.console import Console
-from rich.logging import RichHandler
 from scipy import stats
 from scipy.optimize import curve_fit
 
@@ -47,10 +46,10 @@ import __main__
 
 from . import models, parser
 
-# Timeout circuit fitter functions after X seconds
-TIMEOUT_AFTER = 15
+log = logging.getLogger(__name__)
 
-# >>> Logging utils
+
+# >>> General utils
 
 
 def is_notebook():
@@ -94,7 +93,23 @@ class Settings:
 
 
 def flatten(xs: list) -> list:
-    """Returns a list of all elements in a nested iterable."""
+    """Returns a list of all elements in a nested iterable.
+
+    Parameters
+    ----------
+    xs: list
+        A nested iterable.
+
+    Returns
+    -------
+    list
+        A flattened list.
+
+    Examples
+    --------
+    >>> flatten([1, 2, [3, 4], [5, [6, 7]]])
+    [1, 2, 3, 4, 5, 6, 7]
+    """
 
     def _flatten(xs):
         """Returns a generator that flattens a nested iterable."""
@@ -107,23 +122,36 @@ def flatten(xs: list) -> list:
     return list(_flatten(xs))
 
 
-class _SuppressOutput:
-    def __enter__(self):
-        self._original_stdout = sys.stdout
-        self._original_stderr = sys.stderr
-        sys.stdout = open(os.devnull, "w")
-        sys.stderr = open(os.devnull, "w")
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        sys.stdout.close()
-        sys.stderr.close()
-        sys.stdout = self._original_stdout
-        sys.stderr = self._original_stderr
-
-
 def suppress_output_legacy(func):
-    """Suppresses the output of a function."""
-    # NOTE: This approach only works when stdout/stderr are managed by Python
+    """Suppresses the output of a function.
+
+    Parameters
+    ----------
+    func: callable
+        Input function whose output is to be suppressed.
+
+    Returns
+    -------
+    callable
+        A wrapped function with the output suppressed.
+
+    Notes
+    -----
+    This approach only works when stdout/stderr are managed by Python.
+    """
+
+    class _SuppressOutput:
+        def __enter__(self):
+            self._original_stdout = sys.stdout
+            self._original_stderr = sys.stderr
+            sys.stdout = open(os.devnull, "w")
+            sys.stderr = open(os.devnull, "w")
+
+        def __exit__(self, exc_type, exc_val, exc_tb):
+            sys.stdout.close()
+            sys.stderr.close()
+            sys.stdout = self._original_stdout
+            sys.stderr = self._original_stderr
 
     @wraps(func)
     def wrapped(*args, **kwargs):
@@ -165,13 +193,36 @@ def parse_initial_guess(
     p0: Union[np.ndarray, dict[str, float], list[float]],
     circuit: str,
 ) -> np.ndarray:
-    """Parses the initial guess for circuit parameters."""
+    """Parses the initial guess for circuit parameters from various formats
+    and returns an array of parameters.
+
+    Parameters
+    ----------
[…]
