### 2024-03-27 9a1172f
Augment metrics to accept axis
