+++ src/autoeis/metrics.py
@@ -12,10 +12,11 @@ Collection of functions for calculating metrics.
     r2_score
 
 """
+
 import numpy as np
 
 
-def mape_score(y_true, y_pred):
+def mape_score(y_true, y_pred, axis=0):
     """
     Calculates the generalized MAPE (Mean Absolute Percentage Error) score.
 
@@ -25,6 +26,8 @@ def mape_score(y_true, y_pred):
         Ground truth (true) values.
     y_pred : np.ndarray
         Predicted values.
+    axis : int, optional
+        Axis along which to calculate the MAPE score. Default is 0.
 
     Returns
     -------
@@ -35,10 +38,11 @@ def mape_score(y_true, y_pred):
     -----
     This function handles complex numbers in the input arrays.
     """
-    return np.mean(np.abs((y_true - y_pred) / y_true)) * 100
+    # NOTE: abs is needed to handle complex numbers
+    return np.mean(np.abs((y_true - y_pred) / y_true), axis=axis) * 100
 
 
-def mse_score(y_true, y_pred):
+def mse_score(y_true, y_pred, axis=0):
     """
     Calculates the generalized MSE (Mean Squared Error) score.
 
@@ -48,6 +52,8 @@ def mse_score(y_true, y_pred):
         Ground truth (true) values.
     y_pred : np.ndarray
         Predicted values.
+    axis : int, optional
+        Axis along which to calculate the MSE score. Default is 0.
 
     Returns
     -------
@@ -58,10 +64,11 @@ def mse_score(y_true, y_pred):
     -----
     This function handles complex numbers in the input arrays.
     """
-    return np.mean(np.abs(y_true - y_pred) ** 2)
+    # NOTE: abs is needed to handle complex numbers
+    return np.mean(np.abs(y_true - y_pred) ** 2, axis=axis)
 
 
-def rmse_score(y_true, y_pred):
+def rmse_score(y_true, y_pred, axis=0):
     """
     Calculates the generalized RMSE (Root Mean Squared Error) score.
 
@@ -71,6 +78,8 @@ def rmse_score(y_true, y_pred):
         Ground truth (true) values.
     y_pred : np.ndarray
         Predicted values.
+    axis : int, optional
+        Axis along which to calculate the RMSE score. Default is 0.
 
     Returns
     -------
@@ -81,10 +90,10 @@ def rmse_score(y_true, y_pred):
     -----
     This function handles complex numbers in the input arrays.
     """
-    return np.sqrt(mse_score(y_true, y_pred))
+    return np.sqrt(mse_score(y_true, y_pred, axis=axis))
 
 
-def r2_score(y_true, y_pred):
+def r2_score(y_true, y_pred, axis=0):
     """
     Calculates the generalized R2 score.
 
@@ -94,6 +103,8 @@ def r2_score(y_true, y_pred):
         Ground truth (true) values.
     y_pred : np.ndarray
         Predicted values.
+    axis : int, optional
+        Axis along which to calculate the R2 score. Default is 0.
 
     Returns
     -------
@@ -104,6 +115,10 @@ def r2_score(y_true, y_pred):
     -----
     This function handles complex numbers in the input arrays.
     """
-    ssr = np.sum(np.abs(y_true - y_pred) ** 2)
+    assert y_true.squeeze().ndim == 1
+    other_axes = [i for i in range(y_pred.ndim) if i != axis]
+    y_true = np.expand_dims(y_true, axis=other_axes)
+    # NOTE: abs is needed to handle complex numbers
+    ssr = np.sum(np.abs(y_true - y_pred) ** 2, axis=axis)
     sst = np.sum(np.abs(y_true - np.mean(y_true)) ** 2)
     return 1 - ssr / sst
