### 2024-03-27 cac5cab
Refactor metrics module to follow DRY
