### 2024-03-27 c6d5115
Add missing type hints
