+++ src/autoeis/core.py
@@ -499,9 +499,9 @@ def perform_bayesian_inference(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
     p0: Union[np.ndarray, dict, list[dict], list[np.ndarray]] = None,
-    num_warmup=2500,
-    num_samples=1000,
-    num_chains=1,
+    num_warmup: int = 2500,
+    num_samples: int = 1000,
+    num_chains: int = 1,
     seed: Union[int, jax.Array] = None,
     progress_bar: bool = True,
     refine_p0: bool = False,
@@ -593,9 +593,9 @@ def _perform_bayesian_inference(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
     p0: Union[np.ndarray[float], dict[str, float]] = None,
-    num_warmup=2500,
-    num_samples=1000,
-    num_chains=1,
+    num_warmup: int = 2500,
+    num_samples: int = 1000,
+    num_chains: int = 1,
     seed: Union[int, jax.Array] = None,
     progress_bar: bool = True,
 ) -> tuple[numpyro.infer.mcmc.MCMC, int]:
@@ -674,9 +674,9 @@ def _perform_bayesian_inference_batch(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
     p0: list[dict[str, float]] = None,
-    num_warmup=2500,
-    num_samples=1000,
-    num_chains=1,
+    num_warmup: int = 2500,
+    num_samples: int = 1000,
+    num_chains: int = 1,
     seed: Union[int, jax.Array] = None,
     progress_bar=True,
 ):
