### 2024-03-28 6115213
Add `print_inference_results` function to pretty print a summary of inference results
