+++ src/autoeis/legacy.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import time
 
@@ -17,7 +18,7 @@ from tqdm.auto import tqdm
 import autoeis.visualization as viz
 from autoeis import utils
 
-log = utils.get_logger(__name__)
+log = logging.getLogger(__name__)
 
 
 def circuit_to_function(circuit: str, use_jax=True) -> callable:
@@ -26,9 +27,11 @@ def circuit_to_function(circuit: str, use_jax=True) -> callable:
     eqn = generate_mathematical_expression(circuit_df)["Mathematical expressions"][0]
     if use_jax:
         eqn = eqn.replace("np", "jnp")
+
     def _fn(X, F):
         assert utils.count_parameters(circuit) == len(X), "Invalid number of parameters."
         return eval(eqn)
+
     return jax.jit(_fn) if use_jax else _fn
 
 
@@ -72,10 +75,7 @@ def generate_mathematical_expression(df_circuits: pd.DataFrame) -> pd.DataFrame:
             elif test_results[m][0] == "L":
                 circuit = circuit.replace(test_results_2[m], "(2*1j*np.pi*F*X)")
             elif test_results[m][0] == "P":
-                circuit = circuit.replace(
-                    test_results_2[m],
-                    "(1/(X*(2*1j*np.pi*F)**(Y)))"
-                )
+                circuit = circuit.replace(test_results_2[m], "(1/(X*(2*1j*np.pi*F)**(Y)))")
 
         new_temp_circuit = []
         counter = 0
@@ -98,9 +98,7 @@ def generate_mathematical_expression(df_circuits: pd.DataFrame) -> pd.DataFrame:
 
 
 def ohmic_resistance_filter_legacy(
-    circuits: pd.DataFrame,
-    ohmic_resistance: float,
-    rtol=0.5
+    circuits: pd.DataFrame, ohmic_resistance: float, rtol=0.5
 ) -> pd.DataFrame:
     """Filters the circuits whose ohmic resistance doesn't match a desired value."""
     for i in range(len(circuits["circuitstring"])):
@@ -137,7 +135,9 @@ def ohmic_resistance_filter_legacy(
     return circuits
 
 
-def apply_heuristic_rules_legacy(circuits: pd.DataFrame, ohmic_resistance: float) -> pd.DataFrame:
+def apply_heuristic_rules_legacy(
+    circuits: pd.DataFrame, ohmic_resistance: float
+) -> pd.DataFrame:
     """Apply heuristic rules to filter the generated ECMs.
 
     Parameters
@@ -177,7 +177,9 @@ def model_evaluation(results):
 
     # FIXME: Remove next line once confirmed that `loc` is correctly used.
     # evaluation_results["Consistency"] = pd.to_numeric(evaluation_results["Consistency"], errors="coerce")
-    evaluation_results.loc[:, "Consistency"] = pd.to_numeric(evaluation_results["Consistency"], errors="coerce")
+    evaluation_results.loc[:, "Consistency"] = pd.to_numeric(
+        evaluation_results["Consistency"], errors="coerce"
+    )
     evaluation_results.loc[evaluation_results["Consistency"].isna(), "Consistency"] = np.inf
 
     def absdiff(x):
@@ -188,10 +190,14 @@ def model_evaluation(results):
 
     # FIXME: Remove next line once confirmed that `loc` is correctly used.
     # evaluation_results["Consistency"] = evaluation_results["Consistency"].apply(absdiff)
-    evaluation_results.loc[:, "Consistency"] = evaluation_results["Consistency"].apply(absdiff)    
+    evaluation_results.loc[:, "Consistency"] = evaluation_results["Consistency"].apply(
+        absdiff
+    )
[…]
