### 2024-03-28 c7ab3b8
Minor refactor
