### 2024-03-28 ff773a2
Add jinja2>=3.1.2 to deps to make pandas happy (required by DataFrame.style)
