### 2024-03-29 e287983
Use pipe operator instead of Union for type hints
