+++ src/autoeis/utils.py
@@ -21,11 +21,10 @@ import logging
 import os
 import re
 import sys
-from collections.abc import Iterable
+from collections.abc import Callable, Iterable, Mapping
 from contextlib import contextmanager
 from dataclasses import dataclass
 from functools import wraps
-from typing import Union
 
 import jax  # NOQA: F401
 import jax.numpy as jnp
@@ -92,7 +91,7 @@ class Settings:
     progress_bar: bool = True
 
 
-def flatten(xs: list) -> list:
+def flatten(xs: Iterable) -> list:
     """Returns a list of all elements in a nested iterable.
 
     Parameters
@@ -122,7 +121,7 @@ def flatten(xs: list) -> list:
     return list(_flatten(xs))
 
 
-def suppress_output_legacy(func):
+def suppress_output_legacy(func: Callable) -> Callable:
     """Suppresses the output of a function.
 
     Parameters
@@ -190,7 +189,7 @@ def suppress_output():
 
 
 def parse_initial_guess(
-    p0: Union[np.ndarray, dict[str, float], list[float]],
+    p0: np.ndarray | dict[str, float] | list[float],
     circuit: str,
 ) -> np.ndarray:
     """Parses the initial guess for circuit parameters from various formats
@@ -198,7 +197,7 @@ def parse_initial_guess(
 
     Parameters
     ----------
-    p0: Union[np.ndarray, dict[str, float], list[float]]
+    p0: np.ndarray | dict[str, float] | list[float]
         The initial guess for the circuit parameters.
     circuit: str
         The circuit string.
@@ -231,7 +230,7 @@ def fit_circuit_parameters_legacy(
     circuit: str,
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
-    p0: Union[np.ndarray[float], dict[str, float]] = None,
+    p0: np.ndarray[float] | dict[str, float] = None,
     iters: int = 1,
     maxfev: int = 1000,
 ) -> dict[str, float]:
@@ -246,7 +245,7 @@ def fit_circuit_parameters_legacy(
         Frequencies corresponding to the impedance data.
     Z : np.ndarray[complex]
         Impedance data.
-    p0 : Union[np.ndarray[float], dict[str, float]], optional
+    p0 : np.ndarray[float] | dict[str, float], optional
         Initial guess for the circuit parameters. Default is None.
     iters : int, optional
         Maximum number of iterations for the circuit fitter. Default is 1.
@@ -299,7 +298,7 @@ def fit_circuit_parameters(
     circuit: str,
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
-    p0: Union[np.ndarray[float], dict[str, float]] = None,
+    p0: np.ndarray[float] | dict[str, float] = None,
     iters: int = 1,
     maxfev: int = 1000,
     ftol: float = 1e-13,
@@ -315,7 +314,7 @@ def fit_circuit_parameters(
         Frequencies corresponding to the impedance data.
     Z : np.ndarray[complex]
         Impedance data.
-    p0 : Union[np.ndarray[float], dict[str, float]], optional
+    p0 : np.ndarray[float] | dict[str, float], optional
         Initial guess for the circuit parameters. Default is None.
     iters : int, optional
         Maximum number of iterations for the circuit fitter. Default is 1.
@@ -376,9 +375,7 @@ def fit_circuit_parameters(
     return dict(zip(variables, p0))
 
 
-def eval_circuit(
-    circuit: str, f: Union[np.ndarray, float], p: np.ndarray
-) -> np.ndarray[complex]:
+def eval_circuit(circuit: str, f: np.ndarray | float, p: np.ndarray) -> np.ndarray[complex]:
     """Returns the impedance of a circuit at a given frequency and parameters.
 
     Parameters
[…]
