### 2024-03-29 00bf98b
Update docstrings with new function name
