### 2024-03-29 f826114
Replace Union with pipe operator for type hints
