### 2024-03-29 867fc97
Compute fitness metrics as part of full analysis
