### 2024-03-29 0bc58f9
Don't perform BI twice when there's only one circuit (nasty bug!)
