### 2024-03-29 2fc8ee0
Add unit tests for `parse_ec_output`
