### 2024-03-29 7224ef3
Properly deprecate the old circuit finder
