### 2024-03-29 c3e397f
Update GEP functions to use (freq, Z) signature instead of (Z, freq) to be consistent with the rest of the pkg
