+++ src/autoeis/core.py
@@ -120,8 +120,8 @@ def compute_ohmic_resistance(freq: np.ndarray[float], Z: np.ndarray[complex]) ->
 
 # TODO: Needs heavy refactoring
 def preprocess_impedance_data(
-    impedance: np.ndarray[complex],
     freq: np.ndarray[float],
+    Z: np.ndarray[complex],
     threshold: float = 5e-2,
     plot: bool = False,
 ) -> tuple[np.ndarray[complex], np.ndarray[float], float]:
@@ -131,11 +131,11 @@ def preprocess_impedance_data(
     Parameters
     ----------
     impedance : np.ndarray[complex]
-        Impedance measurements.
+        Impedance measurements as a complex array.
     freq : np.ndarray[float]
         Frequencies corresponding to impedance measurements.
     threshold : float
-        Controls the filtering effect during KK validation.
+        Controls the strictness of the filtering effect during KK validation.
     plot : bool, optional
         If True, a plot of the processed data will be generated. Default is False.
 
@@ -149,8 +149,8 @@ def preprocess_impedance_data(
     log.info("Pre-processing impedance data using KK filter.")
 
     # Fetch the real and imaginary part of the impedance data
-    Re_Z = impedance.real
-    Im_Z = impedance.imag
+    Re_Z = Z.real
+    Im_Z = Z.imag
 
     # Filter 1 - High Frequency Region
     # Find index where phase_Zwe == minimum, remove all high frequency imag values below zero
@@ -169,7 +169,7 @@ def preprocess_impedance_data(
             mask_phase[i] = False
 
     freq = freq[index[0] :]
-    Z = impedance[index[0] :]
+    Z = Z[index[0] :]
     Re_Z = Re_Z[index[0] :]
     Im_Z = Im_Z[index[0] :]
 
@@ -259,8 +259,8 @@ def preprocess_impedance_data(
 
 
 def generate_equivalent_circuits(
-    impedance: np.ndarray[complex],
     freq: np.ndarray[float],
+    Z: np.ndarray[complex],
     iters: int = 100,
     complexity: int = 12,
     tol: float = 1e-2,
@@ -269,14 +269,15 @@ def generate_equivalent_circuits(
     population_size: int = 100,
     seed: int = None,
 ) -> pd.DataFrame:
-    """Generates candidate circuits that fit impedance data using genetic algorithm.
+    """Generates candidate circuits that fit the impedance data using
+    evolutionary algorithms.
 
     Parameters
     ----------
-    Z : np.ndarray[complex]
-        Impedance measurements.
     freq : np.ndarray[float]
         Frequencies corresponding to impedance measurements.
+    Z : np.ndarray[complex]
+        Impedance measurements as a complex array.
     iters : int, optional
         Number of ECM generation iterations (default is 100).
     complexity : int, optional
@@ -311,7 +312,7 @@ def generate_equivalent_circuits(
     }
 
     ecm_generator = _generate_ecm_parallel_julia if parallel else _generate_ecm_serial
-    circuits = ecm_generator(impedance, freq, iters, ec_kwargs, seed)
+    circuits = ecm_generator(Z, freq, iters, ec_kwargs, seed)
 
     # Convert output to DataFrame with columns ("circuitstring", "Parameters")
     circuits = io.parse_ec_output(circuits)
@@ -322,15 +323,21 @@ def generate_equivalent_circuits(
     return circuits
 
 
-def _generate_ecm_serial(impedance, freq, iters, ec_kwargs, seed) -> list[str]:
-    """Generates candidate circuits in serial."""
+def _generate_ecm_serial(
+    freq: np.ndarray[float],
+    Z: np.ndarray[complex],
+    iters: int,
+    ec_kwargs: dict,
+    seed: int,
+) -> list[str]:
[…]
