+++ pyproject.toml
@@ -48,6 +48,7 @@ dependencies = [
     "tqdm",
     "multiprocess",
     "jinja2>=3.1.2",
+    "deprecated",
 ]
 [project.optional-dependencies]
 build = ["hatch"]
