### 2024-03-29 a5ffcb6
Apply (freq, Z) signature throughout the pkg
