+++ tests/test_io.py
@@ -9,7 +9,7 @@ def test_get_assets_path():
 
 
 def test_load_test_dataset():
-    Z, freq = ae.io.load_test_dataset()
+    freq, Z = ae.io.load_test_dataset()
     assert Z.shape == freq.shape
     assert len(Z) == len(freq) > 0
 
