+++ src/autoeis/core.py
@@ -118,144 +118,71 @@ def compute_ohmic_resistance(freq: np.ndarray[float], Z: np.ndarray[complex]) ->
     return R
 
 
-# TODO: Needs heavy refactoring
 def preprocess_impedance_data(
     freq: np.ndarray[float],
     Z: np.ndarray[complex],
-    threshold: float = 5e-2,
-    plot: bool = False,
+    tol_linKK: float = 5e-2,
+    high_freq_threshold: float = 1e3,
 ) -> tuple[np.ndarray[complex], np.ndarray[float], float]:
-    """Preprocess impedance data, filtering out data with a positive
-    imaginary part in high frequencies and applying KK validation.
+    """Preprocesses/cleans up impedance measurements.
+
+    The preprocessing does the following steps:
+        - Discard invalid high frequency measurements (see Notes section)
+        - Filter out data with a positive imaginary part in high frequencies
+        - Enforce the Kramers-Kronig validation (aka Lin-KK)
 
     Parameters
     ----------
-    impedance : np.ndarray[complex]
-        Impedance measurements as a complex array.
     freq : np.ndarray[float]
         Frequencies corresponding to impedance measurements.
-    threshold : float
-        Controls the strictness of the filtering effect during KK validation.
-    plot : bool, optional
-        If True, a plot of the processed data will be generated. Default is False.
+    Z : np.ndarray[complex]
+        Impedance measurements as a complex array.
+    tol_linKK : float
+        Tolerance for acceptable measurements based on linKK residuals.
+    high_freq_threshold : float
+        Lower bound for what is considered a high frequency measurement.
 
     Returns
     -------
-    tuple
+    tuple[np.ndarray[complex], np.ndarray[float], float]
         - Z: Filtered impedance data.
         - freq: Frequencies corresponding to the filtered measurements.
         - rmse: Root mean square error of KK validated data vs. measurements.
     """
-    log.info("Pre-processing impedance data using KK filter.")
+    log.info("Preprocessing/cleaning up impedance data.")
+    n0 = len(freq)
 
-    # Fetch the real and imaginary part of the impedance data
-    Re_Z = Z.real
-    Im_Z = Z.imag
+    # Make sure frequency is sorted in descending order
+    Z = Z[np.argsort(freq)[::-1]]
+    freq = freq[np.argsort(freq)[::-1]]
 
-    # Filter 1 - High Frequency Region
-    # Find index where phase_Zwe == minimum, remove all high frequency imag values below zero
-    # Find index: 10khz - 100khz
-    # ?: What's the logic behind this?
-    # BUG: ???
-    try:
-        pos = np.where((1000 <= freq) & (freq <= 1000000))
-        # Find minimum phase value, note returns as a tuple
-        (index,) = np.where(abs(Im_Z[pos]) == abs(Im_Z[pos]).min())
-    except ValueError:
-        index = [0]
-    mask_phase = [True] * len(Im_Z)
-    for i in range(len(Im_Z)):
-        if i < index:
-            mask_phase[i] = False
+    # Heuristic 1: @freq->∞: |Z.im|->0 => highest_valid_freq = freq @ np.argmin(|Z.im|)
+    high_freq = freq > high_freq_threshold
+    idx_highest_valid_freq = np.argmin(np.abs(Z.imag[high_freq]))
+    # Filter out frequencies above the highest valid frequency (only works if freq is sorted)
+    freq = freq[idx_highest_valid_freq:]
[…]
