+++ tests/test_core.py
@@ -29,9 +29,19 @@ def test_compute_ohmic_resistance_missing_high_freq():
     np.testing.assert_allclose(R, Zreal_at_high_freq)
 
 
+def test_preprocess_impedance_data():
+    freq, Z = io.load_test_dataset()
+    freq_prep, Z_prep, rmse = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
+    assert len(Z_prep) == len(freq_prep)
+    assert len(Z_prep) == 60
+    freq_prep, Z_prep, rmse = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-3)
+    assert len(Z_prep) == len(freq_prep)
+    assert len(Z_prep) == 50
+
+
 def test_gep_serial():
-    Z, freq = io.load_test_dataset()
-    Z, freq, rmse = core.preprocess_impedance_data(Z, freq, threshold=5e-2)
+    freq, Z = io.load_test_dataset()
+    freq, Z, rmse = core.preprocess_impedance_data(freq, Z, threshold=5e-2)
     kwargs = {
         "iters": 2,
         "complexity": 12,
@@ -40,14 +50,14 @@ def test_gep_serial():
         "tol": 1e10,
         "parallel": False,
     }
-    circuits = core.generate_equivalent_circuits(Z, freq, **kwargs)
+    circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
     assert len(circuits) == kwargs["iters"]
     assert isinstance(circuits, pd.DataFrame)
 
 
 def test_gep_parallel():
-    Z, freq = io.load_test_dataset()
-    Z, freq, rmse = core.preprocess_impedance_data(Z, freq, threshold=5e-2)
+    freq, Z = io.load_test_dataset()
+    freq, Z, rmse = core.preprocess_impedance_data(freq, Z, threshold=5e-2)
     kwargs = {
         "iters": 2,
         "complexity": 12,
@@ -56,13 +66,13 @@ def test_gep_parallel():
         "tol": 1e10,
         "parallel": True,
     }
-    circuits = core.generate_equivalent_circuits(Z, freq, **kwargs)
+    circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
     assert len(circuits) == kwargs["iters"]
     assert isinstance(circuits, pd.DataFrame)
 
 
 def test_filter_implausible_circuits():
-    Z, freq = io.load_test_dataset()
+    freq, Z = io.load_test_dataset()
     circuits_unfiltered = io.load_test_circuits()
     N1 = len(circuits_unfiltered)
     circuits = core.filter_implausible_circuits(circuits_unfiltered)
@@ -71,7 +81,7 @@ def test_filter_implausible_circuits():
 
 
 def test_bayesian_inference_single():
-    Z, freq = io.load_test_dataset()
+    freq, Z = io.load_test_dataset()
     circuits = io.load_test_circuits(filtered=True)
     circuit = circuits.iloc[0].circuitstring
     p0 = circuits.iloc[0].Parameters
@@ -83,7 +93,7 @@ def test_bayesian_inference_single():
 
 
 def test_bayesian_inference_batch():
-    Z, freq = io.load_test_dataset()
+    freq, Z = io.load_test_dataset()
     # Only test first three circuits to save time in CI
     circuits = io.load_test_circuits(filtered=True).iloc[:3]
     mcmc_results = core.perform_bayesian_inference(circuits, freq, Z, refine_p0=False)
@@ -95,7 +105,7 @@ def test_bayesian_inference_batch():
 
 @pytest.mark.skip(reason="This test is too slow!")
 def test_perform_full_analysis():
-    Z, freq = io.load_test_dataset()
[…]
