### 2024-03-29 2a58820
Update `preprocess_impedance_data` to return a Box, not a tuple
