### 2024-03-29 f819171
Update unit tests based on recent return type change
