### 2024-03-30 c228edc
Grant write permission to PyPI action [no ci]
