### 2024-03-30 9e50bf6
Fix example script in README [no ci]
