### 2024-03-31 AUTODIAL/AutoEIS#96 @dap-biospec
@ma-sadeghi  Package us is now locating Julia versions "~1.6.1, ~1.7, ~1.8, ~1.9, =1.10.0"  but returns the same "OSError: [WinError 1] Incorrect function." Subsequent code returns "name 'ae' is not defined"
