### 2024-03-31 #6256 @dap-biospec
@ma-sadeghi @lucydot I am still getting the same error.
I will try it on Linux, but Win functionality will have to be resolved or specific workarounds described.

Overall, this project needs substantial work on documentation, tutorials, examples, pitfalls etc. to come to a publishable level. I hope to provide more specifics once it runs on my side.
