### 2024-04-01 5b36ec1
Set up logger before module imports
