### 2024-04-01 86b9dc4
Make `suppress_output` compatible with Windows
