+++ .github/workflows/gh-pages.yml
@@ -22,6 +22,8 @@ jobs:
         run: |
             python -m pip install --upgrade pip
             pip install -e .[doc]
+            # Install Julia dependencies at first import
+            python -c "import autoeis"            
 
       # Build the documentation
       - name: Build the documentation
