+++ .github/workflows/nightly.yml
@@ -38,7 +38,8 @@ jobs:
         run: |
             python -m pip install --upgrade pip
             pip install -e .[test]
-            python -m autoeis install
+            # Install Julia dependencies at first import
+            python -c "import autoeis"
       - name: "Run tests"
         run: |
           pytest -v tests/
