+++ .github/workflows/notebooks.yml
@@ -34,7 +34,8 @@ jobs:
         run: |
             python -m pip install --upgrade pip
             pip install -e .[test]
-            python -m autoeis install
+            # Install Julia dependencies at first import
+            python -c "import autoeis"
       - name: "Run tests"
         run: |
           pytest -v --nbmake --nbmake-timeout=3600 examples/*.ipynb
