### 2024-04-01 bcf2e33
Patch `suppress_output` to work in Jupyter too
