### 2024-04-01 AUTODIAL/AutoEIS#96 @ma-sadeghi
@dap-biospec Sorry for the frustration. I got my hands on a Windows machine this morning, and figured out the root cause. It's now fixed, please upgrade `autoeis`, and should be good to go.

PS. The reason our Windows CI was not failing was because we were using the CLI approach to install Julia dependencies (`python -m autoeis install`), which worked fine, but the install-at-first-import didn't, which wasn't tested. Now, all CIs install Julia dependencies at import, so it's now fully covered in case something goes wrong in the future.
