### 2024-04-01 aa2b608
Update unit tests based on API change
