+++ tests/test_core.py
@@ -31,52 +31,55 @@ def test_compute_ohmic_resistance_missing_high_freq():
 
 def test_preprocess_impedance_data():
     freq, Z = io.load_test_dataset()
-    results = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
-    freq_prep, Z_prep = results.freq, results.Z
+    # Test various tolerances for linKK validation
+    freq_prep, Z_prep = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
     assert len(Z_prep) == len(freq_prep)
     assert len(Z_prep) == 60
-    results = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-3)
-    freq_prep, Z_prep = results.freq, results.Z
+    freq_prep, Z_prep = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-3)
     assert len(Z_prep) == len(freq_prep)
     assert len(Z_prep) == 50
+    # Test return_aux=True
+    _, _, aux = core.preprocess_impedance_data(freq, Z, return_aux=True)
+    assert list(aux.keys()) == ["res", "rmse"]
+    assert list(aux["res"].keys()) == ["real", "imag"]
 
 
-def test_gep_serial():
-    freq, Z = io.load_test_dataset()
-    results = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
-    freq, Z = results.freq, results.Z
-    kwargs = {
-        "iters": 2,
-        "complexity": 12,
-        "population_size": 5,
-        "generations": 5,
-        "tol": 1e10,
-        "parallel": False,
-    }
-    circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
-    assert len(circuits) == kwargs["iters"]
-    assert isinstance(circuits, pd.DataFrame)
+def test_gep():
+    def test_gep_serial():
+        freq, Z = io.load_test_dataset()
+        freq, Z = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
+        kwargs = {
+            "iters": 2,
+            "complexity": 2,
+            "population_size": 5,
+            "generations": 2,
+            "tol": 1e10,
+            "parallel": False,
+        }
+        circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
+        assert len(circuits) == kwargs["iters"]
+        assert isinstance(circuits, pd.DataFrame)
 
+    def test_gep_parallel():
+        freq, Z = io.load_test_dataset()
+        freq, Z = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
+        kwargs = {
+            "iters": 2,
+            "complexity": 2,
+            "population_size": 5,
+            "generations": 2,
+            "tol": 1e10,
+            "parallel": True,
+        }
+        circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
+        assert len(circuits) == kwargs["iters"]
+        assert isinstance(circuits, pd.DataFrame)
 
-def test_gep_parallel():
-    freq, Z = io.load_test_dataset()
-    results = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
-    freq, Z = results.freq, results.Z
-    kwargs = {
-        "iters": 2,
-        "complexity": 12,
-        "population_size": 5,
-        "generations": 5,
-        "tol": 1e10,
-        "parallel": True,
-    }
-    circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
-    assert len(circuits) == kwargs["iters"]
-    assert isinstance(circuits, pd.DataFrame)
+    test_gep_serial()
[…]
