### 2024-04-01 5206402
Resolve ec-path in CLI installer before passing it to pyjuliapkg
