+++ src/autoeis/cli.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 import click
 
 from .julia_helpers import install_backend, install_julia
@@ -17,5 +19,8 @@ def autoeis_installer(context):
 )
 @autoeis_installer.command("install", help="Install Julia dependencies for AutoEIS.")
 def install_cli(ec_path):
+    if ec_path is not None:
+        # Expand ~ in path if present
+        ec_path = Path(ec_path).expanduser()
     install_julia()
     install_backend(ec_path=ec_path)
