### 2024-04-01 7b4595d
Minor reorganizing, fix serial GEP to run in quiet mode
