+++ src/autoeis/julia_helpers.py
@@ -1,4 +1,5 @@
 import logging
+from pathlib import Path
 
 import juliapkg
 from juliapkg.deps import can_skip_resolve
@@ -20,13 +21,20 @@ def install_julia(quiet=True):
         import juliacall
 
 
-def install_backend(ec_path=None, quiet=True):
-    """Installs Julia dependencies for AutoEIS."""
+def install_backend(ec_path: Path = None, quiet=True):
+    """Installs Julia dependencies for AutoEIS.
+
+    Parameters
+    ----------
+    ec_path : Path, optional
+        Path to the local copy of EquivalentCircuits. Default is None. If None,
+        the remote version will be used.
+    """
     is_julia_installed(error=True)
 
     kwargs = {"name": "EquivalentCircuits", "uuid": "da5bd070-f609-4e16-a30d-de86b3faa756"}
     if ec_path is not None:
-        kwargs["path"] = ec_path
+        kwargs["path"] = str(ec_path)
         kwargs["dev"] = True
     else:
         if __equivalent_circuits_jl_version__.startswith("v"):
@@ -44,7 +52,13 @@ def install_backend(ec_path=None, quiet=True):
 
 
 def init_julia(quiet=False):
-    """Initializes Julia and returns the Main module."""
+    """Initializes Julia and returns the Main module.
+
+    Raises
+    ------
+    ImportError
+        If Julia is not installed.
+    """
     is_julia_installed(error=True)
     if quiet:
         with suppress_output():
@@ -56,7 +70,27 @@ def init_julia(quiet=False):
 
 
 def import_package(package_name, Main, error=False):
-    """Imports a package in Julia and returns the module."""
+    """Imports a package in Julia and returns the module.
+
+    Parameters
+    ----------
+    package_name : str
+        Name of the Julia package to import.
+    Main : juliacall.Main
+        Julia Main module.
+    error : bool, optional
+        If True, raises an error if the package is not found. Default is False.
+
+    Returns
+    -------
+    module
+        The imported Julia module.
+
+    Raises
+    ------
+    ImportError
+        If the package is not found and error is True.
+    """
     from juliacall import JuliaError
 
     try:
@@ -69,7 +103,23 @@ def import_package(package_name, Main, error=False):
 
 
 def import_backend(Main=None):
-    """Imports EquivalentCircuits package from Julia."""
+    """Imports EquivalentCircuits package from Julia.
+
+    Parameters
+    ----------
+    Main : juliacall.Main, optional
+        Julia Main module. Default is None.
+
+    Returns
+    -------
+    module
+        The imported Julia module.
+
+    Raises
+    ------
+    ImportError
+        If Julia is not installed or the package is not found.
+    """
     Main = init_julia() if Main is None else Main
     is_backend_installed(Main=Main, error=True)
     return import_package("EquivalentCircuits", Main)
@@ -93,7 +143,28 @@ def is_julia_installed(error=False):
 
 
 def is_backend_installed(Main=None, error=False):
-    """Asserts that EquivalentCircuits.jl is installed."""
+    """Asserts that EquivalentCircuits.jl is installed.
+
+    Parameters
+    ----------
+    Main : juliacall.Main, optional
+        Julia Main module. Default is None. If None, the Main module will be
+        initialized using `init_julia()`.
+    error : bool, optional
+        If True, raises an error if the package is not found. Default is False.
+    install : bool, optional
+        If True, installs the package if it is not found. Default is False.
+
+    Returns
+    -------
+    bool
+        True if the package is installed, False otherwise.
+
+    Raises
+    ------
+    ImportError
+        If Julia is not installed or the package is not found and error is True.
+    """
     Main = init_julia() if Main is None else Main
     if import_package("EquivalentCircuits", Main, error=False) is not None:
         return True
