### 2024-04-01 7ab3739
Add warning when Julia needs to be resolved
