+++ pyproject.toml
@@ -95,7 +95,7 @@ include = ["src/autoeis"]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-addopts = "-ra -v"
+addopts = "-ra -vv --durations=5"
 python_files = "*.py"
 python_classes = "*Test"
 python_functions = "test_*"
