### 2024-04-05 3730084
Add documentation on supported data formats
