### 2024-04-05 2e37948
Add instructions on input data format
