### 2024-04-09 #6256 @dap-biospec
@ma-sadeghi @lucydot Yes, Julia autodependency is now resolved. I made multiple attempts to execute autoeis demo on both Win64 (python 3.12, three different nodes)  and Linux (3.10) over the past week , none of which complete.  Most runs ended with KeyError 'MCMC'.  Until the demo is updated and verified to work on all supported platforms I see no possibility to proceed with the review, let alone recommend it for publication.
