### 2024-04-10 #6256 @lucydot
Hi @dap-biospec  - thanks for your time spent testing and reporting back.

@ma-sadeghi  - note that we do not require JOSS submissions to be compatible with all OS's. ie, Windows compatibility is not a requirement for acceptance (but we do require clear guidance in the documentation as to which OS's are supported). Of course, it needs to run somewhere - and it seems that it is not compatible with the Win64 or Linux OS's that @dap-biospec has tested on.

Can you specify which OS's you have tested on and have verified it works? To aid with @dap-biospec's review it sounds that Linux 3.10 might be a good one to target for compatibility.
