+++ pyproject.toml
@@ -61,6 +61,9 @@ doc = [
     "sphinx",
     "sphinx-copybutton",
     "sphinx-autodoc2",
+    "nbsphinx>=0.9.3",
+    "sphinx-autobuild>=2024.2.4",
+    "myst-nb>=1.0.0",
 ]
 [project.urls]
 "Homepage" = "https://github.com/AUTODIAL/AutoEIS"
@@ -85,6 +88,7 @@ dev-dependencies = [
     "sphinx-autodoc2>=0.5.0",
     "nbmake>=1.5.3",
     "jupyterlab>=4.1.5",
+    "myst-nb>=1.0.0",
 ]
 
 [tool.hatch.version]
