### 2024-04-10 205914c
Add examples to docs
