### 2024-04-10 9572510
Install brew in docs action
