### 2024-04-12 f08175c
bode plot
