### 2024-04-12 306e8fd
add instructions on input formats
