### 2024-04-13 470a263
Add example notebook on ecm basics
