+++ doc/examples.rst
@@ -3,13 +3,12 @@ Examples
 
 The following examples demonstrate how to use AutoEIS to perform EIS analysis. You can download the examples from the main `AutoEIS repository <https://github.com/AUTODIAL/AutoEIS>`__. To run the examples, you need to install `jupyter notebook <https://jupyter.org/>`__. If you have any questions or issues running the examples, please open an `issue <https://github.com/AUTODIAL/AutoEIS/issues>`__ or directly submit a `pull request <https://github.com/AUTODIAL/AutoEIS/pulls>`__. We would greatly appreciate any contributions from the community.
 
-Note that ``AutoEIS`` is still under development, and we will add more examples as we progress. If you have any suggestions for new examples, please open an `issue <https://github.com/AUTODIAL/AutoEIS/issues>`__ or directly submit a `pull request <https://github.com/AUTODIAL/AutoEIS/pulls>`__.
+AutoEIS is still under development, and we will add more examples as we make progress. In the meantime, feel free to explore the functionalities via the `API reference <modules>`__. If you have any suggestions for new examples, please open an `issue <https://github.com/AUTODIAL/AutoEIS/issues>`__ or directly submit a `pull request <https://github.com/AUTODIAL/AutoEIS/pulls>`__.
 
-Demos
------
 
 .. nbgallery::
-    :name: bayesian-workflow
+    :name: examples-gallery
     :maxdepth: 1
 
-    examples/demos/basic_workflow
+    examples/circuit_basics
+    examples/test
