### 2024-04-13 8182ff0
Revise the documentation of inputs
