### 2024-04-13 9394ad0
Add basic workflow to list of examples
