### 2024-04-13 37aedbf
Run notebooks and save results inplace before deploying docs
