+++ doc/conf.py
@@ -67,7 +67,7 @@ nbsphinx_execute_arguments = [
     "--InlineBackend.figure_formats={'svg', 'pdf'}",
     "--InlineBackend.rc=figure.dpi=96",
 ]
-nbsphinx_execute = "always"
+nbsphinx_execute = "auto"
 nbsphinx_prompt_width = "0"
 nbsphinx_allow_errors = False
 
