+++ pyproject.toml
@@ -64,6 +64,7 @@ doc = [
     "sphinx-autodoc2",
     "sphinx-autobuild>=2024.2.4",
     "nbsphinx>=0.9.3",
+    "nbconvert",
 ]
 [project.urls]
 "Homepage" = "https://github.com/AUTODIAL/AutoEIS"
@@ -90,6 +91,9 @@ dev-dependencies = [
     "sphinx-autodoc2>=0.5.0",
     "sphinx-autobuild>=2024.2.4",
     "jupyterlab>=4.1.5",
+    "lcapy>=1.22",
+    "nbsphinx>=0.9.3",
+    "nbconvert>=7.16.3",
 ]
 
 [tool.hatch.version]
