+++ _sources/examples/circuit_basics.ipynb.txt
@@ -12,7 +12,14 @@
   {
    "cell_type": "code",
    "execution_count": 1,
-   "metadata": {},
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-04-13T19:41:04.280664Z",
+     "iopub.status.busy": "2024-04-13T19:41:04.280470Z",
+     "iopub.status.idle": "2024-04-13T19:41:07.976872Z",
+     "shell.execute_reply": "2024-04-13T19:41:07.976411Z"
+    }
+   },
    "outputs": [],
    "source": [
     "import autoeis as ae\n",
@@ -31,7 +38,14 @@
   {
    "cell_type": "code",
    "execution_count": 2,
-   "metadata": {},
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-04-13T19:41:07.979477Z",
+     "iopub.status.busy": "2024-04-13T19:41:07.979308Z",
+     "iopub.status.idle": "2024-04-13T19:41:07.981527Z",
+     "shell.execute_reply": "2024-04-13T19:41:07.981190Z"
+    }
+   },
    "outputs": [],
    "source": [
     "circuit = \"R1-[P2,R3]-C4-[[R5,C6],[L7,R8]]-R2\""
@@ -46,27 +60,23 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 8,
-   "metadata": {},
+   "execution_count": 3,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-04-13T19:41:07.983503Z",
+     "iopub.status.busy": "2024-04-13T19:41:07.983361Z",
+     "iopub.status.idle": "2024-04-13T19:41:07.994502Z",
+     "shell.execute_reply": "2024-04-13T19:41:07.994136Z"
+    }
+   },
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "Suggestion: add a constraint between nodes (25, 27) and (28, 26) for vertical graph\n"
[…]
