### 2024-04-13 bc7e75a
Add lcapy to doc deps
