+++ pyproject.toml
@@ -26,12 +26,14 @@ classifiers = [
 dependencies = [
     "arviz",
     "click",
+    "deprecated",
     "dill",
     "impedance",
     "ipython",
     "ipykernel",
     "ipywidgets",
     "jax",
+    "jinja2>=3.1.2",
     "julia",
     "juliacall==0.9.18",
     "juliapkg",
@@ -40,31 +42,29 @@ dependencies = [
     "numpy",
     "numpyro",
     "pandas",
+    "python-box",
     "psutil",
     "pyparsing>=3",
     "rich",
     "scikit-learn>=1.4",
     "seaborn",
     "tqdm",
-    "multiprocess",
-    "jinja2>=3.1.2",
-    "deprecated",
-    "python-box",
 ]
 [project.optional-dependencies]
 build = ["hatch"]
-test = ["pytest", "nbmake"]
+test = ["nbmake", "pytest"]
 doc = [
     "furo",
+    "lcapy",
     "linkify-it-py",
-    "myst-nb>=1.0.0",
+    "myst-nb",
     "myst-parser",
+    "nbsphinx",
+    "nbconvert",
     "sphinx",
     "sphinx-copybutton",
     "sphinx-autodoc2",
-    "sphinx-autobuild>=2024.2.4",
-    "nbsphinx>=0.9.3",
-    "nbconvert",
+    "sphinx-autobuild",
 ]
 [project.urls]
 "Homepage" = "https://github.com/AUTODIAL/AutoEIS"
@@ -79,21 +79,21 @@ build-backend = "hatchling.build"
 [tool.rye]
 managed = true
 dev-dependencies = [
-    "hatch>=1.9.4",
-    "pytest>=8.1.1",
-    "nbmake>=1.5.3",
     "furo>=2024.1.29",
+    "hatch>=1.9.4",
+    "jupyterlab>=4.1.5",
+    "lcapy>=1.22",
     "linkify-it-py>=2.0.3",
     "myst-nb>=1.0.0",
     "myst-parser>=2.0.0",
+    "nbmake>=1.5.3",
+    "nbsphinx>=0.9.3",
+    "nbconvert>=7.16.3",
+    "pytest>=8.1.1",
     "sphinx>=7.2.6",
     "sphinx-copybutton>=0.5.2",
     "sphinx-autodoc2>=0.5.0",
     "sphinx-autobuild>=2024.2.4",
-    "jupyterlab>=4.1.5",
-    "lcapy>=1.22",
-    "nbsphinx>=0.9.3",
-    "nbconvert>=7.16.3",
 ]
 
 [tool.hatch.version]
