### 2024-04-15 2776ac4
Factor act Bode plot as standalone function
