+++ src/autoeis/visualization.py
@@ -7,6 +7,8 @@ Collection of functions for visualizing EIS data and results.
    :toctree: generated/
 
     draw_circuit
+    plot_bode
+    plot_nyquist
     plot_impedance_combo
     plot_linKK_residuals
     plot_nyquist
@@ -36,6 +38,8 @@ log = logging.getLogger(__name__)
 
 __all__ = [
     "draw_circuit",
+    "plot_bode",
+    "plot_nyquist",
     "plot_impedance_combo",
     "plot_linKK_residuals",
     "plot_nyquist",
@@ -111,7 +115,7 @@ def draw_circuit(circuit: str) -> mpl.figure.Figure:
 def plot_nyquist(
     Z: np.ndarray[complex],
     fmt: str = "o-",
-    size: int = 4,
+    markersize: int = 6,
     color: str = None,
     alpha: int = 1,
     label: str = None,
@@ -125,8 +129,8 @@ def plot_nyquist(
         Impedance data.
     fmt: str, optional
         Format of the markers in the plot. Default is "o-".
-    size: int, optional
-        Size of the markers in the plot. Default is 4.
+    markersize: int, optional
+        Size of the markers in the plot. Default is 6.
     color: str, optional
         Color of the markers in the plot. Default is None.
     alpha: int, optional
@@ -149,7 +153,7 @@ def plot_nyquist(
         color = fmt[0]
         fmt = fmt[1:]
 
-    ax.plot(Z.real, -Z.imag, fmt, c=color, markersize=size, label=label, alpha=alpha)
+    ax.plot(Z.real, -Z.imag, fmt, c=color, markersize=markersize, label=label, alpha=alpha)
     ax.set_xlabel("Re(Z)")
     ax.set_ylabel("-Im(Z)")
     ax.axis("equal")
@@ -160,13 +164,73 @@ def plot_nyquist(
     return ax.figure, ax
 
 
+def plot_bode(
+    freq: np.ndarray[float],
+    Z: np.ndarray[complex],
+    fmt=".-",
+    markersize=6,
+    deg: bool = True,
+    ax: plt.Axes = None,
+) -> tuple[plt.Figure, plt.Axes]:
+    """Plots the Bode plot for the impedance data.
+
+    Parameters
+    ----------
+    freq: np.ndarray[float]
+        Frequencies corresponding to the impedance data.
+    Z: np.ndarray[complex]
+        Impedance data.
+    fmt: str, optional
+        Format of the markers in the plot. Default is ".-".
+    markersize: int, optional
+        Size of the markers in the plot. Default is 6.
+    deg: bool, optional
+        If True, plots the Bode plot in degrees. Default is True.
+    ax: plt.Axes, optional
+        Axes to plot on. Default is None.
+
+    Returns
+    -------
+    tuple[plt.Figure, plt.Axes]
+        Figure and axes of the plot.
+    """
+    if ax is None:
+        fig, ax = plt.subplots(figsize=(5.5, 3.5))
+
+    ax.plot(freq, np.abs(Z), fmt, label=r"$|Z|$", markersize=markersize)
+    ax.set_xscale("log")
+    ax2 = ax.twinx()
+    ax2.plot(
+        freq, np.angle(Z, deg=deg), fmt, markersize=markersize, color="b", label=r"$\phi$"
+    )
+    ax.set_xlabel("frequency (Hz)")
+    ax.set_ylabel(r"$|Z|$")
+    ax2.set_ylabel(rf"$\phi$ ({('deg' if deg else 'rad')})")
+
+    # Color y-axes for better visibility
+    ax2.yaxis.label.set_color("b")
[…]
