### 2024-04-15 3d007e0
Fix incorrect Bode plots (#110)
- Add `plot_bode` as a standalone function
