+++ _sources/generated/generated/autoeis.visualization.plot_bode.rst.txt
@@ -1,0 +1,6 @@
+plot\_bode
+==========
+
+.. currentmodule:: autoeis.visualization
+
+.. autofunction:: plot_bode
