+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-04-13T23:02:39.808531Z",
-     "iopub.status.busy": "2024-04-13T23:02:39.807953Z",
-     "iopub.status.idle": "2024-04-13T23:02:43.978840Z",
-     "shell.execute_reply": "2024-04-13T23:02:43.978337Z"
+     "iopub.execute_input": "2024-04-15T03:44:51.474637Z",
+     "iopub.status.busy": "2024-04-15T03:44:51.473488Z",
+     "iopub.status.idle": "2024-04-15T03:44:55.007608Z",
+     "shell.execute_reply": "2024-04-15T03:44:55.007168Z"
     }
    },
    "outputs": [],
@@ -40,10 +40,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-04-13T23:02:43.981627Z",
-     "iopub.status.busy": "2024-04-13T23:02:43.981465Z",
-     "iopub.status.idle": "2024-04-13T23:02:43.989834Z",
-     "shell.execute_reply": "2024-04-13T23:02:43.989480Z"
+     "iopub.execute_input": "2024-04-15T03:44:55.010108Z",
+     "iopub.status.busy": "2024-04-15T03:44:55.009947Z",
+     "iopub.status.idle": "2024-04-15T03:44:55.018177Z",
+     "shell.execute_reply": "2024-04-15T03:44:55.017810Z"
     }
    },
    "outputs": [],
@@ -63,10 +63,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-04-13T23:02:43.991680Z",
-     "iopub.status.busy": "2024-04-13T23:02:43.991582Z",
-     "iopub.status.idle": "2024-04-13T23:02:46.099498Z",
-     "shell.execute_reply": "2024-04-13T23:02:46.099049Z"
+     "iopub.execute_input": "2024-04-15T03:44:55.020048Z",
+     "iopub.status.busy": "2024-04-15T03:44:55.019954Z",
+     "iopub.status.idle": "2024-04-15T03:44:57.140315Z",
+     "shell.execute_reply": "2024-04-15T03:44:57.139834Z"
     }
    },
    "outputs": [
@@ -110,10 +110,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-04-13T23:02:46.101995Z",
-     "iopub.status.busy": "2024-04-13T23:02:46.101612Z",
-     "iopub.status.idle": "2024-04-13T23:02:46.105311Z",
-     "shell.execute_reply": "2024-04-13T23:02:46.104935Z"
+     "iopub.execute_input": "2024-04-15T03:44:57.142924Z",
+     "iopub.status.busy": "2024-04-15T03:44:57.142487Z",
+     "iopub.status.idle": "2024-04-15T03:44:57.152087Z",
+     "shell.execute_reply": "2024-04-15T03:44:57.151740Z"
     }
    },
    "outputs": [
@@ -144,10 +144,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-04-13T23:02:46.107258Z",
-     "iopub.status.busy": "2024-04-13T23:02:46.107135Z",
-     "iopub.status.idle": "2024-04-13T23:02:46.109659Z",
-     "shell.execute_reply": "2024-04-13T23:02:46.109278Z"
+     "iopub.execute_input": "2024-04-15T03:44:57.154011Z",
[…]
