### 2024-04-25 3d4206b
Disable progress bar by default
