### 2024-04-26 6e7928f
Add workflow as banner to docs homepage
