### 2024-04-26 8eeeebd
perform_bayesian_inference now accepts priors as optional arg
