### 2024-04-26 bad73b4
Log exception message when inference fails
