### 2024-04-26 e4cef2a
Remove unused logic
