### 2024-05-01 2de6ea8
Make a deep copy of priors before sampling (to keep JAX happy)
