### 2024-05-01 bc370fc
Return a deep-copied instance of priors dict (to keep JAX happy)
