### 2024-05-03 5d45fab
Add verbose as optional arg to cli install
