+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-01T20:44:32.998591Z",
-     "iopub.status.busy": "2024-05-01T20:44:32.998053Z",
-     "iopub.status.idle": "2024-05-01T20:44:36.985543Z",
-     "shell.execute_reply": "2024-05-01T20:44:36.985029Z"
+     "iopub.execute_input": "2024-05-03T06:13:51.627851Z",
+     "iopub.status.busy": "2024-05-03T06:13:51.627261Z",
+     "iopub.status.idle": "2024-05-03T06:13:55.497618Z",
+     "shell.execute_reply": "2024-05-03T06:13:55.497100Z"
     }
    },
    "outputs": [],
@@ -40,10 +40,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-01T20:44:36.988446Z",
-     "iopub.status.busy": "2024-05-01T20:44:36.988247Z",
-     "iopub.status.idle": "2024-05-01T20:44:36.996412Z",
-     "shell.execute_reply": "2024-05-01T20:44:36.996075Z"
+     "iopub.execute_input": "2024-05-03T06:13:55.500264Z",
+     "iopub.status.busy": "2024-05-03T06:13:55.500112Z",
+     "iopub.status.idle": "2024-05-03T06:13:55.508521Z",
+     "shell.execute_reply": "2024-05-03T06:13:55.508167Z"
     }
    },
    "outputs": [],
@@ -63,10 +63,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-01T20:44:36.998316Z",
-     "iopub.status.busy": "2024-05-01T20:44:36.998226Z",
-     "iopub.status.idle": "2024-05-01T20:44:39.167022Z",
-     "shell.execute_reply": "2024-05-01T20:44:39.166368Z"
+     "iopub.execute_input": "2024-05-03T06:13:55.510411Z",
+     "iopub.status.busy": "2024-05-03T06:13:55.510316Z",
+     "iopub.status.idle": "2024-05-03T06:13:57.614786Z",
+     "shell.execute_reply": "2024-05-03T06:13:57.614335Z"
     }
    },
    "outputs": [
@@ -110,10 +110,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-01T20:44:39.170651Z",
-     "iopub.status.busy": "2024-05-01T20:44:39.170157Z",
-     "iopub.status.idle": "2024-05-01T20:44:39.174379Z",
-     "shell.execute_reply": "2024-05-01T20:44:39.173867Z"
+     "iopub.execute_input": "2024-05-03T06:13:57.617252Z",
+     "iopub.status.busy": "2024-05-03T06:13:57.616874Z",
+     "iopub.status.idle": "2024-05-03T06:13:57.620547Z",
+     "shell.execute_reply": "2024-05-03T06:13:57.620197Z"
     }
    },
    "outputs": [
@@ -144,10 +144,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-01T20:44:39.176496Z",
-     "iopub.status.busy": "2024-05-01T20:44:39.176386Z",
-     "iopub.status.idle": "2024-05-01T20:44:39.179429Z",
-     "shell.execute_reply": "2024-05-01T20:44:39.178994Z"
+     "iopub.execute_input": "2024-05-03T06:13:57.622508Z",
[…]
