### 2024-05-07 e9c6268
Fit circuit parameters if still uninitialized before inference
