+++ src/autoeis/core.py
@@ -588,18 +588,21 @@ def perform_bayesian_inference(
         else:
             p0 = circuits["Parameters"].tolist()
         circuits = circuits["circuitstring"].tolist()
+
+    # Refine the initial guess if requested or if p0 is still uninitialized
+    if refine_p0 or p0[0] is None:
+        for i, (circuit, p0_) in enumerate(zip(circuits, p0)):
+            p0[i] = utils.fit_circuit_parameters(circuit, freq, Z, p0=p0_, iters=25)
+            # If circuit fitter didn't converge, use the initial guess
+            p0[i] = p0_ if p0[i] is None else p0[i]
+
+    # Validate inputs' types and lengths
     for circuit, p0_ in zip(circuits, p0):
         assert isinstance(circuit, str), f"Circuit must be a string: {circuit}"
         valid_p0_types = (dict, np.ndarray, type(None))
         assert isinstance(p0_, valid_p0_types), f"Invalid p0 type: {p0_}"
         num_params = len(parser.get_parameter_labels(circuit))
         assert len(p0_) == num_params, f"Invalid p0 length: {p0_}"
-
-    if refine_p0:
-        for i, (circuit, p0_) in enumerate(zip(circuits, p0)):
-            p0[i] = utils.fit_circuit_parameters(circuit, freq, Z, p0=p0_)
-            # If circuit fitter didn't converge, use the initial guess
-            p0[i] = p0_ if p0[i] is None else p0[i]
 
     # Short-circuit if no circuits are provided
     if len(circuits) == 0:
