+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-03T06:13:51.627851Z",
-     "iopub.status.busy": "2024-05-03T06:13:51.627261Z",
-     "iopub.status.idle": "2024-05-03T06:13:55.497618Z",
-     "shell.execute_reply": "2024-05-03T06:13:55.497100Z"
+     "iopub.execute_input": "2024-05-07T19:07:10.252860Z",
+     "iopub.status.busy": "2024-05-07T19:07:10.252745Z",
+     "iopub.status.idle": "2024-05-07T19:07:15.252878Z",
+     "shell.execute_reply": "2024-05-07T19:07:15.252332Z"
     }
    },
    "outputs": [],
@@ -40,10 +40,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-03T06:13:55.500264Z",
-     "iopub.status.busy": "2024-05-03T06:13:55.500112Z",
-     "iopub.status.idle": "2024-05-03T06:13:55.508521Z",
-     "shell.execute_reply": "2024-05-03T06:13:55.508167Z"
+     "iopub.execute_input": "2024-05-07T19:07:15.255637Z",
+     "iopub.status.busy": "2024-05-07T19:07:15.255478Z",
+     "iopub.status.idle": "2024-05-07T19:07:15.261221Z",
+     "shell.execute_reply": "2024-05-07T19:07:15.260806Z"
     }
    },
    "outputs": [],
@@ -63,10 +63,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-03T06:13:55.510411Z",
-     "iopub.status.busy": "2024-05-03T06:13:55.510316Z",
-     "iopub.status.idle": "2024-05-03T06:13:57.614786Z",
-     "shell.execute_reply": "2024-05-03T06:13:57.614335Z"
+     "iopub.execute_input": "2024-05-07T19:07:15.263096Z",
+     "iopub.status.busy": "2024-05-07T19:07:15.263002Z",
+     "iopub.status.idle": "2024-05-07T19:07:18.057096Z",
+     "shell.execute_reply": "2024-05-07T19:07:18.055990Z"
     }
    },
    "outputs": [
@@ -110,10 +110,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-03T06:13:57.617252Z",
-     "iopub.status.busy": "2024-05-03T06:13:57.616874Z",
-     "iopub.status.idle": "2024-05-03T06:13:57.620547Z",
-     "shell.execute_reply": "2024-05-03T06:13:57.620197Z"
+     "iopub.execute_input": "2024-05-07T19:07:18.061752Z",
+     "iopub.status.busy": "2024-05-07T19:07:18.061055Z",
+     "iopub.status.idle": "2024-05-07T19:07:18.066995Z",
+     "shell.execute_reply": "2024-05-07T19:07:18.066182Z"
     }
    },
    "outputs": [
@@ -144,10 +144,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-03T06:13:57.622508Z",
-     "iopub.status.busy": "2024-05-03T06:13:57.622386Z",
-     "iopub.status.idle": "2024-05-03T06:13:57.630966Z",
-     "shell.execute_reply": "2024-05-03T06:13:57.630584Z"
+     "iopub.execute_input": "2024-05-07T19:07:18.070715Z",
[…]
