### 2024-05-08 44b01c5
feat: Refactor perform_bayesian_inference to perform parameter refitting in parallel
