### 2024-05-08 fe7c6ed
chore: Update n_jobs calculation in distribute_task function
