### 2024-05-13 ba30c2d
build: Lift version requirement on pyjuliapkg (fixed upstream)
