### 2024-05-14 4216cb8
fix: Disable juliacall IPython extension b/c it clashes with out `quiet` mode
