+++ src/autoeis/core.py
@@ -54,8 +54,9 @@ warnings.filterwarnings("ignore", category=Warning, module="arviz.*")
 log = logging.getLogger(__name__)
 
 # Initialize Julia runtime
-julia_helpers.ensure_julia_deps_ready()
-jl = julia_helpers.init_julia()
+os.environ["PYTHON_JULIACALL_AUTOLOAD_IPYTHON_EXTENSION"] = "no"
+julia_helpers.ensure_julia_deps_ready(quiet=True)
+jl = julia_helpers.init_julia(quiet=True)
 ec = julia_helpers.import_backend(jl)
 
 __all__ = [
