+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,32 +14,13 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-13T14:44:54.005794Z",
-     "iopub.status.busy": "2024-05-13T14:44:54.005614Z",
-     "iopub.status.idle": "2024-05-13T14:45:00.169933Z",
-     "shell.execute_reply": "2024-05-13T14:45:00.168888Z"
+     "iopub.execute_input": "2024-05-14T05:09:21.903992Z",
+     "iopub.status.busy": "2024-05-14T05:09:21.903464Z",
+     "iopub.status.idle": "2024-05-14T05:09:26.323181Z",
+     "shell.execute_reply": "2024-05-14T05:09:26.322313Z"
     }
    },
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "Error in callback _flush_stdio (for post_execute), with arguments args (),kwargs {}:\n"
-     ]
-    },
-    {
-     "ename": "ValueError",
-     "evalue": "I/O operation on closed file.",
-     "output_type": "error",
-     "traceback": [
-      "\u001b[0;31m---------------------------------------------------------------------------\u001b[0m",
-      "\u001b[0;31mValueError\u001b[0m                                Traceback (most recent call last)",
-      "File \u001b[0;32m~/.julia/packages/PythonCall/S5MOg/src/JlWrap/any.jl:223\u001b[0m, in \u001b[0;36m__call__\u001b[0;34m(self, *args, **kwargs)\u001b[0m\n\u001b[1;32m    221\u001b[0m     return ValueBase.__dir__(self) + self._jl_callmethod($(pyjl_methodnum(pyjlany_dir)))\n\u001b[1;32m    222\u001b[0m def __call__(self, *args, **kwargs):\n\u001b[0;32m--> 223\u001b[0m     return self._jl_callmethod($(pyjl_methodnum(pyjlany_call)), args, kwargs)\n\u001b[1;32m    224\u001b[0m def __bool__(self):\n\u001b[1;32m    225\u001b[0m     return True\n",
-      "\u001b[0;31mValueError\u001b[0m: I/O operation on closed file."
-     ]
-    }
-   ],
+   "outputs": [],
    "source": [
     "import autoeis as ae\n",
     "ae.visualization.set_plot_style()"
@@ -59,32 +40,13 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-13T14:45:00.173329Z",
-     "iopub.status.busy": "2024-05-13T14:45:00.173087Z",
-     "iopub.status.idle": "2024-05-13T14:45:00.187264Z",
-     "shell.execute_reply": "2024-05-13T14:45:00.186796Z"
+     "iopub.execute_input": "2024-05-14T05:09:26.327818Z",
+     "iopub.status.busy": "2024-05-14T05:09:26.327475Z",
+     "iopub.status.idle": "2024-05-14T05:09:26.330959Z",
+     "shell.execute_reply": "2024-05-14T05:09:26.330086Z"
     }
    },
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "Error in callback _flush_stdio (for post_execute), with arguments args (),kwargs {}:\n"
-     ]
-    },
-    {
-     "ename": "ValueError",
-     "evalue": "I/O operation on closed file.",
-     "output_type": "error",
-     "traceback": [
[…]
