### 2024-05-14 3d8ef4e
feat: Add `identify_duplicate_circuits` utility function for identifying duplicate circuits
