### 2024-05-14 afcb2eb
chore: Update `distribute_task` function to handle errors during multiprocessing
