### 2024-05-14 a41eb64
fix: Take abs of the error when evaluating convergence criterion in `fit_circuit_parameters`
