### 2024-05-14 1009dbb
fix: Update `fit_circuit_parameters` to handle arbitrary exceptions not just `RuntimeError`s
