### 2024-05-14 fdf140e
fix: Update `perform_bayesian_inference` to handle exceptions during p0 refinement
