### 2024-05-14 68b2c8c
fix: Update p0 type assertion to allow for None (when circuit fitter fails)
