### 2024-05-14 31239f8
refactor: Use `distribute_task` helper to run parallel inference
