### 2024-05-14 f9b4fcd
chore: Add TODO: deal with failed inferences during Bayesian Inference
