### 2024-05-14 77adcf6
fix: Don't use geometric mean for WAIC (sum) b/c of WAIC could be positive too (leading to NaNs)
