### 2024-05-15 c8d6e24
refactor: Use geometric instead of arithmetic mean for error calculation in `fit_circuit_parameters`
