+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-14T18:20:25.069002Z",
-     "iopub.status.busy": "2024-05-14T18:20:25.068862Z",
-     "iopub.status.idle": "2024-05-14T18:20:30.576375Z",
-     "shell.execute_reply": "2024-05-14T18:20:30.575784Z"
+     "iopub.execute_input": "2024-05-15T06:30:59.889961Z",
+     "iopub.status.busy": "2024-05-15T06:30:59.889550Z",
+     "iopub.status.idle": "2024-05-15T06:31:05.167450Z",
+     "shell.execute_reply": "2024-05-15T06:31:05.166002Z"
     }
    },
    "outputs": [],
@@ -40,10 +40,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-14T18:20:30.581148Z",
-     "iopub.status.busy": "2024-05-14T18:20:30.580898Z",
-     "iopub.status.idle": "2024-05-14T18:20:30.583438Z",
-     "shell.execute_reply": "2024-05-14T18:20:30.582773Z"
+     "iopub.execute_input": "2024-05-15T06:31:05.175041Z",
+     "iopub.status.busy": "2024-05-15T06:31:05.174284Z",
+     "iopub.status.idle": "2024-05-15T06:31:05.178624Z",
+     "shell.execute_reply": "2024-05-15T06:31:05.177877Z"
     }
    },
    "outputs": [],
@@ -63,10 +63,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-14T18:20:30.586791Z",
-     "iopub.status.busy": "2024-05-14T18:20:30.586603Z",
-     "iopub.status.idle": "2024-05-14T18:20:33.148495Z",
-     "shell.execute_reply": "2024-05-14T18:20:33.147467Z"
+     "iopub.execute_input": "2024-05-15T06:31:05.182385Z",
+     "iopub.status.busy": "2024-05-15T06:31:05.182183Z",
+     "iopub.status.idle": "2024-05-15T06:31:08.815276Z",
+     "shell.execute_reply": "2024-05-15T06:31:08.813828Z"
     }
    },
    "outputs": [
@@ -110,10 +110,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-14T18:20:33.153710Z",
-     "iopub.status.busy": "2024-05-14T18:20:33.153087Z",
-     "iopub.status.idle": "2024-05-14T18:20:33.233702Z",
-     "shell.execute_reply": "2024-05-14T18:20:33.233193Z"
+     "iopub.execute_input": "2024-05-15T06:31:08.820926Z",
+     "iopub.status.busy": "2024-05-15T06:31:08.820484Z",
+     "iopub.status.idle": "2024-05-15T06:31:08.826528Z",
+     "shell.execute_reply": "2024-05-15T06:31:08.825735Z"
     }
    },
    "outputs": [
@@ -144,10 +144,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-14T18:20:33.236715Z",
-     "iopub.status.busy": "2024-05-14T18:20:33.236544Z",
-     "iopub.status.idle": "2024-05-14T18:20:33.239691Z",
-     "shell.execute_reply": "2024-05-14T18:20:33.239087Z"
+     "iopub.execute_input": "2024-05-15T06:31:08.833241Z",
[…]
