### 2024-05-16 #6256 @ma-sadeghi
Hi @lucydot,

Sorry for the late reply; things have been hectic recently.

We did fix the issue with the notebook, but we want to add more granular notebooks. Currently, we have one big notebook explaining too much. I'll ping you here once those are ready. The ETA is about 1-2 weeks. I hope that's okay. I really appreciate your patience with this review.
