### 2024-05-17 85595ec
refactor: Update `set_plot_style` to include console width parameter (only used when in VS Code)
