+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-16T05:33:50.756868Z",
-     "iopub.status.busy": "2024-05-16T05:33:50.756477Z",
-     "iopub.status.idle": "2024-05-16T05:33:55.204075Z",
-     "shell.execute_reply": "2024-05-16T05:33:55.203545Z"
+     "iopub.execute_input": "2024-05-17T08:26:20.524415Z",
+     "iopub.status.busy": "2024-05-17T08:26:20.524256Z",
+     "iopub.status.idle": "2024-05-17T08:26:26.362188Z",
+     "shell.execute_reply": "2024-05-17T08:26:26.361389Z"
     }
    },
    "outputs": [],
@@ -40,10 +40,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-16T05:33:55.208672Z",
-     "iopub.status.busy": "2024-05-16T05:33:55.208355Z",
-     "iopub.status.idle": "2024-05-16T05:33:55.211074Z",
-     "shell.execute_reply": "2024-05-16T05:33:55.210567Z"
+     "iopub.execute_input": "2024-05-17T08:26:26.366943Z",
+     "iopub.status.busy": "2024-05-17T08:26:26.366565Z",
+     "iopub.status.idle": "2024-05-17T08:26:26.369685Z",
+     "shell.execute_reply": "2024-05-17T08:26:26.368941Z"
     }
    },
    "outputs": [],
@@ -63,10 +63,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-16T05:33:55.213771Z",
-     "iopub.status.busy": "2024-05-16T05:33:55.213624Z",
-     "iopub.status.idle": "2024-05-16T05:33:57.411606Z",
-     "shell.execute_reply": "2024-05-16T05:33:57.411115Z"
+     "iopub.execute_input": "2024-05-17T08:26:26.373203Z",
+     "iopub.status.busy": "2024-05-17T08:26:26.372909Z",
+     "iopub.status.idle": "2024-05-17T08:26:29.774749Z",
+     "shell.execute_reply": "2024-05-17T08:26:29.774029Z"
     }
    },
    "outputs": [
@@ -110,10 +110,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-16T05:33:57.414164Z",
-     "iopub.status.busy": "2024-05-16T05:33:57.413737Z",
-     "iopub.status.idle": "2024-05-16T05:33:57.417341Z",
-     "shell.execute_reply": "2024-05-16T05:33:57.416957Z"
+     "iopub.execute_input": "2024-05-17T08:26:29.778767Z",
+     "iopub.status.busy": "2024-05-17T08:26:29.778098Z",
+     "iopub.status.idle": "2024-05-17T08:26:29.784195Z",
+     "shell.execute_reply": "2024-05-17T08:26:29.783629Z"
     }
    },
    "outputs": [
@@ -144,10 +144,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-16T05:33:57.419338Z",
-     "iopub.status.busy": "2024-05-16T05:33:57.419207Z",
-     "iopub.status.idle": "2024-05-16T05:33:57.427807Z",
-     "shell.execute_reply": "2024-05-16T05:33:57.427417Z"
+     "iopub.execute_input": "2024-05-17T08:26:29.787490Z",
[…]
