### 2024-06-06 76b1889
chore: Update deprecated API to set retina backend for mpl
