### 2024-06-06 f73eae6
refactor: Validate circuits before checking equivalence
