+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,14 +14,15 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-17T08:26:20.524415Z",
-     "iopub.status.busy": "2024-05-17T08:26:20.524256Z",
-     "iopub.status.idle": "2024-05-17T08:26:26.362188Z",
-     "shell.execute_reply": "2024-05-17T08:26:26.361389Z"
+     "iopub.execute_input": "2024-06-06T14:44:55.527926Z",
+     "iopub.status.busy": "2024-06-06T14:44:55.527292Z",
+     "iopub.status.idle": "2024-06-06T14:44:59.815071Z",
+     "shell.execute_reply": "2024-06-06T14:44:59.814561Z"
     }
    },
    "outputs": [],
    "source": [
+    "import numpy as np\n",
     "import autoeis as ae\n",
     "ae.visualization.set_plot_style()"
    ]
@@ -40,10 +41,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-17T08:26:26.366943Z",
-     "iopub.status.busy": "2024-05-17T08:26:26.366565Z",
-     "iopub.status.idle": "2024-05-17T08:26:26.369685Z",
-     "shell.execute_reply": "2024-05-17T08:26:26.368941Z"
+     "iopub.execute_input": "2024-06-06T14:44:59.818272Z",
+     "iopub.status.busy": "2024-06-06T14:44:59.818075Z",
+     "iopub.status.idle": "2024-06-06T14:44:59.826120Z",
+     "shell.execute_reply": "2024-06-06T14:44:59.825771Z"
     }
    },
    "outputs": [],
@@ -63,10 +64,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-05-17T08:26:26.373203Z",
-     "iopub.status.busy": "2024-05-17T08:26:26.372909Z",
-     "iopub.status.idle": "2024-05-17T08:26:29.774749Z",
-     "shell.execute_reply": "2024-05-17T08:26:29.774029Z"
[…]
