### 2024-06-20 efa43a4
refactor: Make matplotlib args in `plot_nyquist` and `plot_bode functions` kwarg-only
