### 2024-06-20 401cd74
refactor: Add `is_iterable` and `is_nested_iterable` helpers to utils
