### 2024-06-20 b78b8a0
refactor: Add `InferenceResult` and `ImpedanceData` container classes
