### 2024-06-20 09f3dd2
refactor: Add whole bunch of objects validators: circuit, p0, priors, seed, and impedance
