### 2024-06-20 beba800
refactor: Major refactor of `perform_bayesian_inference`, now can do SCSD, MCSD, and SCMD
