+++ src/autoeis/core.py
@@ -27,9 +27,8 @@ import numpy as np
 import numpyro
 import pandas as pd
 import psutil
-from box import Box
 from deprecated import deprecated
-from impedance.validation import linKK
+
 from jax import config
 from mpire import WorkerPool
 from numpyro.distributions import Distribution
@@ -40,15 +39,14 @@ from tqdm.auto import tqdm
 from autoeis import io, julia_helpers, metrics, parser, utils
 from autoeis.models import circuit_regression_complex
 
+from .utils import InferenceResult
+
 # Enforce FP64, otherwise circuit fitter fails (because FP32 gradients not sufficient?)
 config.update("jax_enable_x64", True)
 # Tell JAX to use CPUs to avoid the annoying "GPU might be present" warning
 config.update("jax_platforms", "cpu")
 # EIS datasets are not big-enough -> CPU much faster than GPU
 numpyro.set_platform("cpu")
-
-# TODO: Suppress output until ECSHackWeek/impedance.py/issues/280 is fixed
-linKK = utils.suppress_output_legacy(linKK)
 
 warnings.filterwarnings("ignore", category=Warning, module="arviz.*")
 log = logging.getLogger(__name__)
@@ -61,7 +59,6 @@ ec = julia_helpers.import_backend(jl)
 
 __all__ = [
     "perform_full_analysis",
-    "preprocess_impedance_data",
     "generate_equivalent_circuits",
     "filter_implausible_circuits",
     "perform_bayesian_inference",
@@ -529,9 +526,9 @@ def _validate_seed(seed, num_splits=1):
 
 
 def perform_bayesian_inference(
-    circuits: pd.DataFrame | Iterable[str] | str,
-    freq: np.ndarray[float],
-    Z: np.ndarray[complex],
+    circuit: str | Iterable[str] | pd.DataFrame,
+    freq: np.ndarray[float] | Iterable[float],
+    Z: np.ndarray[complex] | Iterable[complex],
     p0: Iterable[float]
     | Mapping[str, float]
     | Iterable[Iterable[float]]
@@ -543,17 +540,23 @@ def perform_bayesian_inference(
     seed: int | jax.Array = None,
     progress_bar: bool = True,
     refine_p0: bool = False,
-) -> list[tuple[MCMC, int]]:
+    parallel: bool = True,
+) -> list[InferenceResult] | InferenceResult:
     """Performs Bayesian inference on the circuits based on impedance data.
 
     Parameters
     ----------
-    circuits : pd.DataFrame | Iterable[str] | str
-        Dataframe containing circuits or list of circuit strings.
-    freq: np.ndarray[float]
-        Frequency data corresponding to the impedance data.
-    Z : np.ndarray[complex]
-        Impedance data as a complex array.
+    circuit : str | Iterable[str] | pd.DataFrame
+        Dataframe containing circuits, list of circuit strings, or a single
+        circuit string.
+    freq: np.ndarray[float] | Iterable[float]
+        Frequency array corresponding to the impedance data. If a list of
+        frequencies is provided, separate inferences will be performed for
+        each frequency array.
+    Z : np.ndarray[complex] | Iterable[np.ndarray[complex]]
+        Impedance data as a complex array. If a list of impedance arrays is
+        provided, separate inferences will be performed for each impedance
+        array.
     p0 : Iterable[float] \
             | Mapping[str, float] \
             | Iterable[Iterable[float]] \
@@ -574,94 +577,128 @@ def perform_bayesian_inference(
     refine_p0 : bool, optional
         If True, the initial guess for the circuit parameters will be refined
         using the circuit fitter (default is False).
+    parallel : bool, optional
+        If True, the MCMC chains will be run in parallel (default is True).
 
     Returns
[…]
