### 2024-06-20 2a763a6
test: Add unit tests for SCSD, MCSD, SCMD inferences
