+++ tests/test_core.py
@@ -1,53 +1,52 @@
+import autoeis as ae
 import numpy as np
-import numpyro
 import pandas as pd
 import pytest
-from autoeis import core, io, utils
 
 
 def test_compute_ohmic_resistance():
     circuit_string = "R1-[P2,P3-R4]"
-    circuit_fn = utils.generate_circuit_fn_impedance_backend(circuit_string)
+    circuit_fn = ae.utils.generate_circuit_fn_impedance_backend(circuit_string)
     R1 = 250
     parameters = np.array([R1, 1e-3, 0.1, 5e-5, 0.8, 10])
     freq = np.logspace(-3, 3, 1000)
     Z = circuit_fn(freq, parameters)
-    R = core.compute_ohmic_resistance(freq, Z)
+    R = ae.core.compute_ohmic_resistance(freq, Z)
     np.testing.assert_allclose(R, R1, rtol=0.15)
 
 
 def test_compute_ohmic_resistance_missing_high_freq():
     circuit_string = "R1-[P2,P3-R4]"
-    circuit_fn = utils.generate_circuit_fn_impedance_backend(circuit_string)
+    circuit_fn = ae.utils.generate_circuit_fn_impedance_backend(circuit_string)
     R1 = 250
     parameters = np.array([R1, 1e-3, 0.1, 5e-5, 0.8, 10])
     freq = np.logspace(-3, 0, 1000)
     Z = circuit_fn(freq, parameters)
-    R = core.compute_ohmic_resistance(freq, Z)
+    R = ae.core.compute_ohmic_resistance(freq, Z)
     # When high frequency measurements are missing, Re(Z) @ max(freq) is good approximation
     Zreal_at_high_freq = Z.real[np.argmax(freq)]
     np.testing.assert_allclose(R, Zreal_at_high_freq)
 
 
 def test_preprocess_impedance_data():
-    freq, Z = io.load_test_dataset()
+    freq, Z = ae.io.load_test_dataset()
     # Test various tolerances for linKK validation
-    freq_prep, Z_prep = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
+    freq_prep, Z_prep = ae.utils.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
     assert len(Z_prep) == len(freq_prep)
     assert len(Z_prep) == 60
-    freq_prep, Z_prep = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-3)
+    freq_prep, Z_prep = ae.utils.preprocess_impedance_data(freq, Z, tol_linKK=5e-3)
     assert len(Z_prep) == len(freq_prep)
     assert len(Z_prep) == 50
     # Test return_aux=True
-    _, _, aux = core.preprocess_impedance_data(freq, Z, return_aux=True)
+    _, _, aux = ae.utils.preprocess_impedance_data(freq, Z, return_aux=True)
     assert list(aux.keys()) == ["res", "rmse"]
     assert list(aux["res"].keys()) == ["real", "imag"]
 
 
 def test_gep():
     def test_gep_serial():
-        freq, Z = io.load_test_dataset()
-        freq, Z = core.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
+        freq, Z = ae.io.load_test_dataset()
+        freq, Z = ae.utils.preprocess_impedance_data(freq, Z, tol_linKK=5e-2)
         kwargs = {
             "iters": 2,
             "complexity": 2,
@@ -56,13 +55,12 @@ def test_gep():
             "tol": 1e10,
             "parallel": False,
         }
-        circuits = core.generate_equivalent_circuits(freq, Z, **kwargs)
[…]
