+++ tests/test_utils.py
@@ -1,6 +1,5 @@
+import autoeis as ae
 import numpy as np
-import numpyro
-from autoeis import core, io, julia_helpers, parser, utils
 
 # Real numbers
 x1 = np.random.rand(10)
@@ -13,13 +12,13 @@ y2 = x2 + np.zeros(10) * 1j
 circuit_string = "R1-[P2,R3]"
 p0_dict = {"R1": 250, "P2w": 1e-3, "P2n": 0.5, "R3": 10.0}
 p0_vals = list(p0_dict.values())
-circuit_fn_gt = utils.generate_circuit_fn_impedance_backend(circuit_string)
+circuit_fn_gt = ae.utils.generate_circuit_fn_impedance_backend(circuit_string)
 freq = np.logspace(-3, 3, 1000)
 Z = circuit_fn_gt(freq, p0_vals)
 
 
 def test_fit_circuit_parameters_without_x0():
-    p_dict = utils.fit_circuit_parameters(circuit_string, freq, Z, iters=10)
+    p_dict = ae.utils.fit_circuit_parameters(circuit_string, freq, Z, iters=10)
     p_fit = list(p_dict.values())
     assert np.allclose(p_fit, p0_vals, rtol=0.01)
 
@@ -27,20 +26,20 @@ def test_fit_circuit_parameters_without_x0():
 def test_fit_circuit_parameters_with_x0():
     # Add some noise to the initial guess to test robustness
     p0 = p0_vals + np.random.rand(len(p0_vals)) * p0_vals * 0.5
-    p_dict = utils.fit_circuit_parameters(circuit_string, freq, Z, p0)
+    p_dict = ae.utils.fit_circuit_parameters(circuit_string, freq, Z, p0)
     p_fit = list(p_dict.values())
     assert np.allclose(p_fit, p0_vals, rtol=0.01)
 
 
 def test_generate_circuit_fn():
     circuit = "R0-C1-[P2,R3]-[P4-[P5,C6],[L7,R8]]"
-    num_params = parser.count_parameters(circuit)
+    num_params = ae.parser.count_parameters(circuit)
     freq = np.array([1, 10, 100])
     p = np.random.rand(num_params)
-    circuit_fn = utils.generate_circuit_fn(circuit)
+    circuit_fn = ae.utils.generate_circuit_fn(circuit)
     Z_py = circuit_fn(freq, p)
-    Main = julia_helpers.init_julia()
-    ec = julia_helpers.import_backend(Main)
+    Main = ae.julia_helpers.init_julia()
+    ec = ae.julia_helpers.import_backend(Main)
     Z_jl = np.array([ec.get_target_impedance(circuit, p, f) for f in freq])
     np.testing.assert_allclose(Z_py, Z_jl)
 
@@ -55,7 +54,7 @@ def test_circuit_complexity():
         "R1-[R2,R3]-[[C4,L5]-P6]-[R7,[R8,[C9,L10]]]": [0, 1, 1, 2, 2, 1, 1, 2, 3, 3],
     }
     for cstr, cc in circuit_complexity_dict.items():
-        assert utils.circuit_complexity(cstr) == cc
+        assert ae.utils.circuit_complexity(cstr) == cc
 
 
 def test_are_circuits_equivalent():
@@ -68,28 +67,25 @@ def test_are_circuits_equivalent():
     ]
     for row in testset:
         c1, c2, eq = row
-        assert utils.are_circuits_equivalent(c1, c2) == eq
+        assert ae.utils.are_circuits_equivalent(c1, c2) == eq
 
 
 def test_eval_posterior_predictive():
     # Load test dataset
-    freq, Z = io.load_test_dataset()
-    circuits = io.load_test_circuits(filtered=True)
+    freq, Z = ae.io.load_test_dataset()
[…]
