### 2024-06-24 c7e66e0
refactor: Add `flush_streams` helper to utils.py
