### 2024-06-25 a23595b
chore: Fix typo in log warning message
