### 2024-06-25 ce896e2
refactor: Add `utils.flush_streams` so progress bars render properly
