### 2024-06-25 95c39bd
chore: Improve post-filtering warning message for empty circuits
