### 2024-06-25 9e1c1ec
feat: Add `is_ndarray_like` helper function to utils.py
