### 2024-06-25 a5a02d6
chore: Update type hints in utils.py to use Iterable instead of list
