### 2024-06-25 41bf562
refactor: Remove unused `circuit` from `parse_initial_guess` args
