### 2024-06-25 418d23b
feat: Add `generate_initial_guess` helper to utils.py
