### 2024-06-26 e204300
refactor: Standardize plotting functions to return ax object(s)
