### 2024-07-22 47907b3
docs: Better error message
