### 2024-07-22 757b62f
refactor: Deep copy circuits dataframe to avoid modifying original
