### 2024-07-22 551d887
bug: p0 must be replicated for distribute_task to work
