### 2024-07-22 f9931e3
bug: Handle edge case where p0 was not found
