### 2024-07-22 b61037b
refactor: Ensure seed is a list for MCMD compatibility
