### 2024-07-22 94d9525
refactor: Propagate recent API changes throughout the codebase
