### 2024-07-22 c46b393
docs: Update docs to include new example notebooks + update basic usage
