### 2024-07-22 20ae908
refactor: Update `perform_full_analysis` per recent API changes
