### 2024-07-22 60de4f6
docs: Remove moved function from docstrings
