### 2024-07-22 0af90a1
feat: Add new toy dataset (battery cycling)
