### 2024-07-22 612d8f0
bug: Fixed edge case where handling R1 messed up with R10, etc.
