### 2024-07-22 8f4b71e
refactor: `validate_parameter` now optionally raises Exception
