+++ src/autoeis/parser.py
@@ -64,7 +64,7 @@ def validate_circuit(circuit: str) -> bool:
     return True  # If all checks pass, the circuit is considered valid
 
 
-def validate_parameter(p: str) -> bool:
+def validate_parameter(p: str, raises: bool = True) -> bool:
     """Checks if a parameter label is valid.
 
     Valid parameter labels: {R,C,L,Pw,Pn}{N} where N is a number, e.g., P1n.
@@ -73,6 +73,9 @@ def validate_parameter(p: str) -> bool:
     ----------
     p : str
         String representation of the parameter label.
+    raises : bool, optional
+        If True, raises an AssertionError on invalid parameter labels.
+        Default is True.
 
     Returns
     -------
@@ -80,13 +83,22 @@ def validate_parameter(p: str) -> bool:
         True if the parameter label is valid, False otherwise.
     """
     # Check if parameter label is a string
-    assert isinstance(p, str), "Parameter label must be a string."
+    if not isinstance(p, str):
+        if raises:
+            raise AssertionError("Parameter label must be a string.")
+        return False
     # Check if parameter label is not empty
-    assert len(p) > 0, "Parameter label is empty."
+    if not p:
+        if raises:
+            raise AssertionError("Parameter label is empty.")
+        return False
     # Check if parameter label is valid
     pattern = r"(?:R\d+|C\d+|L\d+|P\d+[wn])"
-    assert re.fullmatch(pattern, p), f"Invalid parameter label: {p}"
-    return True  # If all checks pass, the parameter label is considered valid
+    if not re.fullmatch(pattern, p):
+        if raises:
+            raise AssertionError(f"Invalid parameter label: {p}")
+        return False
+    return True
 
 
 def parse_component(c: str) -> str:
