### 2024-07-22 99bd911
refactor: Implement repr for `InferenceResult`
