### 2024-07-22 16bb5cd
refactor: `variables` is now inferred from dict keys rather than passed explicitly
