### 2024-07-22 ee94198
refactor: Use 'converged' instead of 'success' for consistency
