### 2024-07-23 583e302
refactor: Don't clutter circuits dataframe with unnecessary columns
