### 2024-07-23 3bc75dc
refactor: Don't sort circuits dataframe, breaks indexing later
