### 2024-07-23 30b5830
bug: Fix edge case where priors is still None when private SCSD function is called
