### 2024-07-23 c4451f6
refactor: `print_inference_results` should rank circuits internally (without modifying circuits dataframe)
