+++ _sources/generated/generated/autoeis.utils.eval_posterior_predictive.rst.txt
@@ -1,0 +1,6 @@
+eval\_posterior\_predictive
+===========================
+
+.. currentmodule:: autoeis.utils
+
+.. autofunction:: eval_posterior_predictive
