+++ _sources/examples/parallel_inference.ipynb.txt
@@ -1,0 +1,1931 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Batch Analysis"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "In this notebook, we'll learn how to analyze EIS data in batch mode. Normally, you have a single set of EIS data, i.e., set of impedance measurements at various frequencies, plus a common circuit model that you want to fit to the data. This is what we call single circuit, single dataset or SCSD in short. However, there are two other modes of analysis that you might encounter in practice:\n",
+    "\n",
+    "- Single circuit, multiple datasets (SCMD): You have multiple datasets, each with its own impedance measurements, but you want to fit the same circuit model to all of them. A good example of this is when you have EIS data for multiple samples which you want to compare, or a single sample under different conditions, e.g., EIS data at different cycles during battery cycling.\n",
+    "\n",
+    "- Multiple circuits, single dataset (MCSD): You have a single dataset, but you want to fit different circuit models to it. This is useful when you want to compare different models to see which one fits the data best, which is by the way the classic use case of AutoEIS itself!"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-07-23T21:35:37.920107Z",
+     "iopub.status.busy": "2024-07-23T21:35:37.919504Z",
+     "iopub.status.idle": "2024-07-23T21:35:41.236573Z",
+     "shell.execute_reply": "2024-07-23T21:35:41.235997Z"
+    },
+    "scrolled": true
+   },
+   "outputs": [],
+   "source": [
+    "import random\n",
+    "\n",
+    "import autoeis as ae\n",
+    "import matplotlib.pyplot as plt\n",
+    "\n",
+    "ae.visualization.set_plot_style()"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Single circuit, multiple datsets (SCMD)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "To test this, we can use a toy dataset that ships with the package. This dataset contains EIS data for a coin cell battery measured at discharged state at various cycles. Let's load the dataset and see what it looks like."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-07-23T21:35:41.239148Z",
+     "iopub.status.busy": "2024-07-23T21:35:41.238880Z",
+     "iopub.status.idle": "2024-07-23T21:35:41.248916Z",
+     "shell.execute_reply": "2024-07-23T21:35:41.248565Z"
+    }
+   },
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Number of cycles: 130\n"
+     ]
+    }
+   ],
+   "source": [
+    "datasets = ae.io.load_battery_dataset()\n",
+    "print(f\"Number of cycles: {len(datasets)}\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
[…]
