### 2024-07-24 392abea
docs: Fix incorrect type hint
