### 2024-07-24 663d73d
api: Deprecate `perform_full_analysis` until it's fixed.
