+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-23T21:23:09.566971Z",
-     "iopub.status.busy": "2024-07-23T21:23:09.566456Z",
-     "iopub.status.idle": "2024-07-23T21:23:13.334473Z",
-     "shell.execute_reply": "2024-07-23T21:23:13.333942Z"
+     "iopub.execute_input": "2024-07-24T02:55:42.320406Z",
+     "iopub.status.busy": "2024-07-24T02:55:42.319901Z",
+     "iopub.status.idle": "2024-07-24T02:55:46.259263Z",
+     "shell.execute_reply": "2024-07-24T02:55:46.258608Z"
     }
    },
    "outputs": [],
@@ -41,10 +41,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-23T21:23:13.337627Z",
-     "iopub.status.busy": "2024-07-23T21:23:13.337289Z",
-     "iopub.status.idle": "2024-07-23T21:23:13.345684Z",
-     "shell.execute_reply": "2024-07-23T21:23:13.345336Z"
+     "iopub.execute_input": "2024-07-24T02:55:46.262497Z",
+     "iopub.status.busy": "2024-07-24T02:55:46.262057Z",
+     "iopub.status.idle": "2024-07-24T02:55:46.271138Z",
+     "shell.execute_reply": "2024-07-24T02:55:46.270480Z"
     }
    },
    "outputs": [],
@@ -64,10 +64,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-23T21:23:13.347584Z",
-     "iopub.status.busy": "2024-07-23T21:23:13.347487Z",
-     "iopub.status.idle": "2024-07-23T21:23:15.753727Z",
-     "shell.execute_reply": "2024-07-23T21:23:15.753129Z"
+     "iopub.execute_input": "2024-07-24T02:55:46.273596Z",
+     "iopub.status.busy": "2024-07-24T02:55:46.273475Z",
+     "iopub.status.idle": "2024-07-24T02:55:48.564605Z",
+     "shell.execute_reply": "2024-07-24T02:55:48.563985Z"
     }
    },
    "outputs": [
@@ -111,10 +111,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-23T21:23:15.756634Z",
-     "iopub.status.busy": "2024-07-23T21:23:15.756135Z",
-     "iopub.status.idle": "2024-07-23T21:23:15.766007Z",
-     "shell.execute_reply": "2024-07-23T21:23:15.765598Z"
+     "iopub.execute_input": "2024-07-24T02:55:48.567196Z",
+     "iopub.status.busy": "2024-07-24T02:55:48.566780Z",
+     "iopub.status.idle": "2024-07-24T02:55:48.570241Z",
+     "shell.execute_reply": "2024-07-24T02:55:48.569896Z"
     }
    },
    "outputs": [
@@ -145,10 +145,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-23T21:23:15.767989Z",
-     "iopub.status.busy": "2024-07-23T21:23:15.767890Z",
-     "iopub.status.idle": "2024-07-23T21:23:15.770407Z",
-     "shell.execute_reply": "2024-07-23T21:23:15.770085Z"
+     "iopub.execute_input": "2024-07-24T02:55:48.572217Z",
[…]
