+++ _sources/examples/circuit_basics.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-24T02:55:42.320406Z",
-     "iopub.status.busy": "2024-07-24T02:55:42.319901Z",
-     "iopub.status.idle": "2024-07-24T02:55:46.259263Z",
-     "shell.execute_reply": "2024-07-24T02:55:46.258608Z"
+     "iopub.execute_input": "2024-07-24T02:58:43.684659Z",
+     "iopub.status.busy": "2024-07-24T02:58:43.684128Z",
+     "iopub.status.idle": "2024-07-24T02:58:47.242675Z",
+     "shell.execute_reply": "2024-07-24T02:58:47.241755Z"
     }
    },
    "outputs": [],
@@ -41,10 +41,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-24T02:55:46.262497Z",
-     "iopub.status.busy": "2024-07-24T02:55:46.262057Z",
-     "iopub.status.idle": "2024-07-24T02:55:46.271138Z",
-     "shell.execute_reply": "2024-07-24T02:55:46.270480Z"
+     "iopub.execute_input": "2024-07-24T02:58:47.246736Z",
+     "iopub.status.busy": "2024-07-24T02:58:47.246274Z",
+     "iopub.status.idle": "2024-07-24T02:58:47.248762Z",
+     "shell.execute_reply": "2024-07-24T02:58:47.248366Z"
     }
    },
    "outputs": [],
@@ -64,10 +64,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-24T02:55:46.273596Z",
-     "iopub.status.busy": "2024-07-24T02:55:46.273475Z",
-     "iopub.status.idle": "2024-07-24T02:55:48.564605Z",
-     "shell.execute_reply": "2024-07-24T02:55:48.563985Z"
+     "iopub.execute_input": "2024-07-24T02:58:47.250680Z",
+     "iopub.status.busy": "2024-07-24T02:58:47.250556Z",
+     "iopub.status.idle": "2024-07-24T02:58:49.395632Z",
+     "shell.execute_reply": "2024-07-24T02:58:49.395159Z"
     }
    },
    "outputs": [
@@ -111,10 +111,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-24T02:55:48.567196Z",
-     "iopub.status.busy": "2024-07-24T02:55:48.566780Z",
-     "iopub.status.idle": "2024-07-24T02:55:48.570241Z",
-     "shell.execute_reply": "2024-07-24T02:55:48.569896Z"
+     "iopub.execute_input": "2024-07-24T02:58:49.398117Z",
+     "iopub.status.busy": "2024-07-24T02:58:49.397709Z",
+     "iopub.status.idle": "2024-07-24T02:58:49.407060Z",
+     "shell.execute_reply": "2024-07-24T02:58:49.406646Z"
     }
    },
    "outputs": [
@@ -145,10 +145,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2024-07-24T02:55:48.572217Z",
-     "iopub.status.busy": "2024-07-24T02:55:48.572112Z",
-     "iopub.status.idle": "2024-07-24T02:55:48.580909Z",
-     "shell.execute_reply": "2024-07-24T02:55:48.580520Z"
+     "iopub.execute_input": "2024-07-24T02:58:49.408899Z",
[…]
