### 2024-09-11 #6256 @dap-biospec
@ma-sadeghi @lucydot Code is still broken and cannot be evaluated.
Sample output on Windows:
```
>>> circuits = ae.perform_full_analysis(freq, Z, iters=100, parallel=True)
Traceback (most recent call last):
  File "", line 1, in
  File "[...]\Python312\Lib\site-packages\autoeis\core.py", line 898, in perform_full_analysis
    raise NotImplementedError(
NotImplementedError: This function has some issues and is not yet ready to use. Regardless, it is recommended to use the step-by-step approach for more control. Please refer to the documentation for more details.
```

Sample output on Centos 9:
```
>>> circuits = ae.perform_full_analysis(freq, Z, iters=100, parallel=True)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "[...]/python3.12/site-packages/autoeis/core.py", line 898, in perform_full_analysis
    raise NotImplementedError(
NotImplementedError: This function has some issues and is not yet ready to use. Regardless, it is recommended to use the step-by-step approach for more control. Please refer to the documentation for more details.
```

AutoEIS home page has a prominent warning that function [perform_full_analysis] is disabled. Therefore, it is not surprising that usage tutorial is broken.
