### 2024-09-11 #6256 @ma-sadeghi
@dap-biospec Thanks for getting back to us. As mentioned in the Exception message, that function is deprecated and is not to be used. Please visit the examples page under our documentation.

I'll make sure to update the Basic Usage page to remove it altogether to not create further confusion for the users. (Basically, that function was never meant to be part of the package, sorry for the confusion)

Apart from this, you should be able to continue with the review.
