### 2024-09-12 #6256 @ma-sadeghi
@dap-biospec I just updated both the documentation and the README file to make it clear for the users not to use the envelope function.
