### 2024-10-03 #6256 @dap-biospec
@ma-sadeghi It appears that you attempted to resolve issues by eliminating introductory section from depository all together.  This is not a solution to the bugs in the code. Please include basic step-by-step tutorial that can be clearly found and followed from the landing page of the depository. It may be prudent to include verifiable demo section for each claim that you are making in the manuscript. Please include all the required steps in a markup page(s), leaving linked jupyter notebooks only for  illustrations that cannot be included in the markup files.  As @lucydot rightfully pointed out, please ask an un-initiated colleague to follow written tutorial without any additional support and verify that they obtain all the results that you claim and expect to see.
