### 2024-10-04 #6256 @ma-sadeghi
@dap-biospec Thanks for the feedback. Here’s how we’ve addressed your points:

The perform_full_analysis function was originally meant as a convenience helper, but it turned out to add unnecessary maintenance complexity and made debugging harder by masking individual steps. We decided to remove it, but this doesn’t affect AutoEIS’s core functionality.

The "basic usage" section was removed from the README to keep the landing page clean and focused. We’ve linked to the full documentation, which covers everything in detail:

<kbd> <img src="https://github.com/user-attachments/assets/7989f6c0-14dc-4274-94bb-92b20e45152a"> </kbd> <kbd> <img src="https://github.com/user-attachments/assets/4b768b8d-eeb5-4951-8231-d74a939351ab"> </kbd>
Is there a reason you’d prefer it on the landing page?

On Jupyter notebooks vs. Markdown: Good point. Notebooks can be slow to render. To handle that, we convert them to HTML in the documentation, so they’re easy to view. Users can still grab the actual notebooks from the repo if they want to run them. We avoid Markdown for examples because it’s harder to automate testing for it, and Jupyter notebooks are tested on every push.

Per your and @lucydot's suggestion, we had colleague independently install and use AutoEIS, and they were able to go through the examples.

Hope these address your concerns.
