### 2024-11-01 #6256 @dap-biospec
@ma-sadeghi  I attempted to reproduce the first three examples and encountered variety of errors and inconsistencies in each of "Circuit models 101", "circuit generation", and "detailed workflow". This starts with the lack of arguments for ae.visualization.set_plot_style() example and continues with runtimes errors and unexpected keywords.

It does not appear that the authors performed literal execution of their tutorial on a clean system. Errors may be due to package dependence or environment configuration.

It also appears that the authors  exported  jupyter notebook as HTML without correcting action steps or providing adequate instructions on the expected behavior.

@lucydot at this time functionality is not verifiable.
