### 2024-11-03 #6256 @ma-sadeghi
@dap-biospec Thank you for taking the time to review.

---

> @ma-sadeghi I attempted to reproduce the first three examples and encountered various errors and inconsistencies in "Circuit models 101," "circuit generation," and "detailed workflow." Issues include missing arguments for ae.visualization.set_plot_style() and further runtime errors and unexpected keywords.

All of our example notebooks are automatically tested on GitHub virtual machines. These tests are designed to catch any cell that fails to run. The tests are two-fold: [One](https://github.com/AUTODIAL/AutoEIS/blob/main/.github/workflows/notebooks.yml) is executed on every pull request, while [another](https://github.com/AUTODIAL/AutoEIS/blob/main/.github/workflows/gh-pages.yml) runs on every push to the repository to update the documentation. All recent tests have passed. However, to double-check, I cloned a fresh repository, created a fresh virtual environment, and successfully ran the "detailed_workflow.ipynb" notebook (which covers core AutoEIS functionalities). I’ve recorded the process, and you can view it [here](https://youtu.be/UdIWxDuJjj8).

> It does not appear that the authors performed a literal execution of their tutorial on a clean system. Errors may be due to package dependency or environment configuration.

As explained above, the examples are tested and run on GitHub virtual machines as part of our CI workflow.

> It also appears that the authors exported the Jupyter notebook as HTML without addressing action steps or providing adequate instructions on the expected behavior.

The snippet below shows the Action responsible for publishing the documentation. This workflow is executed on every push to the repository:

```yaml
      - name: Build the documentation
        run: |
          uv run jupyter nbconvert \
            --ExecutePreprocessor.timeout=3600 \
            --to notebook --execute --inplace examples/*.ipynb
          cd doc
          uv run make html
```

Based on this snippet, prior to generating HTML files, all example notebooks are re-run, and results are updated. If any cell fails, the published HTML reflects the errors (I double checked our online documentation, and there was no error). Additionally, as noted, there’s a separate Action (notebooks.yml) that runs for each pull request. This setup ensures that any changes causing failures are caught and cannot be merged.

---

If you're experiencing specific issues with your local setup, I’d be glad to help you troubleshoot. Feel free to share details about your environment, dependencies, or any error messages, and I can assist in pinpointing the source of the problem.
