### 2024-11-03 #6256 @dap-biospec
@ma-sadeghi  Thank you for confirming that HTML content was literally translated from jupyter notebooks.  This likely explains why HTML instructions do not work, by the same account as printing of dynamic HTML content to PDF does not yield working pages.

Please review previous request to verify that literal execution of your published HTML instructions yield expected demonstrations.  A "literal"  means that a reader that  has absolutely no prior knowledge of your package can follow your instructions from this page  https://github.com/AUTODIAL/AutoEIS/ verbatim (i.e. copy-paste) and arrive at the same results.  Please also note that code validation by script, or auto-generation of HTML from notebooks in not sufficient to make "as published" tutorial adequate. Hence, the earlier request to test the instructions by uninitiated human.
