### 2024-11-04 #6256 @lucydot
@dap-biospec thank you for continued efforts testing the code and @ma-sadeghi for clear account of testing and verification setup.

Testing the code via Github Actions is a widely used tool as it should reflect what occurs for when software is freshly installed. It is peculiar that the GH Action runs without error, yet a human following the same procedure does get errors.

Do you think this could this be a difference in Python and/or Julia versions @ma-sadeghi ? I can see the GH workflows `tests.yml` and `notebooks.yml` use julia version v1.10 and python version v3.10.

@dap-biospec  - could you let us know which python version you are using? Are you running in a clean python environment (for example, with a virtual environment or conda)? This might help us understand if there are version conflicts.
