### 2024-11-04 #6256 @dap-biospec
@ma-sadeghi is making a small but consequential omission in the online tutorial section.  Code that may run in jupyper notebook may not run, or produce the same output, outside of it. Hence, simple HTML export does not work as an online documentation.  Furthermore, execution of code by copying and pasting in plain python prompt does not yield described results, as outlined above.

As a publication, this work should  be self-sufficient and accessible to any reader, including non-programming chemists.
