### 2024-11-11 #6256 @lucydot
@ma-sadeghi  - this sounds like a good plan. What you summarise is also what I think @dap-biospec is suggesting.
