### 2024-11-22 #6256 @ma-sadeghi
Hi @lucydot, thanks for you patience. I did successfully reproduce it in a fresh Python REPL. You can find the screencast [here](https://youtu.be/9f8KI2wxdNw).
