### 2024-11-25 #6256 @dap-biospec
@ma-sadeghi Here are screenshots of REPL executions.  As the screencast is not a part of the manuscript, I am going only by the actual posted steps.

First, "Circuits 101" leads to a prompt:
![image](https://github.com/user-attachments/assets/2bcdcce3-0281-4aee-8a33-eeb7c466c10d)
In the screenshots below this step was bypassed.

While you mention lcapy and LaTeX in "Circuits 101", this should be clearly spelled in the installation section similarly to Julia if you are relying on it to a similar extent.  Alternatively, you can have separate and clearly marked pages with or without visualization.

"Circuit generation":
![image](https://github.com/user-attachments/assets/7b21f224-3219-45fb-9dea-00864e764c99)

"Detailed workflow":
![image](https://github.com/user-attachments/assets/32d65bad-488e-4775-9840-22a217ddcb61)

Please let me know when this is addressed in the code or in the tutorial and I will repeat.
