### 2024-12-07 #6256 @lucydot
Hi @dap-biospec and @ma-sadeghi  -

I've had a run through installation and tutorials myself. In general I try to avoid doing this as editor, but given the considerable time that has passed, and effort already invested, I think it is sensible to take a more direct approach so that we can wrap this up.

## Installation

I created an empty conda environment, installed pip, and then followed the instructions given: `pip install -U autoeis`. It proceeded for me without hiccup.

> While you mention lcapy and LaTeX in "Circuits 101", this should be clearly spelled in the installation section similarly to Julia if you are relying on it to a similar extent

I agree with this suggestion from @dap-biospec. To me it makes sense to keep all installation steps on the installation page @ma-sadeghi. It could be included as an optional sub-section for those who want to visualise the circuit (making clear that for most autoeis functionality these additional steps, which are a little more involved than a pip-install, are not needed).

## Jupyter vs REPL

I think there is an assumption that the tutorials are being run through a Jupyter Notebook - for example, `ae.visualization.plot_impedance_combo(freq, Z)` does not save (or otherwise display) a plot when following the tutorials with REPL. I think this assumption is fair for many (perhaps even most) users.

@ma-sadeghi if you are working on the assumption that tutorials are completed within a Jupyter Notebook, you could state this at the top of each tutorial page. To support those who want to use REPL, you could point to the nice screencast you have developed and demonstrate how to save a figure using `plt.savefig` function  - I know vast majority of researchers will know this, but may help students/those just starting out with Python/matplotlib.

## Functionality

@dap-biospec I could not reproduce the errors you had in "Circuit generation" and "Detailed workflow". This difference seems unusual to me; perhaps this is a bug which has been fixed since starting the review - I suggest re-installing autoeis (ideally within a virtual environment / conda), if you have not already done so, and re-running through these tutorials.
