### 2024-12-17 #6256 @dap-biospec
@ma-sadeghi I updated autoeis to 0.0.34.
"ae.visualization.set_plot_style()" still produces user prompt.
If bypassed, circuit generation successfully completed  " ae.core.generate_equivalent_circuits(freq, Z, **kwargs)". I will continue testing further, but the prompt must be addresses in tutorial or the software.
