### 2024-12-17 #6256 @dap-biospec
@lucydot For a depository which is used solely by the authors and close colleagues, any assumptions are warranted. A peer reviewed publication imposes additional requirements on spelling out assumptions, abbreviations, prerequisites and several other established standard, without which depository does not rise to a level of a publication. While I can easily guess that  ae.visualization.plot_impedance_combo(freq, Z) and similar calls may not have the same output in REPL as in  jupyter, a) we cannot exclude electrochemists without extensive programming background from the audience and b) the example code must proceed through all steps without errors.

Furthermore, I have substantial concerns about rolling code changes. Release says 0.0.34, list says 0.0.36, paper submitted for review still says 0.0.17, if I am not mistaken.  In my opinion, the submitted paper must be locked to the release under review with the optional information on obtaining the latest version. Otherwise, the accepted paper can point to a materially different code, however well intended the changes were.
