### 2024-12-19 #6256 @lucydot
> A peer reviewed publication imposes additional requirements on spelling out assumptions, abbreviations, prerequisites and several other established standard, without which depository does not rise to a level of a publication.

I agree, hence the suggestion to state whether Jupyter Notebooks are being used at the top of the tutorial.

>  In my opinion, the submitted paper must be locked to the release under review with the optional information on obtaining the latest version.

It is very usual for code to be updated and new versions released during the JOSS review process. In fact this is encouraged - it means the review process is working (as the code is being improved through review). When a code passes review, we ask authors to make a final release and the version reported on the paper is updated accordingly.
