### 2025-01-17 27cf3b3
bug: `set_plot_style` breaks if run from non-interactive sessions, e.g., REPL
