### 2025-01-28 #6256 @lucydot
@ma-sadeghi - what are your thoughts on your comments above - are you able to update tutorials so that they run as written for REPL and Jupyter?

@dap-biospec - has your testing raised any further issues? If the Jupyter vs REPL issue is resolved are you happy that stated functionality is demonstrated?
