### 2025-02-06 #6256 @ma-sadeghi
Hi @lucydot, I apologize for the delay. I agree, here's what we've done to address @dap-biospec's comments:

- There was a subtle bug that broke REPL workflows. The bug is now [fixed](27cf3b34083bf14bb7c498ada99327c7fb8d4939). If you're interested in the details: The reason the bug didn't surface in my screen cast was that I was using an IPython REPL, not Python's native REPL.
- That said, because of the interactive nature of typical AutoEIS workflows, we still strongly encourage users to use a proper interactive environment, e.g., IPython Notebook, Jupyter Lab, VS Code, etc. So, we added a prominent warning to both the README and the examples' landing page to communicate this to users.

![Image](https://github.com/user-attachments/assets/e8e0c4a5-15ea-479e-971a-7d40c8280e02)
![Image](https://github.com/user-attachments/assets/97dd7827-7920-42b9-a844-9ae1172e6fec)

Hope this addresses the issue.
