### 2025-02-12 #6256 @lucydot
@dap-biospec - has your testing raised any further issues? Please see message above from @ma-sadeghi; they have made clear that examples are expected to run in a Jupyter Notebook (with the assumption being that basic adaptations to the example as written would enable integration into Python REPL or script).
