### 2025-02-21 #6256 @dap-biospec
@ma-sadeghi  While the previous errors were cleared, additional errors have been encountered and examples are still not reproducible.

1. "Example notebooks" pages are still a plain HTML conversion of the content of Jupyter notebooks, with no notices that each notebook must be copy/pasted into particular environment or the included URL of notebook to download and run. Thus, HTML examples are misleading.
2. Example notebooks  are not renderable on Github depository and no instructions are provided to a broad scientific community on handling them, whether on the ReadMe page or examples section.
3.  Example calls or code snippets that require graphical environment are still not clearly marked as such. An umbrella warning, provided in the installation section, is not sufficient to identify environment prerequisites. The authors must know environment dependence of each call and should not offload it to the reader.
4. ae.visualization.draw_circuit(circuit) causes runtime error for unresolved dependency form pdflatex, which is not automatically resolved or addressed in manual installation instructions. Thus, the pattern of implicit dependency expectations, raised several times before, continues. This occurs in both Python command line and Jupyter environments.
5. In cases where code executes, reported values and circuits are different from those provided in the example notebooks, with no explanation provided for the discrepancy. (ae.core.generate_equivalent_circuits(freq, Z, **kwargs) and ae.core.filter_implausible_circuits(circuits_unfiltered) ) Thus, claimed results are not reproducible.
6. Runtime errors continue at several steps in the example code,  ending up with multiple errors an lock-up at ae.core.perform_bayesian_inference(circuits, freq, Z) step:
 Refining Initial Guess   0% ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 0/3  [ 0:00:23 < -:--:-- , ? it/s ][15:57:03] WARNING  Removing Julia environment directory:                                           julia_helpers.py:222
                    <source>\.julia\environments\pyjuliapkg
Refining Initial Guess   0% ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 0/3  [ 0:00:24 < -:--:-- , ? it/s ]Traceback (most recent call last):
Refining Initial Guess   0% ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 0/3  [ 0:00:24 < -:--:-- , ? it/s ]  File "D:\Python\Python312\Lib\site-packages\autoeis\julia_helpers.py", line 203, in ensure_julia_deps_ready
    _ensure_julia_deps_ready(quiet)
  File "D:\Python\Python312\Lib\site-packages\autoeis\julia_helpers.py", line 189, in _ensure_julia_deps_ready
    Main = init_julia(quiet=quiet)
           ^^^^^^^^^^^^^^^^^^^^^^^
  File "D:\Python\Python312\Lib\site-packages\autoeis\julia_helpers.py", line 70, in init_julia
    from juliacall import Main
  File "D:\Python\Python312\Lib\site-packages\juliacall\__init__.py", line 251, in <module>
    init()
  File "D:\Python\Python312\Lib\site-packages\juliacall\__init__.py", line 227, in init
    res = jl_eval(script.encode('utf8'))
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
OSError: [WinError 541541187] Windows Error 0x20474343
[errors truncated]

Further testing was halted as irrelevant.
