### 2025-02-21 #6256 @dap-biospec
@lucydot After over a year of efforts to bring the manuscript in line with expectations for a peer-reviewed research article, I arrive at the conclusion that the authors are either unable or unwilling to make a niche GitHub depository accessible to a broad research community beyond its current users. Furthermore,  I do not believe that the content of this manuscript makes a substantial new contribution beyond what was  already published in [this](https://iopscience.iop.org/article/10.1149/1945-7111/aceab2/meta) article and, thus, does not warrant a separate publication. Based on the lack of novel contribution, lack of reproducibility, and poor documentation I recommend rejection of this manuscript.
