### 2025-02-27 #6256 @lucydot
A note here to say that I'm in discussion with authors and editors re: the best way to proceed.
I will be at a conference until the 10th March, so may be slower to reply than usual until I'm back in the office.
