### 2025-03-10 #6256 @lucydot
Hi @ma-sadeghi - the reviewer comments and my thoughts are below:

> 1. "Example notebooks" pages are still a plain HTML conversion of the content of Jupyter notebooks, with no notices that each notebook must be copy/pasted into particular environment or the included URL of notebook to download and run. Thus, HTML examples are misleading.

As long as you are clear at the top of each tutorial page that to run the code as exactly written you will need a Jupyter notebook, I judge point 1 to be addressed.

 > 2. Example notebooks are not renderable on Github depository and no instructions are provided to a broad scientific community on handling them, whether on the ReadMe page or examples section.

The documentation approach you take (Jupyter which is compiled and served through Github to a docs website) is quite standard in the field. No further action need be taken on this.

Points 3 and 4 have been fixed during the review process.

> 5. In cases where code executes, reported values and circuits are different from those provided in the example notebooks, with no explanation provided for the discrepancy. (ae.core.generate_equivalent_circuits(freq, Z, **kwargs) and ae.core.filter_implausible_circuits(circuits_unfiltered) ) Thus, claimed results are not reproducible.

Which you responded:

> This is expected. Equivalent circuits are generated using an evolutionary algorithm, which is a stochastic process. So, you'd expect a different pool of circuits every time. (Of course, you'd expect the algorithm to converge to a few good candidates if you generate a large pool of circuits via iters keywords)

This randomness  was not obvious to me, if you have not already done so, it may be worth flagging this up in the tutorial docs.

> 6. Runtime errors continue at several steps in the example code, ending up with multiple errors an lock-up at ae.core.perform_bayesian_inference(circuits, freq, Z) step:

I will run through these examples myself to see if I can reproduce the errors.
