### 2025-03-10 #6256 @lucydot
Hi @ma-sadeghi :

I've just ran through the "Batch Analysis" workflow, as this contains the code (`ae.core.perform_bayesian_inference(circuits, freq, Z)`) where the error reported above was generated. I ran through the other tutorials earlier in the review process.

I used a Python REPL, following the reviewers approach, making small changes where necessary, such as leaving out `ae.visualization.set_plot_style()`) commands. I do not reproduce the same error - the code runs as expected following the tutorial.

You also now have clear guidance to use a Jupyter Notebook environment, so there is little room for confusion. Those who are unable to adapt the code for the REPL can use the supported environment.

RE: the error reported above. It may be that the reviewer did not have the correct Julia installation, or that it is an issue with older Windows OS (see for instance: https://discourse.julialang.org/t/convert-string-encoding-to-utf-8/57782/2). I am running this with Julia 1.9.1, in line with what your docs states as being necessary (a Julia version > 1.9), and on a mac. If you are testing on a recent Windows OS that is already more than what I usually see, where it is often assumed the user is running on a Unix system.

In short, I can verify functionality, and am happy with docs after the changes made during review.  I'm going to generate a checklist now and work through it. I've discussed this approach with the wider editorial team; they are happy for me to proceed as reviewer.
