### 2025-03-15 #6256 @ma-sadeghi
@lucydot Thanks for taking the time to complete the review. I'll go through the final checklist soon and will update this thread.

**PS 1**. Per your suggestion, a paragraph was added to the circuit generation notebook explaining why generated circuits are not deterministic. We also remembered that `generate_equivalent_circuits` accepts a `seed` argument for reproducibility. This is also now documented in the same notebook. (the API was already documented under API Reference.)

**PS 2**. The issue with `ae.visualization.set_plot_style()` in a native Python REPL should now be fixed by this [commit](https://github.com/AUTODIAL/AutoEIS/commit/27cf3b34083bf14bb7c498ada99327c7fb8d4939). It was related to `matplotlib` inline backend, which is now dynamically set depending on the environment.
