### 2025-04-09 #6256 @lucydot
@ma-sadeghi two DOIs need to be added to the references: for the DearEIS and Nature Review Methods Primers.
