### 2025-04-10 #6256 @ma-sadeghi
> Hi [@ma-sadeghi](https://github.com/ma-sadeghi) ; it looks like you have completed almost everything, are you happy for me to do final checks of the paper my end?

Hi @lucydot. We made some minor changes to the manuscript, including changing the title to a more informative one: "AutoEIS: Automated equivalent circuit modeling from electrochemical impedance spectroscopy data using statistical machine learning". The missing DOIs have also been added, and unused references removed from `paper.bib`. Should be good to go on our end.

Just two things:

1. Regarding "Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.": do you want me to rename the Zenodo entry to the actual title of the paper? Currently, it's just the name of the repository (the default value): "AUTODIAL/AutoEIS"
2. Since we've modified the paper title, do we need to rename this thread or the bot automatically handles that when adding the entry to https://joss.theoj.org/?

Thanks for all your help throughout the process.
