### 2025-04-24 #6256 @lucydot
Hi @ma-sadeghi

> Regarding "Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.": do you want me to rename the Zenodo entry to the actual title of the paper? Currently, it's just the name of the repository (the default value): "AUTODIAL/AutoEIS"

Yes, exactly. The title of the Zenodo entry and the paper title must exactly match

> Since we've modified the paper title, do we need to rename this thread or the bot automatically handles that when adding the entry to https://joss.theoj.org/?

I've manually updated the thread title for consistency.
