### 2025-05-01 AUTODIAL/AutoEIS#141 @Kevin-Mattheus-Moerman
Example notebook basic_workflow.ipynb fails with "no plausible circuits found"
When running `basic_workflow.ipynb` I get the following with the default `kwargs`:

```python
    kwargs = {
        "iters": 12,
        "complexity": 12,
        "population_size": 100,
        "generations": 30,
        "terminals": "RLP",
        "tol": 1e-3,
        "parallel": True
    }
```

![Image](https://github.com/user-attachments/assets/00842323-6204-45c1-8528-0b292ecfd69d)

So no plausible circuits are found. I did not have time to check an exhaustive set of `kwargs` that would be best, but I do get some circuits returned if I use `"tol:1e-2"` and `"iters": 25`.

The intended behaviour here surely, is that the basic example should reliable run without user interference/trouble shooting. Hence I suggest you optimise these `kwargs` to a "very safe" set that should always run.

Separately, the notebook features `use_pregenerated_circuits = False` but from the note text above it sounds like its default should be `use_pregenerated_circuits = True`.

I would implement both the `use_pregenerated_circuits = True` and a general set of suitable `kwargs` to address the above.
