### 2025-05-01 #6256 @Kevin-Mattheus-Moerman
as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:

👉 Although your README features basic statements on welcoming contributions, we recommend that you add more detailed contributing guidelines. You can add these and mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

The review was completed by the handling editor @lucydot. I can also add that I have myself successfully installed the software and evaluated the notebook examples provided (on Ubuntu Linux and using VS Codium). I have posted one issue that you should address in relation to the examples: https://github.com/AUTODIAL/AutoEIS/issues/141.

- [ ] Found a minor typo `infereces` in `...let's see how many of the infereces converged` in the notebook `circuit_generation.ipynb`.
- [ ] In the paper, check if `ohmic` should be written with a capital letter i.e. `Ohmic`.
