### 2025-05-09 0d28173
fix: use safer default kwargs to ensure some circuits are found
