### 2025-05-09 #6256 @ma-sadeghi
Hi @Kevin-Mattheus-Moerman. I just pushed three commits:

- Fix type in parallel inference notebook
- Fix typo in paper.md
- Use safer kwargs in the basic workflow notebook
