### 2025-05-14 #6256 @ma-sadeghi
@lucydot Added contributing guide and a mention in README. Also addressed the issue raised by @Kevin-Mattheus-Moerman (AUTODIAL/AutoEIS#141), awaiting his confirmation whether or not the change I made fixed the issue. Thanks!
